% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_heatmap.R
\name{absoluteProbabilityHeatmap}
\alias{absoluteProbabilityHeatmap}
\title{Renders a heatmap based on a population matrix from the factorized counts matrix.}
\usage{
absoluteProbabilityHeatmap(counts, celda.mod, main = NA)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_CG".}

\item{main}{The title of the plot. Default NULL.}
}
\description{
Renders a heatmap based on a population matrix from the factorized counts matrix.
}
