% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{appendCeldaList}
\alias{appendCeldaList}
\title{Append two celdaList objects}
\usage{
appendCeldaList(list1, list2)
}
\value{
A celdaList object. This object contains all resList entries and runParam records from both lists.
}
\description{
Returns a single celdaList representing the combination of two provided celdaList objects.
}
\examples{
appended.list = appendCeldaList(celda.CG.grid.search.res, celda.CG.grid.search.res)
}
