% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\docType{data}
\name{available_models}
\alias{available_models}
\title{available models}
\format{An object of class \code{character} of length 3.}
\usage{
available_models
}
\description{
available models
}
\keyword{datasets}
