% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{availableModels}
\alias{availableModels}
\title{available models}
\format{An object of class \code{character} of length 3.}
\usage{
availableModels
}
\description{
available models
}
\examples{
data(availableModels)
availableModels
}
\keyword{datasets}
