% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calc.perplexity}
\alias{calc.perplexity}
\title{calc.perplexity}
\format{list}
\source{
\url{http://github.com/compbiomed/celda}
}
\usage{
calc.perplexity
}
\description{
The result of running the resamplePerplexity function on the pbmc_res model
}
\details{
Perplexity plot created with the pbmc_res model.

Generated by Yusuke Koga
}
\keyword{datasets}
