% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calc.perplexity2}
\alias{calc.perplexity2}
\title{calc.perplexity2}
\format{list}
\source{
\url{http://github.com/compbiomed/celda}
}
\usage{
calc.perplexity2
}
\description{
The result of running the calculatePerplexityWithResamplingfunction on the pbmc_res model. Done on K = 10:20, L = 10:50
}
\details{
Perplexity plot created with the pbmc_res model, 2nd-pass.

Generated by Yusuke Koga
}
\keyword{datasets}
