% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{calcGibbsProb}
\alias{calcGibbsProb}
\title{Calculate Log Likelihood For Single Set of Cluster Assignments (Gene Clustering)}
\usage{
calcGibbsProb(ix, r, z, k, a, b, g)
}
\arguments{
\item{ix}{The index in z corresponding to the cell currently being considered during Gibbs sampling}

\item{r}{A numeric count matrix}

\item{z}{A numeric vector of cluster assignments}

\item{k}{The number of clusters being considered}

\item{a}{Vector of non-zero concentration parameters for sample <-> cluster assignment Dirichlet distribution}

\item{b}{Vector of non-zero concentration parameters for cluster <-> gene assignment Dirichlet distribution}

\item{g}{The number of cell states ("topics")}
}
\description{
This function calculates the log-likelihood of a cell's membership to each possible clusters,
given the cluster assignment for all other cells.
}
\examples{
TODO 
}
