% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{calcLL_gene_clustering}
\alias{calcLL_gene_clustering}
\title{Calculate Log Likelihood For A Set of Gene Clusterings (Gene Clustering)}
\usage{
calcLL_gene_clustering(counts, z, k, alpha, beta, gamma)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{z}{A numeric vector of cluster assignments}

\item{k}{The number of clusters being considered}

\item{alpha}{Vector of non-zero concentration parameters for sample <-> cluster assignment Dirichlet distribution}

\item{beta}{Vector of non-zero concentration parameters for cluster <-> gene assignment Dirichlet distribution}

\item{gamma}{The number of cell states ("topics")}
}
\description{
This function calculates the log likelihood of each clustering of genes generated
over multiple iterations of Gibbs sampling.
}
\examples{
TODO
}
\keyword{likelihood}
\keyword{log}

