% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_performance.R
\name{calculate_dissimilarity_matrix}
\alias{calculate_dissimilarity_matrix}
\title{Calculate a dissimilarity matrix between columns of a counts matrix.}
\usage{
calculate_dissimilarity_matrix(counts, normalize = T,
  dist.metric = "euclidian")
}
\arguments{
\item{counts}{A counts matrix used to run celda}

\item{normalize}{Whether to normalize the provided matrix automatically. Defaults to TRUE.}

\item{dist.metric}{Which dissimilarity metric to use. Options are "euclidian", "cosine". Defaults to "euclidian"}
}
\description{
In order to make distances between cells more comparable despite an invidiual cell's depth of sequencing,
this function requires that the counts matrix be normalized within cells (to CPM), 
and subsequently across cells. This is performed automatically (and toggled by the _normalize_ parameter) 
on the provided counts matrix, unless otherwise specified.
}
