% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{calculate_loglik_from_variables.celda_C}
\alias{calculate_loglik_from_variables.celda_C}
\title{calcLLFromVariables for celda Cell clustering function}
\usage{
calculate_loglik_from_variables.celda_C(counts, s, z, K, alpha, beta)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{s}{A vector indicating the sample for each cell (column) in the count matrix}

\item{z}{A numeric vector of cluster assignments}

\item{K}{An integer or range of integers indicating the desired number of cell clusters (for celda_C / celda_CG models)}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution}

\item{beta}{Non-zero concentration parameter for gene Dirichlet distribution}
}
\description{
calcLLFromVariables for celda Cell clustering function
}
