% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{calculate_loglik_from_variables.celda_CG}
\alias{calculate_loglik_from_variables.celda_CG}
\title{calculate log lileklihood for the celda Cell and Gene clustering function}
\usage{
calculate_loglik_from_variables.celda_CG(counts, s, z, y, K, L, alpha, beta,
  delta, gamma)
}
\arguments{
\item{counts}{A numeric count matrix.}

\item{s}{Vector of sample labels.}

\item{z}{A numeric vector of cell cluster assignments.}

\item{y}{A numeric vector of gene cluster assignments.}

\item{K}{The number of cell populations.}

\item{L}{The number of clusters being considered.}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution.}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell.}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state.}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state.}
}
\description{
calculate log lileklihood for the celda Cell and Gene clustering function
}
