% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{calculate_loglik_from_variables}
\alias{calculate_loglik_from_variables}
\title{Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.}
\usage{
calculate_loglik_from_variables(model, ...)
}
\arguments{
\item{model}{Model to use for calculating log-likelihood of assignments; one of ("celda_C", "celda_CG", "celda_G")}
}
\value{
The log-likelihood of the provided cluster assignment for the provided counts matrix.
}
\description{
Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.
}
