% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_performance.R
\name{calculate_silhouette_width}
\alias{calculate_silhouette_width}
\title{Calculate the silhouette width for a given celda model.}
\usage{
calculate_silhouette_width(celda.mod, dis.matrix, param = "z")
}
\arguments{
\item{celda.mod}{A celda_C, celda_G, or celda_CG object}

\item{dis.matrix}{A dissimilarity matrix, generated from the counts matrix used to generate the celda.mod parameter}

\item{param}{Whether to calculate based off the z (K) or y (L) clustering for a celda_CG object. Defaults to "z".}
}
\value{
An object of class silhouette
}
\description{
Calculate the silhouette width for a given celda model.
}
