% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{calculateLoglikFromVariables.celda_C}
\alias{calculateLoglikFromVariables.celda_C}
\title{Calculate the celda_C log likelihood for user-provided cluster assignments}
\usage{
calculateLoglikFromVariables.celda_C(counts, s, z, K, alpha, beta)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{s}{A vector indicating the sample for each cell (column) in the count matrix}

\item{z}{A numeric vector of cluster assignments}

\item{K}{The total number of clusters in z}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution}

\item{beta}{Non-zero concentration parameter for gene Dirichlet distribution}

\item{...}{Additional parameters}
}
\description{
Calculate the celda_C log likelihood for user-provided cluster assignments
}
