% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{calculateLoglikFromVariables.celda_CG}
\alias{calculateLoglikFromVariables.celda_CG}
\title{Calculate log lileklihood for the celda Cell and Gene clustering model, given a set of cell / gene cluster assignments}
\usage{
\method{calculateLoglikFromVariables}{celda_CG}(counts, s, z, y, K, L, alpha,
  beta, delta, gamma, ...)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{s}{A numeric vector of sample labels corresponding to the samples each cell in the count matrix originates from}

\item{z}{A numeric vector of cell cluster assignments}

\item{y}{A numeric vector of gene cluster assignments}

\item{K}{The number of cell clusters}

\item{L}{The number of gene clusters}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution.}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell.}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state.}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state.}

\item{...}{Additional parameters}
}
\description{
Calculate log lileklihood for the celda Cell and Gene clustering model, given a set of cell / gene cluster assignments
}
