% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{calculateLoglikFromVariables}
\alias{calculateLoglikFromVariables}
\title{Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.}
\usage{
calculateLoglikFromVariables(counts, celda.mod, ...)
}
\arguments{
\item{counts}{The counts matrix used to generate the provided cluster assignments.}

\item{celda.mod}{Celda model. Options available in `celda::available.models`.}

\item{...}{Additional parameters.}
}
\value{
The log-likelihood of the provided cluster assignment for the provided counts matrix.
}
\description{
Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.
}
\examples{
celda.sim = simulateCells(model="celda_CG")
celda.mod = "celda_CG"
loglik = calculateLoglikFromVariables(celda.sim$counts, celda.mod, 
                                      sample.label=celda.sim$sample.label,
                                      z=celda.sim$z, y=celda.sim$y,
                                      K=celda.sim$K, L=celda.sim$L,
                                      alpha=celda.sim$alpha, beta=celda.sim$beta,
                                      gamma=celda.sim$gamma, delta=celda.sim$delta)
}
