% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{calculatePerplexity}
\alias{calculatePerplexity}
\title{Calculate the perplexity from a single celda run}
\usage{
calculatePerplexity(counts, celda.mod, new.counts = NULL)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C", "celda_G" or "celda_CG".}

\item{new.counts}{A new counts matrix used to calculate perplexity. If NULL, perplexity will be calculated for the 'counts' matrix. Default NULL.}
}
\value{
The perplexity for the provided data and model.
}
\description{
Perplexity can be seen as a measure of how well a provided set of 
cluster assignments fit the count data being modeled.
}
\details{
Perplexity is defined in LDA as the exp of the log likelihood of the model
divided by the total amount of word tokens. The corresponding perplexity
for the celda models are derived in their respective model description 
documents. These documents will not be made publicly available until 
after publication of celda.
}
