% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{calculatePerplexityWithCrossValidation}
\alias{calculatePerplexityWithCrossValidation}
\title{Calculate and visualize perplexity of all models in a celda_list, with k-fold cross validation}
\usage{
calculatePerplexityWithCrossValidation(celda.list, counts, folds, title = "")
}
\arguments{
\item{celda.list}{A celda_list object as returned from *celda()*}

\item{counts}{The counts matrix used to generate the provided celda.list}

\item{folds}{Numvber of folds for cross validation}

\item{title}{Title to be appended to the perplexity plot title. Default is "".}
}
\value{
A list with a data frame summarizing all of the calculated perplexities
}
\description{
Calculate and visualize perplexity of all models in a celda_list, with k-fold cross validation
}
