% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{calculatePerplexityWithResampling}
\alias{calculatePerplexityWithResampling}
\title{Calculate and visualize perplexity of all models in a celda_list, with resampling}
\usage{
calculatePerplexityWithResampling(counts, celda.list, resample = 5,
  seed = 12345)
}
\arguments{
\item{counts}{The counts matrix used to generate the provided celda.list}

\item{celda.list}{An object returned from 'celdaGridSearch'}

\item{resample}{The number of times to resample the counts matrix for evaluating perplexity. Default 5.}

\item{seed}{Parameter to set.seed() for random number generation. Default 12345.}
}
\value{
A list with a data frame summarizing all of the calculated perplexities, and a ggplot2 object visualizing them.
}
\description{
Calculates the perplexity (a measure of how well celda's cluster assignments fit the data) for every celda model in a celda_list generated by the celda() function. Perplexity of the cluster assignments given the provided count matrix,
as well as resamplings of that count matrix, provide a distribution of perplexities and a better sense of the quality of a given K/L choice.
}
