% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{cC.decomposeCounts}
\alias{cC.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cC.decomposeCounts(counts, s, z, K)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{s}{An integer vector indicating the sample label for each cell (column) in the count matrix}

\item{z}{A numeric vector of cluster assignments}

\item{K}{The total number of clusters in z}
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
