% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{cCG.decomposeCounts}
\alias{cCG.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cCG.decomposeCounts(counts, s, z, y, K, L)
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
