% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{cD.calcEMDecontamination}
\alias{cD.calcEMDecontamination}
\title{This function updates decontamination}
\usage{
cD.calcEMDecontamination(omat, phi, eta, theta, z, K, beta, delta)
}
\arguments{
\item{omat}{Numeric/Integer matrix. Observed count matrix, rows represent features and columns represent cells}

\item{phi}{Numeric matrix. Rows represent features and columns represent cell populations}

\item{eta}{Numeric matrix. Rows represent features and columns represent cell populations}

\item{theta}{Numeric vector. Proportion of truely expressed transctripts}

\item{z}{Integer vector. Cell population labels}

\item{K}{Integer. Number of cell populations}

\item{beta}{Numeric. Concentration parameter for Phi}

\item{delta}{Numeric. Concentration parameter for Theta}
}
\description{
This function updates decontamination
}
