% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.C.mod}
\alias{celda.C.mod}
\title{celda.C.mod}
\format{A celda_C object}
\usage{
celda.C.mod
}
\description{
celda_C results generated from celda.C.sim
}
\examples{
celda.C.mod = celda_C(celda.C.sim$counts, K=celda.C.sim$K, nchains=1)
}
\keyword{datasets}
