% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.C.sim}
\alias{celda.C.sim}
\title{celda.C.sim}
\format{A list of counts and properties as returned from simulateCells()}
\usage{
celda.C.sim
}
\description{
An example simulated count matrix from the celda_C model.
}
\examples{
celda.C.sim = simulateCells("celda_C")
}
\keyword{datasets}
