% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.CG.grid.search.res}
\alias{celda.CG.grid.search.res}
\title{celda.CG.grid.search.res}
\format{An object as returned from celdaGridSearch()}
\usage{
celda.CG.grid.search.res
}
\description{
Example results of celdaGridSearch on celda.CG.sim
}
\examples{

}
\keyword{datasets}
