% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.CG.mod}
\alias{celda.CG.mod}
\title{celda.CG.mod}
\format{A celda_CG object}
\usage{
celda.CG.mod
}
\description{
celda_CG results generated from celda.CG.sim
}
\examples{
celda.CG.mod = celda_CG(celda.CG.sim$counts, K=celda.CG.sim$K,
                        L=celda.CG.sim$L, nchains=1)
}
\keyword{datasets}
