% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.CG.sim}
\alias{celda.CG.sim}
\title{celda.CG.sim}
\format{A list of counts and properties as returned from simulateCells()}
\usage{
celda.CG.sim
}
\description{
An example simulated count matrix from the celda_CG model.
}
\examples{
celda.CG.sim = simulateCells("celda_CG")
}
\keyword{datasets}
