% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.G.mod}
\alias{celda.G.mod}
\title{celda.G.mod}
\format{A celda_G object}
\usage{
celda.G.mod
}
\description{
celda_G results generated from celda.G.sim
}
\examples{
celda.G.mod = celda_G(celda.G.sim$counts, L=celda.G.sim$L, nchains=1)
}
\keyword{datasets}
