% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.G.sim}
\alias{celda.G.sim}
\title{celda.G.sim}
\format{A list of counts and properties as returned from simulateCells()}
\usage{
celda.G.sim
}
\description{
An example simulated count matrix from the celda_G model.
}
\examples{
celda.G.sim = simulateCells("celda_G")
}
\keyword{datasets}
