% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda.R
\name{celda}
\alias{celda}
\title{Run the celda Bayesian hierarchical model on a matrix of counts.}
\usage{
celda(counts, model, sample.label = NULL, nchains = 1, cores = 1,
  seed = 12345, verbose = F, logfile = "", ...)
}
\arguments{
\item{counts}{A count matrix}

\item{model}{Which celda sub-model to run. Options include "celda_C" (cell clustering), "celda_G" (gene clustering), "celda_CG" (gene and cell clustering)}

\item{sample.label}{A numeric vector indicating the originating sample for each cell (column) in the count matrix. By default, every cell will be assumed to be from an independent sample.}

\item{nchains}{The number of chains of Gibbs sampling to run for every combination of K/L parameters. Defaults to 1}

\item{cores}{The number of cores to use for parallell Gibbs sampling. Defaults to 1.}

\item{seed}{The base seed for random number generation}

\item{verbose}{Print log messages during celda chain execution}

\item{logfile}{Path to output file for logging messages from worker threads. By default, messages are redirected to stderr of the main process.}

\item{...}{Model specific parameters}
}
\value{
Object of class "celda_list", which contains results for all model parameter combinations and summaries of the run parameters
}
\description{
Yields assigments of genes/cells to clusters, depending on the provided model type.
}
