% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celda_heatmap.celda_C}
\alias{celda_heatmap.celda_C}
\title{celda_heatmap for celda Cell clustering function}
\usage{
\method{celda_heatmap}{celda_C}(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_C"}

\item{counts}{A numeric count matrix}

\item{...}{extra parameters passed onto the render_celda_heatmap}
}
\description{
celda_heatmap for celda Cell clustering function
}
