% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{celda_heatmap.celda_CG}
\alias{celda_heatmap.celda_CG}
\title{celda_heatmap for celda Cell and Gene clustering model}
\usage{
\method{celda_heatmap}{celda_CG}(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{A celda model object of "Celda_CG"}

\item{counts}{A count matrix}

\item{...}{extra parameters passed onto render_celda_heatmap}
}
\description{
celda_heatmap for celda Cell and Gene clustering model
}
