% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celda_heatmap.celda_G}
\alias{celda_heatmap.celda_G}
\title{celda_heatmap for celda Gene clustering function}
\usage{
\method{celda_heatmap}{celda_G}(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_G"}

\item{counts}{A numeric count matrix}
}
\description{
celda_heatmap for celda Gene clustering function
}
