% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{celda_heatmap}
\alias{celda_heatmap}
\title{Render a stylable heatmap of count data based on celda clustering results.}
\usage{
celda_heatmap(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{A celda model object (of class "celda_C", "celda_G" or "celda_CG")}

\item{counts}{the counts matrix}

\item{z}{A numeric vector of cluster assignments for cell. Resolved automatically from celda object when available.}

\item{y}{A numeric vector of cluster assignments for gene. Resolved automatically from celda object when available.}

\item{scale.log}{specify the transformation type of the matrix for (semi-)heatmap, can be "log","row"(z-acore by row),"col"(z-score by column), etc. #To be completed}

\item{scale.row}{specify the transformation type of the matrix for (semi-)heatmap, can be "log","row"(z-acore by row),"col"(z-score by column), etc. #To be completed}

\item{z.trim}{two element vector to specify the lower and upper cutoff of the z-score normalization result by default it is set to NULL so no trimming will be done.}

\item{scale_fun}{specify the function for scaling}

\item{cluster.row}{boolean values determining if rows should be clustered}

\item{cluster.column}{boolean values determining if columns should be clustered}
}
\description{
Render a stylable heatmap of count data based on celda clustering results.
}

