% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaCMod}
\alias{celdaCMod}
\title{celdaCMod}
\format{
A celda_C object
#@examples
data(celdaCSim)
celdaCMod <- celda_C(celdaCSim$counts, K = celdaCSim$K, nchains = 1)
}
\usage{
celdaCMod
}
\description{
Old celda_C results generated from celdaCSim
}
\keyword{datasets}
