% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaCSim}
\alias{celdaCSim}
\title{celdaCSim}
\format{
A list of counts and properties as returned from old simulateCells().
}
\usage{
data(celdaCSim)
}
\description{
An old example simulated count matrix from the celda_C model.
}
\keyword{datasets}
