% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaGMod}
\alias{celdaGMod}
\title{celdaGMod}
\format{
A celda_G object
#@examples
data(celdaGSim)
celdaGMod <- celda_G(celdaGSim$counts, L = celdaGSim$L, nchains = 1)
}
\usage{
celdaGMod
}
\description{
Old celda_G results generated from celdaGsim
}
\keyword{datasets}
