% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaGSim}
\alias{celdaGSim}
\title{celdaGSim}
\format{
A list of counts and properties as returned from old simulateCells()
}
\usage{
data(celdaGSim)
}
\description{
An old example simulated count matrix from the celda_G model.
}
\keyword{datasets}
