% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\docType{methods}
\name{celdaHeatmap,celda_C-method}
\alias{celdaHeatmap,celda_C-method}
\title{Heatmap for celda_C}
\usage{
\S4method{celdaHeatmap}{celda_C}(counts, celdaMod, featureIx, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Celda object of class `celda_C`.}

\item{featureIx}{Integer vector. Indices of features to plot, such the top
features from a differential expression analysis.}

\item{...}{Additional parameters.}
}
\value{
list A list containing dendrograms and the heatmap grob
}
\description{
Renders an expression heatmap to visualize `celda_C()` results.
 Features to include in the heatmap must be supplied.
}
\examples{
data(celdaCSim, celdaCMod)
celdaHeatmap(celdaCSim$counts, celdaCMod)
}
\seealso{
`celda_C()` for clustering cells and `celdaTsne()` for generating
 2-dimensional coordinates
}
