% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\docType{methods}
\name{celdaHeatmap,celda_CG-method}
\alias{celdaHeatmap,celda_CG-method}
\title{Heatmap for celda_CG}
\usage{
\S4method{celdaHeatmap}{celda_CG}(counts, celda.mod, nfeatures = 25, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class `celda_CG`.}

\item{nfeatures}{Integer. Maximum number of features to select for each module. Default 25.}

\item{...}{Additional parameters.}
}
\value{
A list containing dendrograms and the heatmap grob
}
\description{
Renders an expression heatmap to visualize `celda_CG()` results. The top `nfeatures` for each module will be included in the heatmap.
}
\examples{
celdaHeatmap(celda.CG.sim$counts, celda.CG.mod)
}
\seealso{
`celda_CG()` for clustering features and cells and `celdaTsne()` for generating 2-dimensional coordinates
}
