% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celdaHeatmap.celda_C}
\alias{celdaHeatmap.celda_C}
\title{celdaHeatmap for celda Cell clustering function}
\usage{
\method{celdaHeatmap}{celda_C}(counts, celda.mod, feature.ix, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C".}

\item{feature.ix}{Integer vector. Indices of features to plot, such the top features from a differential expression analysis.}

\item{...}{Additional parameters.}
}
\description{
celdaHeatmap for celda Cell clustering function
}
\examples{
celda.sim = simulateCells("celda_C")
celda.mod = celda_C(celda.sim$counts, K=celda.sim$K)
celdaHeatmap(celda.sim$counts, celda.mod)
}
