% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celdaHeatmap.celda_G}
\alias{celdaHeatmap.celda_G}
\title{celdaHeatmap for celda Gene clustering model}
\usage{
\method{celdaHeatmap}{celda_G}(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_G"}

\item{counts}{A numeric count matrix}

\item{...}{extra parameters passed onto renderCeldaHeatmap}
}
\description{
celdaHeatmap for celda Gene clustering model
}
