% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{celdaLogLik}
\alias{celdaLogLik}
\title{Plots the log likehood over iteration for each chain}
\usage{
celdaLogLik(celda.res, group.by, exclude.iter = 1:10, line.size = 0.5,
  legend.position = c("none", "left", "right", "bottom", "top"),
  scales = c("free_y", "free_x", "free", "fixed"))
}
\arguments{
\item{celda.res}{An object returned from the 'celda' function.}

\item{group.by}{A vector giving the variables to group the chains by.}

\item{exclude.iter}{Integers giving the indicies of iterations to excluce in the plotting. Default 1:10.}

\item{line.size}{Size of the lines for each chain.}

\item{legend.position}{Position of the color legend for the chains. Default "none".}

\item{scales}{Are scales shared across all facets (the default, ‘"fixed"’), or do they vary across rows (‘"free_x"’), columns (‘"free_y"’), or both rows and columns (‘"free"’). See ?ggplot2::facet_grid for more information.}
}
\description{
Plots the log likehood over iteration for each chain
}
