% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaPerplexity,celdaList-method}
\alias{celdaPerplexity,celdaList-method}
\title{Get perplexity for every model in a celdaList}
\usage{
\S4method{celdaPerplexity}{celdaList}(celdaList)
}
\arguments{
\item{celdaList}{An object of class celdaList.}
}
\value{
List. Contains one celdaModel object for each of the parameters
 specified in the `runParams()` of the provided celda list.
}
\description{
Returns perplexity for each model in a celdaList as calculated
 by `perplexity().`
}
\examples{
data(celdaCGGridSearchRes)
celdaCGGridModelPerplexities <- celdaPerplexity(celdaCGGridSearchRes)
}
