% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{celdaProbabilityMap,celda_CG-method}
\alias{celdaProbabilityMap,celda_CG-method}
\title{Probability map for a celda_CG model}
\usage{
\S4method{celdaProbabilityMap}{celda_CG}(
  counts,
  celdaMod,
  level = c("cellPopulation", "sample"),
  ...
)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Celda object of class `celda_CG`.}

\item{level}{Character. One of 'cellPopulation' or 'sample'.
'cellPopulation' will display the absolute probabilities and relative
normalized expression of each module in each cell population. 'sample'
will display the absolute probabilities and relative normalized abundance
of each cell population in each sample. Default 'cellPopulation'.}

\item{...}{Additional parameters.}
}
\value{
A grob containing the specified plots
}
\description{
Renders probability and relative expression heatmaps to
 visualize the relationship between features and cell populations or cell
 populations and samples.
}
\examples{
data(celdaCGSim, celdaCGMod)
celdaProbabilityMap(celdaCGSim$counts, celdaCGMod)
}
\seealso{
`celda_CG()` for clustering features and cells
}
