% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celdaTsne.celda_C}
\alias{celdaTsne.celda_C}
\title{Embeds cells in two dimensions using tSNE based on celda_C results.}
\usage{
\method{celdaTsne}{celda_C}(counts, celda.mod, max.cells = 25000,
  min.cluster.size = 100, initial.dims = 20, perplexity = 20,
  max.iter = 2500, seed = 12345, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C".}

\item{max.cells}{Integer. Maximum number of cells to plot. Cells will be randomly subsampled if ncol(counts) > max.cells. Larger numbers of cells requires more memory. Default 25000.}

\item{min.cluster.size}{Integer. Do not subsample cell clusters below this threshold. Default 100.}

\item{initial.dims}{Integer. PCA will be used to reduce the dimentionality of the dataset. The top 'initial.dims' principal components will be used for tSNE. Default 20.}

\item{perplexity}{Numeric. Perplexity parameter for tSNE. Default 20.}

\item{max.iter}{Integer. Maximum number of iterations in tSNE generation. Default 2500.}

\item{seed}{Integer. Passed to set.seed(). Default 12345.}

\item{...}{Additional parameters.}
}
\value{
A two column matrix of t-SNE coordinates
}
\description{
Embeds cells in two dimensions using tSNE based on celda_C results.
}
\examples{
celda.sim = simulateCells("celda_C")
celda.mod = celda_C(celda.sim$counts, K=celda.sim$K)
tsne.res = celdaTsne(celda.sim$counts, celda.mod)
}
