% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{celdaTsne.celda_CG}
\alias{celdaTsne.celda_CG}
\title{Embeds cells in two dimensions using tSNE based on celda_CG results.}
\usage{
\method{celdaTsne}{celda_CG}(counts, celda.mod, max.cells = 10000,
  min.cluster.size = 100, modules = NULL, perplexity = 20,
  max.iter = 2500, seed = 12345, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda model. Options available in `celda::available.models`.}

\item{max.cells}{Integer. Maximum number of cells to plot. Cells will be randomly subsampled if ncol(conts) > max.cells. Larger numbers of cells requires more memory. Default 10000.}

\item{min.cluster.size}{Integer. Do not subsample cell clusters below this threshold. Default 100.}

\item{modules}{Integer vector. Determines which features modules to use for tSNE. If NULL, all modules will be used. Default NULL.}

\item{perplexity}{Numeric. Perplexity parameter for tSNE. Default 20.}

\item{max.iter}{Integer. Maximum number of iterations in tSNE generation. Default 2500.}

\item{seed}{Integer. Passed to set.seed(). Default 12345.}

\item{...}{Additional parameters.}
}
\description{
Embeds cells in two dimensions using tSNE based on celda_CG results.
}
