% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celdaTsne.celda_G}
\alias{celdaTsne.celda_G}
\title{Embeds cells in two dimensions using tSNE based on celda_CG results.}
\usage{
\method{celdaTsne}{celda_G}(counts, celda.mod, states = NULL,
  perplexity = 20, max.iter = 2500, distance = "hellinger",
  seed = 12345, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_G".}

\item{states}{Numeric vector; determines which gene states to use for tSNE. If NULL, all states will be used. Default NULL.}

\item{perplexity}{Numeric vector; determines perplexity for tSNE. Default 20.}

\item{max.iter}{Integer. Maximum number of iterations of Gibbs sampling to perform. Default 1000.}

\item{distance}{Character. Determines which distance metric to use for tSNE. Options are 'hellinger', 'cosine', 'spearman', and 'euclidean'. Default 'hellinger'.}

\item{seed}{Integer. Passed to set.seed(). Default 12345.}

\item{...}{Additional parameters.}
}
\description{
Embeds cells in two dimensions using tSNE based on celda_CG results.
}
