% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{cG.calcLLFromVariables}
\alias{cG.calcLLFromVariables}
\title{Calculate Log Likelihood For A Set of Gene Clusterings (Gene Clustering)}
\usage{
cG.calcLLFromVariables(counts, y, L, beta, gamma, delta)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{y}{A numeric vector of gene cluster assignments}

\item{L}{The number of clusters being considered}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state}
}
\description{
This function calculates the log likelihood of each clustering of genes generated
over multiple iterations of Gibbs sampling.
}
\examples{
TODO
}
\keyword{likelihood}
\keyword{log}
