% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{cG.decomposeCounts}
\alias{cG.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cG.decomposeCounts(counts, y, L)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{y}{A numeric vector of gene cluster assignments}

\item{L}{The number of clusters being considered}
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
