% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{chainKs}
\alias{chainKs}
\title{Get the K value used for each chain in a celda run.}
\usage{
chainKs(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get the K value used for each chain in a celda run.
}
