% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{chainLs}
\alias{chainLs}
\title{Get the L value used for each chain in a celda run.}
\usage{
chainLs(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get the L value used for each chain in a celda run.
}
