% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{chainSeeds}
\alias{chainSeeds}
\title{Get seeds for all chains from a celda run.}
\usage{
chainSeeds(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get seeds for all chains from a celda run.
}

