% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_list.R
\name{chooseBestChain}
\alias{chooseBestChain}
\title{Determine the best chain among a set of celda_* objects with
otherwise uniform K/L choices.}
\usage{
chooseBestChain(celda.mods, method = "perplexity")
}
\arguments{
\item{celda.mods}{A list of celda class objects (celda_C, celda_CG, celda_G)}

\item{method}{How to choose the best chain. Choices are c("perplexity", "loglik"). Defaults to perplexity. "perplexity" calculates each chain's perplexity as the inverse of the geometric mean, per the original LDA description. "loglik" chooses the chain which reached the maximal log likelihood during Gibbs sampling.}
}
\description{
Determine the best chain among a set of celda_* objects with
otherwise uniform K/L choices.
}
