% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster.plot}
\alias{cluster.plot}
\title{cluster.plot}
\format{ggplot}
\source{
\url{http://github.com/compbiomed/celda}
}
\usage{
cluster.plot
}
\description{
The result of running the gettingClusters function on the pbmc_res model
}
\details{
Cluster plot created with the pbmc_res model.

Generated by Yusuke Koga
}
\keyword{datasets}
