% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{cluster_probability.celda_G}
\alias{cluster_probability.celda_G}
\title{Calculates the conditional probability of each cell belong to each cluster given all other cluster assignments}
\usage{
cluster_probability.celda_G(counts, celda.mod)
}
\arguments{
\item{counts}{The original count matrix used in the model}

\item{celda.mod}{A model returned from the 'celda_G' function}
}
\value{
A list containging a matrix for the conditional cell cluster probabilities.
}
\description{
Calculates the conditional probability of each cell belong to each cluster given all other cluster assignments
}
