% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{clustering}
\alias{clustering}
\title{Get clustering parameters and outcomes from a celda model.}
\usage{
clustering(celda.mod)
}
\value{
List. Contains K, z (for celda_C and celda_CG models), and/or L, y (for celda_G and celda_CG models.)
}
\description{
Returns the K/L parameters provided for modeling, as well as the corresponding z/y results.
}
\examples{
clustering(celda.CG.mod)
}
