% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{clusterProbabilities.celda_C}
\alias{clusterProbabilities.celda_C}
\title{clusterProbabilities for celda Cell clustering function}
\usage{
\method{clusterProbabilities}{celda_C}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_C"}
}
\description{
clusterProbabilities for celda Cell clustering function
}
