% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{clusterProbabilities.celda_CG}
\alias{clusterProbabilities.celda_CG}
\title{clusterProbabilities for the celda Cell and Gene clustering function}
\usage{
\method{clusterProbabilities}{celda_CG}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of "Celda_CG"}
}
\description{
clusterProbabilities for the celda Cell and Gene clustering function
}
