% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{clusterProbabilities.celda_G}
\alias{clusterProbabilities.celda_G}
\title{clusterProbabilities for celda Gene clustering model}
\usage{
\method{clusterProbabilities}{celda_G}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_G"}
}
\description{
clusterProbabilities for celda Gene clustering model
}
