% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{clusterProbabilities}
\alias{clusterProbabilities}
\title{Get the probability of the cluster assignments generated during a celda run.}
\usage{
clusterProbabilities(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object (of class "celda_C", "celda_G" or "celda_CG")}
}
\description{
Get the probability of the cluster assignments generated during a celda run.
}

