% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\docType{methods}
\name{clusterProbability,celda_G-method}
\alias{clusterProbability,celda_G-method}
\title{Conditional probabilities for features in modules from a Celda_G model}
\usage{
\S4method{clusterProbability}{celda_G}(counts, celda.mod, log = FALSE,
  modules = NULL, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class `celda_G`.}

\item{log}{Logical. If FALSE, then the normalized conditional probabilities will be returned. If TRUE, then the unnormalized log probabilities will be returned. Default FALSE.}

\item{...}{Additional parameters.}
}
\value{
A list containging a matrix for the conditional cell cluster probabilities.
}
\description{
Calculates the conditional probability of each feature belonging to each module given all other feature cluster assignments in a `celda_G()` result.
}
\examples{
cluster.prob = clusterProbability(celda.G.sim$counts, celda.G.mod)
}
\seealso{
`celda_G()` for clustering features
}
