% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{clusters,celdaModel-method}
\alias{clusters,celdaModel-method}
\title{Get clustering outcomes from a celda model}
\usage{
\S4method{clusters}{celdaModel}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda model. Options available in `celda::available.models`.}
}
\value{
List. Contains z (for celda_C and celda_CG models) and/or y (for celda_G and celda_CG models)
}
\description{
Returns the z / y results corresponding to the cell / gene cluster labels determined by the provided celda model.
}
\examples{
clusters(celda.CG.mod)
}
