% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{clusters}
\alias{clusters}
\alias{clusters,SingleCellExperiment-method}
\alias{clusters<-}
\alias{clusters<-,SingleCellExperiment-method}
\title{Get or set the cell cluster labels from a celda
 \link[SingleCellExperiment]{SingleCellExperiment} object}
\usage{
clusters(sce)

\S4method{clusters}{SingleCellExperiment}(sce)

clusters(sce) <- value

\S4method{clusters}{SingleCellExperiment}(sce) <- value
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
returned by \link{celda_C}, or \link{celda_CG}.}
}
\value{
List. Contains cell cluster z (for celda_C and celda_CG Models)
 and/or gene module y (for celda_G and celda_CG Models).
}
\description{
Return or set the cell cluster labels determined
 by \link{celda_C} or \link{celda_CG} models.
}
\examples{
data(sceCelda_CG)
clusters(sceCelda_CG)
}
