% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{compare_count_matrix}
\alias{compare_count_matrix}
\title{Check whether a count matrix was the one used in a given celda run}
\usage{
compare_count_matrix(count.matrix, celda.checksum)
}
\arguments{
\item{count.matrix}{A numeric matrix of counts}

\item{celda.checksum}{An MD5 checksum from a celda_list or celda model object (celda_C, celda_G, celda_CG)}
}
\value{
TRUE if provided count matrix matches the one used in the celda run, FALSE otherwise
}
\description{
Compare the MD5 checksum of a provided count.matrix to the count matrix
checksum on a celda_list object, to see if they're the same.
}
