% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{completeClusterHistory.celda_C}
\alias{completeClusterHistory.celda_C}
\title{completeClusterHistory for celda Cell clustering funciton}
\usage{
\method{completeClusterHistory}{celda_C}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_C"}
}
\description{
completeClusterHistory for celda Cell clustering funciton
}
