% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{completeClusterHistory.celda_CG}
\alias{completeClusterHistory.celda_CG}
\title{completeClusterHistory for the celda Cell and Gene clustering function}
\usage{
\method{completeClusterHistory}{celda_CG}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of "Celda_CG"}
}
\description{
completeClusterHistory for the celda Cell and Gene clustering function
}
