% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{completeClusterHistory.celda_G}
\alias{completeClusterHistory.celda_G}
\title{completeClusterHistory for celda Gene clustering function}
\usage{
\method{completeClusterHistory}{celda_G}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_G"}
}
\description{
completeClusterHistory for celda Gene clustering function
}
