% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{completeClusterHistory}
\alias{completeClusterHistory}
\title{Get the complete history of gene / cell / gene & cell cluster assignments 
generated during a celda run, dependent on the model provided.}
\usage{
completeClusterHistory(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object (of class "celda_C", "celda_G" or "celda_CG")}
}
\description{
Get the complete history of gene / cell / gene & cell cluster assignments 
generated during a celda run, dependent on the model provided.
}
