% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{completeLogLik}
\alias{completeLogLik}
\title{Get log-likelihood history}
\usage{
completeLogLik(celda.mod)
}
\value{
Numeric. The log-likelihood at each step of Gibbs sampling used to generate the model.
}
\description{
Retrieves the complete log-likelihood from all iterations of Gibbs sampling used to generate a celda model.
}
\examples{
completeLogLik(celda.CG.mod)
}
