% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contamination.sim}
\alias{contamination.sim}
\title{contamination.sim}
\format{A list}
\usage{
contamination.sim
}
\description{
Generated by simulateObservedMatrix
}
\details{
A toy contamination data generated by simulateObservedMatrix
}
\keyword{datasets}
