% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{countChecksum,celdaList-method}
\alias{countChecksum,celdaList-method}
\title{Get the MD5 hash of the count matrix from the celdaList}
\usage{
\S4method{countChecksum}{celdaList}(celdaList)
}
\arguments{
\item{celdaList}{An object of class celdaList.}
}
\value{
A character string of length 32 containing the MD5 digest of
 the count matrix.
}
\description{
Returns the MD5 hash of the count matrix used to generate the
 celdaList.
}
\examples{
data(celdaCGGridSearchRes)
countChecksum <- countChecksum(celdaCGGridSearchRes)
}
