% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{countChecksum}
\alias{countChecksum}
\title{Get count matrix checksum for comparison}
\usage{
countChecksum(celda.mod)
}
\value{
Character. The MD5 hash of the count matrix used to generate the provided celda model.
}
\description{
Retrieves the MD5 checksum of the count matrix used to generate the provided celda mdoel.
}
\examples{
countChecksum(celda.CG.mod)
}
