% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{countsTimesProbs}
\alias{countsTimesProbs}
\alias{.countsTimesProbs}
\title{Counts matrix times cell population probabilies}
\usage{
.countsTimesProbs(counts, phi)
}
\arguments{
\item{counts}{feature-by-cell matrix}

\item{phi}{feature-by-probability matrix}
}
\description{
Counts matrix times cell population probabilies
}
