% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{decon.calcLL}
\alias{decon.calcLL}
\title{This function calculates the log-likelihood}
\usage{
decon.calcLL(omat, z, phi, eta, theta)
}
\arguments{
\item{omat}{Numeric/Integer matrix. Observed count matrix, rows represent features and columns represent cells}

\item{z}{Integer vector. Cell population labels}

\item{phi}{Numeric matrix. Rows represent features and columns represent cell populations}

\item{eta}{Numeric matrix. Rows represent features and columns represent cell populations}

\item{theta}{Numeric vector. Proportion of truely expressed transcripts}
}
\description{
This function calculates the log-likelihood
}
