% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{DeconX}
\alias{DeconX}
\title{This function updates decontamination}
\usage{
DeconX(omat, z, max.iter = 200, beta = 1e-06, delta = 10)
}
\arguments{
\item{omat}{Numeric/Integer matrix. Observed count matrix, rows represent features and columns represent cells.}

\item{z}{Integer vector. Cell population labels.}

\item{max.iter}{Integer. Maximum iterations of EM algorithm. Default to be 200}

\item{beta}{Numeric. Concentration parameter for Phi. Default to be 1e-6}

\item{delta}{Numeric / Numeric vector. Concentration parameter for Theta. Default to be 10}
}
\description{
This function updates decontamination
}
