% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExp.R
\name{differentialExpression}
\alias{differentialExpression}
\alias{differentialExpression,SingleCellExperiment-method}
\alias{differentialExpression,matrix-method}
\title{Differential expression for cell subpopulations using MAST}
\usage{
differentialExpression(x, ...)

\S4method{differentialExpression}{SingleCellExperiment}(
  x,
  useAssay = "counts",
  altExpName = "featureSubset",
  c1,
  c2 = NULL,
  onlyPos = FALSE,
  log2fcThreshold = NULL,
  fdrThreshold = 1
)

\S4method{differentialExpression}{matrix}(
  x,
  celdaMod,
  c1,
  c2 = NULL,
  onlyPos = FALSE,
  log2fcThreshold = NULL,
  fdrThreshold = 1
)
}
\arguments{
\item{x}{A numeric \link{matrix} of counts or a
\linkS4class{SingleCellExperiment}
with the matrix located in the assay slot under \code{useAssay}.
Rows represent features and columns represent cells. Must contain cluster
labels in \code{celdaClusters(x, altExpName = altExpName)} if \code{x} is a
\linkS4class{SingleCellExperiment} object.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use if \code{x} is a
\link[SingleCellExperiment]{SingleCellExperiment} object. Default "counts".}

\item{altExpName}{The name for the \link[SingleCellExperiment]{altExp} slot
to use. Default "featureSubset".}

\item{c1}{Integer vector. Cell populations to include in group 1 for the
differential expression analysis.}

\item{c2}{Integer vector. Cell populations to include in group 2 for the
differential expression analysis. If NULL, the clusters in the c1 group are
compared to all other clusters. Default NULL.}

\item{onlyPos}{Logical. Whether to only return markers with positive log2
fold change. Default FALSE.}

\item{log2fcThreshold}{Numeric. A number greater than 0 that specifies the
absolute log2 fold change threshold. Only features with absolute value
above this threshold will be returned. If NULL, this filter will not be
applied. Default NULL.}

\item{fdrThreshold}{Numeric. A number between 0 and 1 that specifies the
false discovery rate (FDR) threshold. Only features below this threshold
will be returned. Default 1.}

\item{celdaMod}{Celda object of class `celda_C` or `celda_CG`.}
}
\value{
Data frame containing MAST results including statistics such as
 p-value, log2 fold change, and FDR.
}
\description{
Uses MAST to find differentially expressed features for
 specified cell subpopulations.
}
\examples{
data(sceCeldaCG)
clusterDiffexpRes <- differentialExpression(sceCeldaCG, c1 = c(1, 2))
data(celdaCGSim, celdaCGMod)
clusterDiffexpRes <- differentialExpression(celdaCGSim$counts,
  celdaCGMod,
  c1 = c(1, 2))
}
