% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{distinct_colors}
\alias{distinct_colors}
\title{Generate a distinct palette for coloring clusters}
\usage{
distinct_colors(n, hues = c("red", "cyan", "orange", "blue", "yellow",
  "purple", "green", "magenta"), saturation.range = c(0.4, 1),
  value.range = c(0.4, 1))
}
\arguments{
\item{n}{Integer; Number of colors to generate}

\item{saturation.range}{Numeric vector of length 2 with values between 0 and 1. Default: c(0.25, 1)}

\item{value.range}{Numeric vector of length 2 with values between 0 and 1. Default: c(0.5, 1)}
}
\value{
A vector of distinct colors in HEX
}
\description{
Generate a distinct palette for coloring clusters
}
