% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMultInt}
\alias{eigenMatMultInt}
\title{Fast matrix multiplication for double x int}
\usage{
eigenMatMultInt(A, B)
}
\arguments{
\item{A}{Numeric matrix. Dimension G x K matrix of genes by cell population probabilities.}

\item{B}{Integer matrix. Dimensions G x M. Matrix of gene counts by cell.}
}
\value{
An integer matrix representing the product of A and B
}
\description{
Fast matrix multiplication for double x int
}
