% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMultNumeric}
\alias{eigenMatMultNumeric}
\title{Fast matrix multiplication for double x double}
\usage{
eigenMatMultNumeric(A, B)
}
\arguments{
\item{A}{a double matrix}

\item{B}{an integer matrix}
}
\value{
An integer matrix representing the product of A and B
}
\description{
Fast matrix multiplication for double x double
}
