% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{factorizeMatrix.celda_CG}
\alias{factorizeMatrix.celda_CG}
\title{factorizeMatrix for celda Cell and Gene clustering function}
\usage{
factorizeMatrix.celda_CG(counts, celda.obj, type = c("counts", "proportion",
  "posterior"))
}
\arguments{
\item{counts}{A numerix count matrix}

\item{celda.obj}{object returned from celda_CG function}

\item{type}{one of the "counts", "proportion", or "posterior".}
}
\value{
A list of factorized matrices, of the types requested by the user. NOTE: "population" state matrices are always returned in cell population (rows) x transcriptional states (cols).
}
\description{
factorizeMatrix for celda Cell and Gene clustering function
}
