% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastNormPropLog}
\alias{fastNormPropLog}
\title{Fast normalization for numeric matrix}
\usage{
fastNormPropLog(R_counts, R_alpha)
}
\arguments{
\item{R_counts}{An integer matrix}

\item{R_alpha}{A double value to be added to the matrix as a pseudocount}
}
\value{
A numeric matrix where the columns have been normalized to proportions
}
\description{
Fast normalization for numeric matrix
}
