% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{featureModuleLookup,celda_C-method}
\alias{featureModuleLookup,celda_C-method}
\title{Lookup the module of a feature}
\usage{
\S4method{featureModuleLookup}{celda_C}(counts, celdaMod, feature, exactMatch = TRUE)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Model of class `celda_C`.}

\item{feature}{Character vector. The module assignemnts will be found for
feature names in this vector.}

\item{exactMatch}{Logical. Whether an exact match or a partial match using
`grep()` is required to look up the feature in the rownames of the counts
matrix. Default TRUE.}
}
\value{
List. Each element contains the module of the provided feature.
}
\description{
Finds the module assignments of given features in a `celda_C()`
 model.
}
