% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{featureModuleLookup.celda_CG}
\alias{featureModuleLookup.celda_CG}
\title{Obtain the feature module of a feature of interest}
\usage{
\method{featureModuleLookup}{celda_CG}(counts, celda.mod, feature,
  exact.match = TRUE)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Model of class "celda_G" or "celda_CG".}

\item{feature}{Character vector. Identify feature modules for the specified feature names.}

\item{exact.match}{Logical. Whether to look for exact match of the gene name within counts matrix. Default TRUE.}
}
\value{
List. Each entry corresponds to the feature module determined for the provided features
}
\description{
This function will output the gene module of a specific gene(s) from a celda model
}
\examples{
module = featureModuleLookup(celda.CG.sim$counts, celda.CG.mod, 
                             c("Gene_1", "Gene_XXX"))
}
