% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_list.R
\name{filterCeldaList}
\alias{filterCeldaList}
\title{Select models from a list of results generated in a celda run.}
\usage{
filterCeldaList(celda.list, K = c(), L = c(), chain = c(), index = NULL)
}
\arguments{
\item{celda.list}{Object of class "celda_list". An object containing celda models returned from `celdaGridSearch()`.}

\item{K}{The K parameter for the desired model in the results list. Matches all K by default. Accepts ranges.}

\item{L}{The L parameter for the desired model in the results list. Matches all L by default. Accepts ranges.}

\item{chain}{The desired chain(s) for the specified model, for the specified K/L. Matches all chains by default. Accepts ranges.}

\item{index}{The index of the desired model in the run.parameters in the provided celda_list. Overrides all other parameters if provided. Defaults to NULL.}
}
\value{
A celda model object matching the provided parameters, or a list of celda model objects if multiple models were matched (of class "celda_C", "celda_G", "celda_CG" accordingly), or NA if one is not found.
}
\description{
Convenience function for picking out specific models from a celda_list object.
Models can be selected by various parameters, most importantly the K/L parameters (number of cell
 clusters / number of feature clusters).
}
\examples{
celda.mods = celdaGridSearch(celda::pbmc_select, model="celda_CG", K.to.test=c(10,15), 
                      L=c(50, 100), max.iter=2, nchains=1)
desired.mod = filterCeldaList(celda.mods, K=10, L=100)
}
