% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{finalClusterAssignment.celda_C}
\alias{finalClusterAssignment.celda_C}
\title{finalClusterAssignment for celda Cell clustering funciton}
\usage{
\method{finalClusterAssignment}{celda_C}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C".}
}
\description{
finalClusterAssignment for celda Cell clustering funciton
}
