% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{finalClusterAssignment}
\alias{finalClusterAssignment}
\title{Get the final feature and/or cell cluster assignments determined by a
celda model.}
\usage{
finalClusterAssignment(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}
}
\value{
If celda.mod is of class "celda_C" or "celda_G", the cell / feature cluster assignments, respectively. If celda.mod is of class "celda_CG", a list containing both the cell and feature cluster assignments.
}
\description{
Get the final feature and/or cell cluster assignments determined by a
celda model.
}
\examples{
celda.mod = celda_CG(celda::pbmc_select, K=10, L=50)
final.clustering = finalClusterAssignment(celda.mod)
cell.clusters = final.clustering$z
feature.clusters = final.clustering$y
}
