% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{finalLogLiklihood}
\alias{finalLogLiklihood}
\title{Get the log likelihood from the final iteration of Gibbs sampling
for each chain generated during a celda run.}
\usage{
finalLogLiklihood(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get the log likelihood from the final iteration of Gibbs sampling
for each chain generated during a celda run.
}
