% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{generateCells_gene_clustering}
\alias{generateCells_gene_clustering}
\title{Generate Count Data}
\usage{
generateCells_gene_clustering(C = 100, N.Range = c(500, 5000), G = 1000,
  k = 5, a = 1, b = 1, g = 1, seed = 12345)
}
\arguments{
\item{C}{The number of cells}

\item{N.Range}{The range of counts each gene should have}

\item{G}{The number of genes for which to simulate counts}

\item{a}{Vector of non-zero concentration parameters for sample <-> cluster assignment Dirichlet distribution}

\item{b}{Vector of non-zero concentration parameters for cluster <-> gene assignment Dirichlet distribution}

\item{g}{The number of cell states ("topics")' @param k The number of gene clusters to simulate from}

\item{seed}{Parameter to set.seed() for random number generation}
}
\description{
Generate a simulated count matrix, based off a generative distribution whose 
parameters can be provided by the user.
}
\examples{
TODO
}

