% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_list.R
\name{getBestModel}
\alias{getBestModel}
\title{Select the best model from a celda_list, as determined by final log-likelihood.}
\usage{
getBestModel(celda.list, K, L)
}
\arguments{
\item{celda.list}{A celda_list object returned from celda()}

\item{K}{Limit search for best model to models with this number of cell clusters. Defaults to NULL.}

\item{L}{Limit search for best model to models with this number of gene clusters. Defaults to NULL.}
}
\value{
The celda model object with the highest finalLogLik attribute, meeting any K/L criteria provided
}
\description{
This function returns the celda model (celda_C, celda_G, celda_CG) from a celda_list
with the maxiumim final log-likelihood. If a K or L (or combination) parameter is provided,
the model with these K/L and the highest log-likelihood is returned.
}
