% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{getK.celda_C}
\alias{getK.celda_C}
\title{getK for celda Cell clustering function}
\usage{
\method{getK}{celda_C}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C".}
}
\description{
getK for celda Cell clustering function
}
