% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{getK.celda_CG}
\alias{getK.celda_CG}
\title{getK for the celda Cell and Gene clustering model}
\usage{
\method{getK}{celda_CG}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of "Celda_CG"}
}
\description{
getK for the celda Cell and Gene clustering model
}
