% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{getK.celda_G}
\alias{getK.celda_G}
\title{getK for celda Gene clustering model}
\usage{
\method{getK}{celda_G}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_G".}
}
\description{
getK for celda Gene clustering model
}
