% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{getL.celda_C}
\alias{getL.celda_C}
\title{getL for celda Cell clustering function}
\usage{
\method{getL}{celda_C}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_C"}
}
\description{
getL for celda Cell clustering function
}
