% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{getL.celda_G}
\alias{getL.celda_G}
\title{Get the L (number of feature clusters) parameter provided to a celda model.}
\usage{
\method{getL}{celda_G}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_G".}
}
\value{
Integer. The L provided to the model during initialization.
}
\description{
Get the L (number of feature clusters) parameter provided to a celda model.
}
\examples{
celda.mod = celda_G(celda::pbmc_select, L=50)
mod.l.value = getL(celda.mod)
}
