% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_list.R
\name{getModel}
\alias{getModel}
\title{Select the best model amongst a list of models generated in a celda run.}
\usage{
getModel(celda.list, K = NULL, L = NULL, chain = NULL, best = "loglik")
}
\arguments{
\item{celda.list}{A celda_list object returned from celda()}

\item{K}{The K parameter for the desired model in the results list. Defaults to NULL.}

\item{L}{The L parameter for the desired model in the results list. Defaults to NULL.}

\item{chain}{The desired chain for the specified model. Defaults to NULL. Overrides best parameter if provided.}

\item{best}{Method for choosing best chain automatically. Options are c("perplexity", "loglik"). See documentation for chooseBestModel for details. Defaults to "loglik."}
}
\value{
A celda model object matching the provided parameters (of class "celda_C", "celda_G", "celda_CG" accordingly), or NA if one is not found.
}
\description{
Select the best model amongst a list of models generated in a celda run.
}
