% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{getPerplexity}
\alias{getPerplexity}
\title{Get perplexity for every model in a celdaList}
\usage{
getPerplexity(celda.mod)
}
\value{
List. Contains one celdaModel object for each of the parameters specified in the `runParams()` of the provided celda list.
}
\description{
Returns perplexity for each model in a celdaList as calculated by `perplexity().`
}
\examples{
celda.CG.grid.model.perplexities = getPerplexity(celda.CG.grid.search.res)
}
