% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettingClusters.R
\name{gettingClusters}
\alias{gettingClusters}
\title{Assists in selection of K/L parameter for downstream analysis.}
\usage{
gettingClusters(celda.list, matrix, iterations)
}
\arguments{
\item{celda.list}{celda.list object. Ouput of celda function.}

\item{matrix}{matrix. Counts matrix that was inputted to celda function. Do not normalize prior to using function.}

\item{iterations}{Numeric. Number of iterations to run the function. A higher number will generally output a smoother plot}
}
\description{
Assists in selection of K/L parameter for downstream analysis.
}
