% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{GiniPlot}
\alias{GiniPlot}
\title{Create a bar chart based on Gini coefficients of transcriptional states}
\usage{
GiniPlot(counts, celda.mod, cell_clusters = NULL, label_size = 4,
  bar_col = NULL)
}
\arguments{
\item{counts}{A numeric count matrix.}

\item{celda.mod}{An object of class celda_CG.}

\item{cell_clusters}{A list of cell clusters. Calculate Gini coefficients using cells from specific cell clusters. If NULL, uses all cells (NULL by default).}

\item{label_size}{Numeric. The size of labels of transcrptional states on top of bars (4 by default).}

\item{bar_col}{The color of bars (#FF8080FF by default).}
}
\description{
A bar plot shows the ordered Gini coefficients of transcriptional states. The transcriptional states with high Gini coefficients (i.e. close to 1) 
   indicates that only a few cell clusters highly express the genes in these transcriptional states. It is easier for users to define cell types based on these
   genes in these significant states.
}
