% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{initialSeed}
\alias{initialSeed}
\title{Get seed used to generate model}
\usage{
initialSeed(celda.mod)
}
\value{
Numeric. The random seed used to generate the provided celda model.
}
\description{
Retrieves the random seed used to generate a celda model.
}
\examples{
initialSeed(celda.CG.mod)
}
