% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{logLikelihood.celda_G}
\alias{logLikelihood.celda_G}
\title{Calculate the celda_G log likelihood for user-provided cluster assignments}
\usage{
logLikelihood.celda_G(counts, y, L, beta, delta, gamma)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{y}{Numeric vector. Denotes feature module labels.}

\item{L}{Integer. Number of feature modules.}

\item{beta}{Numeric. Concentration parameter for Phi. Adds a pseudocount to each feature module in each cell. Default 1.}

\item{delta}{Numeric. Concentration parameter for Psi. Adds a pseudocount to each feature in each module. Default 1.}

\item{gamma}{Numeric. Concentration parameter for Eta. Adds a pseudocount to the number of features in each module. Default 1.}

\item{...}{Additional parameters.}
}
\value{
The log-likelihood for the given cluster assignments
}
\description{
This function calculates the log likelihood of each clustering of genes generated
over multiple iterations of Gibbs sampling.
}
\examples{
celda.sim = simulateCells(model="celda_G")
loglik = logLikelihood(celda.sim$counts, model="celda_G", 
                       y=celda.sim$y, L=celda.sim$L,
                       beta=celda.sim$beta, delta=celda.sim$delta,
                       gamma=celda.sim$gamma)
}
\keyword{likelihood}
\keyword{log}
