% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{logLikelihoodcelda_G}
\alias{logLikelihoodcelda_G}
\title{Calculate Celda_G log likelihood}
\usage{
logLikelihoodcelda_G(counts, y, L, beta, delta, gamma)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells.}

\item{y}{Numeric vector. Denotes feature module labels.}

\item{L}{Integer. Number of feature modules.}

\item{beta}{Numeric. Concentration parameter for Phi. Adds a pseudocount to
each feature module in each cell. Default 1.}

\item{delta}{Numeric. Concentration parameter for Psi. Adds a pseudocount to
each feature in each module. Default 1.}

\item{gamma}{Numeric. Concentration parameter for Eta. Adds a pseudocount to
the number of features in each module. Default 1.}
}
\value{
The log-likelihood for the given cluster assignments.
}
\description{
Calculates the log likelihood for user-provided feature module
 clusters using the `celda_G()` model.
}
\examples{
data(celdaGSim)
loglik <- logLikelihoodcelda_G(celdaGSim$counts,
  y = celdaGSim$y,
  L = celdaGSim$L,
  beta = celdaGSim$beta,
  delta = celdaGSim$delta,
  gamma = celdaGSim$gamma
)

loglik <- logLikelihood(celdaGSim$counts,
  model = "celda_G",
  y = celdaGSim$y,
  L = celdaGSim$L,
  beta = celdaGSim$beta,
  delta = celdaGSim$delta,
  gamma = celdaGSim$gamma
)
}
\seealso{
`celda_G()` for clustering features
}
\keyword{likelihood}
\keyword{log}
