% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{lookupGeneModule}
\alias{lookupGeneModule}
\title{Obtain the gene module of a gene of interest}
\usage{
lookupGeneModule(counts, model, gene)
}
\arguments{
\item{counts}{Counts matrix that the celda model was run on}

\item{model}{Celda model, class of "celda_CG" or "celda_G"}

\item{gene}{Genes to check the gene module}
}
\description{
This function will output the gene module of a specific gene(s) from a celda model
}
