% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{lookupTranscriptionalStateofGene}
\alias{lookupTranscriptionalStateofGene}
\title{Obtain the transcriptional state of a gene of interest}
\usage{
lookupTranscriptionalStateofGene(counts, model, gene)
}
\arguments{
\item{counts}{Counts matrix that the celda model was run on}

\item{model}{Celda model, class of "celda_CG" or "celda_G"}

\item{gene}{Genes to check the transcriptional state}
}
\description{
This function will output the transcriptional state of a specific gene(s) from a celda model
}
