% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{matrixNames,celdaModel-method}
\alias{matrixNames,celdaModel-method}
\title{Get feature, cell and sample names from a celda model}
\usage{
\S4method{matrixNames}{celdaModel}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda model. Options available in `celda::available.models`.}
}
\value{
List. Contains row, column, and sample character vectors corresponding to the values provided when the celda model was generated.
}
\description{
Retrieves the row, column, and sample names used to generate a celda model.
}
\examples{
matrixNames(celda.CG.mod)
}
