% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{modelPriors}
\alias{modelPriors}
\title{Get model prior parameters from a celda model.}
\usage{
modelPriors(celda.mod)
}
\value{
List. Contains alpha, beta (for celda_C and celda_CG models), or delta, gamma (for celda_G and celda_CG models).
}
\description{
Returns the model priors (e.g. alpha, beta) provided at model creation for a given celda model.
}
\examples{
modelPriors(celda.CG.mod)
}
