% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{modules}
\alias{modules}
\alias{modules,SingleCellExperiment-method}
\alias{modules<-}
\alias{modules<-,SingleCellExperiment-method}
\title{Get or set the feature module labels from a celda
 \link[SingleCellExperiment]{SingleCellExperiment} object}
\usage{
modules(sce)

\S4method{modules}{SingleCellExperiment}(sce)

modules(sce) <- value

\S4method{modules}{SingleCellExperiment}(sce) <- value
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
returned by \link{celda_G} or \link{celda_CG}.}
}
\value{
List. Contains feature module y (for celda_G and celda_CG Models).
}
\description{
Return or set the feature module cluster labels determined
 by \link{celda_G} or \link{celda_CG} models.
}
\examples{
data(sceCelda_CG)
modules(sceCelda_CG)
}
