% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nonzero}
\alias{nonzero}
\title{get row and column indices of none zero elements in the matrix}
\usage{
nonzero(R_counts)
}
\arguments{
\item{R_counts}{A matrix}
}
\value{
An integer matrix where each row is a row, column indices pair
}
\description{
get row and column indices of none zero elements in the matrix
}
