% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_select}
\alias{pbmc_select}
\title{pbmc_select}
\format{The rows are organized by gene names, while the
columns are organized by barcodes. For this tutorial, the dataset has been
slightly modified so the rownames are comprised of the Ensembl gene ID as
well as the gene name.}
\source{
\url{https://support.10xgenomics.com/single-cell/software/pipelines/latest/rkit}
}
\usage{
pbmc_select
}
\description{
A dataset of 3000 Peripheral Blood Mononuclear Cells (PBMC),
available from 10X Genomics.  The raw dataset can be found at the URL below.
}
\details{
3000 Peripheral Blood Mononuclear Cells (PBMC) from 10X Genomics.
}
\keyword{datasets}
