% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plot_dr_cluster}
\alias{plot_dr_cluster}
\title{Create a scatterplot based on celda cluster labels.}
\usage{
plot_dr_cluster(dim1, dim2, cluster, xlab = "Dimension_1",
  ylab = "Dimension_2")
}
\arguments{
\item{dim1}{Numeric vector; first dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector; second dimension from data dimensionality reduction output.}

\item{cluster}{Vector; Contains cluster labels (e.g. from celda_C or celda_CG).}

\item{xlab}{Character vector, used as label for rows. Default "Dimension_1".}

\item{ylab}{Character vector, used as label for columns. Default "Dimension_2".}
}
\description{
Create a scatterplot based on celda cluster labels.
}
