% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plot_dr_state}
\alias{plot_dr_state}
\title{Create a scatterplot based off of a matrix containing the celda state probabilities per cell.}
\usage{
plot_dr_state(dim1, dim2, matrix, rescale = FALSE, xlab = "Dimension_1",
  ylab = "Dimension_2", color_low = "grey", color_high = "blue")
}
\arguments{
\item{dim1}{Numeric vector; first dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector; second dimension from data dimensionality reduction output.}

\item{matrix}{Cell state probability matrix, will have cell name for column name and state probability for row name.}

\item{rescale}{Boolean. If TRUE z-score normalize the matrix. Default FALSE.}

\item{xlab}{Character vector, used as label for x axis. Default "Dimension_1".}

\item{ylab}{Character vector, used as label for y axis. Default "Dimension_2".}

\item{color_low}{Character vector of R colors available from the colors() function. The color will be used to signify the lowest values on the scale. Default: 'grey'}

\item{color_high}{Character vector of R colors available from the colors() function. The color will be used to signify the highest values on the scale. Default: 'blue'}
}
\description{
Create a scatterplot based off of a matrix containing the celda state probabilities per cell.
}
