% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decontx.R
\name{plotDecontXMarkerExpression}
\alias{plotDecontXMarkerExpression}
\title{Plots expression of marker genes before and after decontamination}
\usage{
plotDecontXMarkerExpression(x, markers, ...)
}
\arguments{
\item{x}{Either a \linkS4class{SingleCellExperiment} or
a matrix-like object of counts.}

\item{markers}{List or vector of marker genes to plot.}

\item{...}{Additional arguments passed to
\code{\link[decontX]{plotDecontXMarkerExpression}} (for example
\code{groupClusters}, \code{assayName}, \code{z}, \code{exactMatch},
\code{by}, \code{log1p}, \code{ncol}, \code{plotDots}, \code{dotSize}).}
}
\value{
Returns a \code{ggplot} object.
}
\description{
Generates a violin plot that shows the expression of marker
genes across clusters or cell types. Can be used to view the expression of
marker genes in different cell types before and after decontamination with
\code{\link{decontX}}. As of celda 1.23.0 this is a thin wrapper around
\code{\link[decontX]{plotDecontXMarkerExpression}} in the \pkg{decontX}
package, which must be installed.
}
\seealso{
\code{\link[decontX]{plotDecontXMarkerExpression}}. See
\code{\link{decontX}} for a full example of how to estimate and plot
contamination.
}
\author{
Shiyi Yang, Joshua Campbell
}
