% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendro.R
\name{plotDendro}
\alias{plotDendro}
\title{Plots dendrogram of `buildTreeHybrid` output}
\usage{
plotDendro(decisionTree, classLabel = NULL, addSensPrec = TRUE,
  leafSize = 24, boxSize = 7, boxColor = "black")
}
\arguments{
\item{classLabel}{A character value. The name of a label to which draw
the path and rules. If NULL (default), the rules for every cluster is shown.}

\item{addSensPrec}{Logical. Print training sensitivities and precisions
for each cluster below leaf label? Default is TRUE.}

\item{leafSize}{A numeric value. Size of text below each leaf. Default is 24.}

\item{boxSize}{A numeric value. Size of rule labels. Default is 7.}

\item{boxColor}{A character value. Color of rule labels. Default is `black`.}

\item{counts}{List object. The output of `celda::buildTreeHybrid`.}
}
\value{
A ggplot2 object
}
\description{
Generates a dendrogram of the rules and performance
(optional) of the decision tree generates by `buildTreeHybrid`.
}
\examples{
data(celdaCGSim, celdaCGMod)
plotHeatmap(celdaCGSim$counts,
    z = clusters(celdaCGMod)$z, y = clusters(celdaCGMod)$y)
}
