% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceState}
\alias{plotDimReduceState}
\title{Create a scatterplot based off of a matrix containing the celda state probabilities per cell.}
\usage{
plotDimReduceState(dim1, dim2, counts, celda.mod, modules = NULL,
  rescale = TRUE, size = 1, xlab = "Dimension_1", ylab = "Dimension_2",
  color_low = "grey", color_mid = NULL, color_high = "blue")
}
\arguments{
\item{dim1}{Numeric vector. First dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector. Second dimension from data dimensionality reduction output.}

\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_G" or "celda_CG".}

\item{modules}{Character vector. Module(s) from celda model to be plotted.}

\item{rescale}{Logical. Whether rows of the matrix should be rescaled to [0,1]. Default TRUE.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default "Dimension_1".}

\item{ylab}{Character vector. Label for the y-axis. Default "Dimension_2".}

\item{color_low}{Character. A color available from `colors()`. The color will be used to signify the lowest values on the scale. Default 'grey'.}

\item{color_mid}{Character. A color available from `colors()`. The color will be used to signify the midpoint on the scale.}

\item{color_high}{Character. A color available from `colors()`. The color will be used to signify the highest values on the scale. Default 'blue'.}
}
\description{
Create a scatterplot based off of a matrix containing the celda state probabilities per cell.
}
