% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDrCluster}
\alias{plotDrCluster}
\title{Create a scatterplot based on celda cluster labels.}
\usage{
plotDrCluster(dim1, dim2, cluster, size = 1, xlab = "Dimension_1",
  ylab = "Dimension_2", specific_clusters = NULL)
}
\arguments{
\item{dim1}{Numeric vector; first dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector; second dimension from data dimensionality reduction output.}

\item{cluster}{Vector; Contains cluster labels (e.g. from celda_C or celda_CG).}

\item{size}{Numeic vector; size of point on plot.Default 1.}

\item{xlab}{Character vector, used as label for rows. Default "Dimension_1".}

\item{ylab}{Character vector, used as label for columns. Default "Dimension_2".}

\item{specific_clusters}{Numeic vector; Contains specific cluster labels.}
}
\description{
Create a scatterplot based on celda cluster labels.
}
