% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDrGene}
\alias{plotDrGene}
\title{Create a scatterplot for each row of a normalized gene expression matrix where x and y axis are from a data dimensionality reduction tool.}
\usage{
plotDrGene(dim1, dim2, matrix, trim = c(-2, 2), rescale = FALSE, size = 1,
  xlab = "Dimension_1", ylab = "Dimension_2", color_low = "grey",
  color_mid = NULL, color_high = "blue")
}
\arguments{
\item{dim1}{Numeric vector; first dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector; second dimension from data dimensionality reduction output.}

\item{matrix}{Counts matrix, will have cell name for column name and gene name for row name.}

\item{trim}{A two element vector to specify the lower and upper cutoff for the data. Occurs after row scaling. Set to NULL to disable. Default c(-2,2).A two element vector to specify the lower and upper cutoff for the data. Occurs after normalization, log transformation, and row scaling. Set to NULL to disable. Default c(-2,2).}

\item{rescale}{Boolean. If TRUE, will rescale counts matrix on a 0~1 scale. Default FALSE.}

\item{size}{Numeic vector; size of point on plot. Default = 1.}

\item{xlab}{Character vector, used as label for x axis. Default "Dimension_1".}

\item{ylab}{Character vector, used as label for y axis. Default "Dimension_2".}

\item{color_low}{Character vector of R colors available from the colors() function. The color will be used to signify the lowest values on the scale. Default: 'grey'}

\item{color_mid}{Character vector of R colors available from the colors() function. The color will be used to signify the midpoint on the scale.}

\item{color_high}{Character vector of R colors available from the colors() function. The color will be used to signify the highest values on the scale. Default: 'blue'}
}
\description{
Create a scatterplot for each row of a normalized gene expression matrix where x and y axis are from a data dimensionality reduction tool.
}
