% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDrGrid}
\alias{plotDrGrid}
\title{Create a scatterplot given two dimensions from a data dimensionality reduction tool (e.g tSNE)}
\usage{
plotDrGrid(dim1, dim2, matrix, size, xlab, ylab, color_low, color_mid,
  color_high, var_label)
}
\arguments{
\item{dim1}{Numeric vector; first dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector; second dimension from data dimensionality reduction output.}

\item{matrix}{Matrix, will contain cells/samples as columns and variable of interest as rows.}

\item{size}{Numeic vector; size of point on plot.}

\item{xlab}{Character vector, used as label for x axis.}

\item{ylab}{Character vector, used as label for y axis.}

\item{color_low}{Character vector of R colors available from the colors() function. The color will be used to signify the lowest values on the scale.}

\item{color_mid}{Character vector of R colors available from the colors() function. The color will be used to signify the midpoint on the scale.}

\item{color_high}{Character vector of R colors available from the colors() function. The color will be used to signify the highest values on the scale.}

\item{var_label}{Character vector, used as label for the scale.}
}
\description{
Create a scatterplot given two dimensions from a data dimensionality reduction tool (e.g tSNE)
}
