% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity.celda_C}
\alias{plotGridSearchPerplexity.celda_C}
\title{Plot perplexity as a function of K from celda_C models}
\usage{
\method{plotGridSearchPerplexity}{celda_C}(celda.list)
}
\arguments{
\item{celda.list}{Object of class 'celda_list'.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters.
}
\description{
Plots perplexity as a function of the cell (K) clusters as generated by celdaGridSearch().
}
\examples{
celda.sim = simulateCells(model="celda_C")
celda.list = celdaGridSearch(celda.sim$counts, model="celda_C", 
                             K.to.test=3:5, max.iter=2, nchains=1)
celda.list = calculatePerplexityWithResampling(celda.sim$counts, celda.list)
plotGridSearchPerplexity(celda.list)
}
