% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity.celda_G}
\alias{plotGridSearchPerplexity.celda_G}
\title{Plot perplexity as a function of L from a celda_G model}
\usage{
\method{plotGridSearchPerplexity}{celda_G}(celda.list)
}
\arguments{
\item{celda.list}{Object of class 'celda_list'.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters.
}
\description{
Plots perplexity as a function of the gene (L) clusters as generated by celdaGridSearch().
}
\examples{
## Simulate a small dataset with 5 cell clusters and 10 feature modules
celda.sim = simulateCells(model="celda_G", L=10, G=100, C=100)
## Run various combinations of parameters with 'celdaGridSearch'
cgs = celdaGridSearch(celda.sim$counts, model="celda_G", params.test=list(L=9:11))
cgs = calculatePerplexityWithResampling(celda.sim$counts, cgs)
plotGridSearchPerplexity(cgs)
}
