% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoEnrich.R
\name{plotPhenoEnrich}
\alias{plotPhenoEnrich}
\title{Phenotype enrichment plot}
\usage{
plotPhenoEnrich(counts, celda.mod, phenoLabels, phenotype = NULL,
  cellCluster = NULL, perspective = c("phenotype", "cellcluster"),
  plot = c("mosaic", "balloon"), ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells.}

\item{celda.mod}{Celda model returned by the \link{celda_C},
\code{\link{celda_CG}} or \code{\link{subsetCeldaList}} functions. Must be
a \code{celda_C} or \code{celda_CG} object.}

\item{phenoLabels}{Character vector of phenotype labels for the cells. Must
have the same length as the column number of count matrix \code{counts}.}

\item{phenotype}{The phenotypes to inspect. If specified, cells will be
categorized into two groups based on whether their phenotype is in this
phenotype group or not. By default, no phenotype grouping is performed.}

\item{cellCluster}{The cell clusters to inspect. If specified, cells will be
categorized into two groups based on whether their assigned cell cluster is
in this cell cluster group or not. By default, no cell cluster grouping is
performed.}

\item{perspective}{The comparing perspective. One of "phenotype" or
"cellcluster". Indicates whether to look at the enrichment of phenotypes in
cell clusters or the enrichment of cell clusters in phenotypes. Default is
"phenotype", which looks at the enrichment of phenotype in cell clusters.}

\item{plot}{The plot of visualizing the result. One of "mosaic" or "balloon".
Uses the \code{\link[vcd]{mosaic}} function for plotting mosaic plot and the
\code{\link[ggpubr]{ggballoonplot}} function for plotting the balloon plot.}

\item{...}{Additional parameters passed to the \code{\link[vcd]{mosaic}} or
\code{\link[ggpubr]{ggballoonplot}} function.}
}
\value{
NULL if \code{plot = "mosaic"} and a ggplot object if
 \code{plot = "balloon"}
}
\description{
Plot the enrichment of phenotypes within cell clusters or
 \emph{vice versa} in a given celda model. Generates a mosaic plot or
 a balloon plot.
}
\examples{
# Use sample labels as a surrogate for phenotype labels
plotPhenoEnrich(celda.CG.sim$counts,
    celda.CG.mod,
    celda.CG.sim$sample.label)

# Look at the enrichment of "Sample_1" in cell cluster 1 in balloon plot
plotPhenoEnrich(counts = celda.CG.sim$counts,
    celda.mod = celda.CG.mod,
    phenoLabels = celda.CG.sim$sample.label,
    phenotype = "Sample_1",
    cellCluster = 1,
    plot = "balloon")

# Look at the enrichment of cell cluster 1 & 2 in
# phenotype "Sample_3" & "Sample_4" in mosaic plot
plotPhenoEnrich(counts = celda.CG.sim$counts,
    celda.mod = celda.CG.mod,
    phenoLabels = celda.CG.sim$sample.label,
    phenotype = c("Sample_3", "Sample_4"),
    cellCluster = c(1, 2),
    perspective = "cellcluster")
}
