% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDecisionTree.R
\name{print.moduleDecisionTree}
\alias{print.moduleDecisionTree}
\title{Print method for moduleDecisionTree}
\usage{
\method{print}{moduleDecisionTree}(x, ...)
}
\arguments{
\item{x}{A moduleDecisionTree object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for moduleDecisionTree
}
