% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{render_iteration_likelihood_plot}
\alias{render_iteration_likelihood_plot}
\title{Visualize log likelihood per iteration of Gibbs sampling}
\usage{
render_iteration_likelihood_plot(celda.obj)
}
\arguments{
\item{celda.obj}{A celda object}
}
\value{
A ggplot object of the likelihood per iteration of Gibbs sampling
}
\description{
Visualize log likelihood per iteration of Gibbs sampling
}
