% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_heatmap.R
\name{renderCeldaHeatmap}
\alias{renderCeldaHeatmap}
\title{Renders a heatmap based on a matrix of counts where rows are genes and columns are cells.}
\usage{
renderCeldaHeatmap(counts, z = NULL, y = NULL, scale.row = scale,
  trim = c(-2, 2), feature.ix = NULL, cell.ix = NULL,
  cluster.gene = TRUE, cluster.cell = TRUE, color_scheme = c("divergent",
  "sequential"), color_scheme_symmetric = TRUE, color_scheme_center = 0,
  col = NULL, annotation_cell = NULL, annotation_gene = NULL,
  annotation_color = NULL, breaks = NULL, legend = TRUE,
  annotation_legend = TRUE, annotation_names_gene = TRUE,
  annotation_names_cell = TRUE, show_genenames = FALSE,
  show_cellnames = FALSE, hclust_method = "ward.D2",
  treeheight_gene = ifelse(cluster.gene, 50, 0),
  treeheight_cell = ifelse(cluster.cell, 50, 0), ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{z}{Numeric vector. Denotes cell population labels.}

\item{y}{Numeric vector. Denotes feature module labels.}

\item{scale.row}{Function; A function to scale each individual row. Set to NULL to disable. Occurs after normalization and log transformation. Defualt is 'scale' and thus will Z-score transform each row.}

\item{trim}{Numeric vector. Vector of length two that specifies the lower and upper bounds for the data. This threshold is applied after row scaling. Set to NULL to disable. Default c(-2,2).}

\item{feature.ix}{Integer vector. Select features for display in heatmap. If NULL, no subsetting will be performed. Default NULL.}

\item{cell.ix}{Integer vector. Select cells for display in heatmap. If NULL, no subsetting will be performed. Default NULL.}

\item{cluster.gene}{Logical. Determines whether rows should be clustered. Default TRUE.}

\item{cluster.cell}{Logical. Determines whether columns should be clustered. Default TRUE.}

\item{color_scheme}{"Character. One of ""divergent"" or ""sequential"". A ""divergent"" scheme is best for highlighting relative data (denoted by 'color_scheme_center') such as gene expression data that has been normalized and centered. A ""sequential"" scheme is best for highlighting data that are ordered low to high such as raw counts or probabilities. Default "divergent".}

\item{color_scheme_symmetric}{Logical. When the color_scheme is "divergent" and the data contains both positive and negative numbers, TRUE indicates that the color scheme should be symmetric from [-max(abs(data)),max(abs(data))]. For example, if the data ranges goes from -1.5 to 2, then setting this to TRUE will force the color scheme to range from -2 to 2. Default TRUE.}

\item{color_scheme_center}{Numeric. Indicates the center of a "divergent" color_scheme. Default 0.}

\item{col}{Color for the heatmap.}

\item{annotation_cell}{Data frame. Additional annotations for each cell will be shown in the column color bars. The format of the data frame should be one row for each cell and one column for each annotation. Numeric variables will be displayed as continuous color bars and factors will be displayed as discrete color bars. Default NULL.}

\item{annotation_gene}{A data frame for the gene annotations (rows).}

\item{annotation_color}{List. Contains color scheme for all annotations. See `?pheatmap` for more details.}

\item{breaks}{Numeric vector. A sequence of numbers that covers the range of values in the normalized `counts`. Values in the normalized `matrix` are assigned to each bin in `breaks`. Each break is assigned to a unique color from `col`. If NULL, then breaks are calculated automatically. Default NULL.}

\item{legend}{Logical. Determines whether legend should be drawn. Default TRUE.}

\item{annotation_legend}{Logical. Whether legend for all annotations should be drawn. Default TRUE.}

\item{annotation_names_gene}{Logical. Whether the names for features should be shown. Default TRUE.}

\item{annotation_names_cell}{Logical. Whether the names for cells should be shown. Default TRUE.}

\item{show_genenames}{Logical. Specifies if feature names should be shown. Default TRUE.}

\item{show_cellnames}{Logical. Specifies if cell names should be shown. Default FALSE.}

\item{hclust_method}{Character. Specifies the method to use for the 'hclust' function. See `?hclust` for possible values. Default "ward.D2".}

\item{treeheight_gene}{Numeric. Width of the feature dendrogram. Set to 0 to disable plotting of this dendrogram. Default: if cluster.gene == TRUE, then treeheight.feature = 50, else treeheight.feature = 0.}

\item{treeheight_cell}{Numeric. Height of the cell dendrogram. Set to 0 to disable plotting of this dendrogram. Default: if cluster.cell == TRUE, then treeheight.cell = 50, else treeheight.cell = 0.}

\item{...}{Other arguments to be passed to underlying pheatmap function.}
}
\description{
Renders a heatmap based on a matrix of counts where rows are genes and columns are cells.
}
