% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{renderInteractiveKLPlot}
\alias{renderInteractiveKLPlot}
\title{Render an interactive figure demonstrating the clustering performance of different K/L parameters}
\usage{
renderInteractiveKLPlot(celda.list, method = "perplexity",
  title = "Model Performance (All Chains)")
}
\arguments{
\item{celda.list}{A list of celda_CG objects returned from celda function}

\item{method}{One of "perplexity" or "loglik", passed through to calculatePerformanceMetric()}

\item{title}{The plot title}
}
\description{
Plot the performance of a list of celda_CG models returned 
from running the celda function. This is visualized by rendering a boxplot for each chain
run for each combination of K/L (cell/gene).
}
