% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{renderIterationLikelihoodPlot}
\alias{renderIterationLikelihoodPlot}
\title{Visualize log likelihood per iteration of Gibbs sampling}
\usage{
renderIterationLikelihoodPlot(celda.obj)
}
\arguments{
\item{celda.obj}{A celda object}
}
\value{
A ggplot object of the likelihood per iteration of Gibbs sampling
}
\description{
Visualize log likelihood per iteration of Gibbs sampling
}
