% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_heatmap.R
\name{renderProbabilityHeatmap}
\alias{renderProbabilityHeatmap}
\title{Render a heatmap based on a population matrix from the factorized counts matrix.}
\usage{
renderProbabilityHeatmap(counts, model, relative = TRUE, scale = TRUE,
  trim = c(-2, 2))
}
\arguments{
\item{counts}{A count matrix where rows are genes and columns are cells, used to generate celda model.}

\item{model}{celda model of class "celda_CG.}

\item{relative}{Boolean, converts the proportions in the population matrix to a relative probability of each transcriptional state in each cellular subpopulation. Default: TRUE}

\item{scale}{Boolean, z-scores the population matrix. default: TRUE}

\item{trim}{A two element vector to specify the lower and upper cutoff for the data. Set to NULL to disable. Default c(-2,2).}
}
\description{
Render a heatmap based on a population matrix from the factorized counts matrix.
}
