% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{resList,celdaList-method}
\alias{resList,celdaList-method}
\title{Get final celda models from a celdaList}
\usage{
\S4method{resList}{celdaList}(celda.mod)
}
\arguments{
\item{celda.mod}{An object of class celdaList.}
}
\value{
List. Contains one celdaModel object for each of the parameters specified in the `runParams()` of the provided celda list.
}
\description{
Returns all models generated during a `celdaGridSearch()` run.
}
\examples{
celda.CG.grid.models = resList(celda.CG.grid.search.res)
}
