% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{resList}
\alias{resList}
\alias{resList,SingleCellExperiment-method}
\alias{resList,celdaList-method}
\title{Get final celdaModels from a celda model \code{SCE} or celdaList
 object}
\usage{
resList(x, altExpName = "featureSubset")

\S4method{resList}{SingleCellExperiment}(x, altExpName = "featureSubset")

\S4method{resList}{celdaList}(x)
}
\arguments{
\item{x}{An object of class \link[SingleCellExperiment:SingleCellExperiment-class]{SingleCellExperiment} or
\code{celdaList}.}

\item{altExpName}{The name for the \link[SingleCellExperiment]{altExp} slot
to use. Default "featureSubset".}
}
\value{
List. Contains one celdaModel object for each of the parameters
 specified in \code{runParams(x)}.
}
\description{
Returns all celda models generated during a
 \link{celdaGridSearch} run.
}
\examples{
data(sceCeldaCGGridSearch)
celdaCGGridModels <- resList(sceCeldaCGGridSearch)
data(celdaCGGridSearchRes)
celdaCGGridModels <- resList(celdaCGGridSearchRes)
}
