% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.cells}
\alias{sample.cells}
\title{A toy count matrix for use with celda.}
\format{A matrix of simulated gene counts with 10 rows (genes) and 10 columns (cells)}
\source{
\url{http://github.com/definitelysean/celda}
}
\usage{
sample.cells
}
\description{
A toy count matrix for use with celda.
}
\keyword{datasets}

