% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{sample.ll}
\alias{sample.ll}
\title{Sample log-likelihood probabilities}
\usage{
sample.ll(ll.probs)
}
\arguments{
\item{counts}{A numeric count matrix}
}
\value{
A single integer in 1:k corresponding to a cluster assignment
}
\description{
Given a set of log-likelihoods for cluster membership, return a single cluster assignment.
}
\examples{
TODO
}
\keyword{likelihood}
\keyword{log}
\keyword{sample}

