% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleCells}
\alias{sampleCells}
\title{sampleCells}
\format{
A matrix of simulated gene counts with 10 rows (genes) and 10
 columns (cells).
}
\source{
\url{http://github.com/campbio/celda}
}
\usage{
data(sampleCells)
}
\description{
A matrix of simulated gene counts.
}
\details{
A toy count matrix for use with celda.

Generated by Josh Campbell.
}
\keyword{datasets}
