% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{sampleLabel,celdaModel-method}
\alias{sampleLabel,celdaModel-method}
\title{Get sample labels from a celda model}
\usage{
\S4method{sampleLabel}{celdaModel}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda model. Options available in `celda::available.models`.}
}
\value{
Character. Contains the sample labels provided at model creation time, or those automatically generated by celda.
}
\description{
Returns the sample labels for the count matrix provided for generation of a given celda model.
}
\examples{
sampleLabel(celda.CG.mod)
}
