% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCelda_C}
\alias{sceCelda_C}
\title{sceCelda_C}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
sceCelda_C
}
\description{
A \linkS4class{SingleCellExperiment} object containing the results of running \link{celda_C} on \link{celdaCSim}.
}
\examples{
data(celdaCSim)
sceCelda_C = celda_C(celdaCSim$counts,
    K = celdaCSim$K,
    sampleLabel = celdaCSim$sampleLabel)
}
\keyword{datasets}
