% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCeldaG}
\alias{sceCeldaG}
\title{sceCeldaG}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
data(sceCeldaG)
}
\description{
A \linkS4class{SingleCellExperiment} object containing the
 results of running \link{selectFeatures} and \link{celda_G} on
 \link{celdaGSim}.
}
\examples{
data(celdaGSim)
sceCeldaG <- selectFeatures(celdaGSim$counts)
sceCeldaG <- celda_G(sceCeldaG, L = celdaGSim$L, nchains = 1)
}
\keyword{datasets}
