% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{seed}
\alias{seed}
\title{Get the random seed for a given celda model.}
\usage{
seed(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object (of class "celda_C", "celda_G" or "celda_CG")}
}
\description{
Get the random seed for a given celda model.
}
