% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_list.R
\name{selectModels}
\alias{selectModels}
\title{Select a set of models from a celda_list objects based off of rows in its run.params attribute.}
\usage{
selectModels(celda.list, run.param.rows)
}
\arguments{
\item{celda.list}{A celda_list object returned from celda()}

\item{run.param.idx}{Row indices in the celda.list's run params corresponding to the desired models.}
}
\value{
A celda_list containing celda model objects matching the provided parameters (of class "celda_C", "celda_G", "celda_CG" accordingly), or NA if one is not found.
}
\description{
Select a set of models from a celda_list objects based off of rows in its run.params attribute.
}
