% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{simulateCells.celda_C}
\alias{simulateCells.celda_C}
\title{simulateCells for celda Cell clustering function}
\usage{
simulateCells.celda_C(S = 10, C.Range = c(10, 100), N.Range = c(100,
  5000), G = 500, K = 5, alpha = 1, beta = 1)
}
\arguments{
\item{S}{Total number of samples}

\item{C.Range}{Vector of length 2 given the range (min,max) of number of cells for each sample to be randomly generated from the uniform distribution}

\item{N.Range}{Vector of length 2 given the range (min,max) of number of counts for each cell to be randomly generated from the uniform distribution}

\item{G}{Total number of Genes to be simulated}

\item{K}{An integer or range of integers indicating the desired number of cell clusters (for celda_C / celda_CG models)}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution}

\item{beta}{Non-zero concentration parameter for gene Dirichlet distribution}
}
\description{
simulateCells for celda Cell clustering function
}
