% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{simulateCellscelda_C}
\alias{simulateCellscelda_C}
\title{Simulate cells from the celda_C model}
\usage{
simulateCellscelda_C(
  model,
  S = 5,
  CRange = c(50, 100),
  NRange = c(500, 1000),
  G = 100,
  K = 5,
  alpha = 1,
  beta = 1,
  seed = 12345
)
}
\arguments{
\item{model}{Character. Options available in `celda::availableModels`.}

\item{S}{Integer. Number of samples to simulate. Default 5.}

\item{CRange}{Vector of length 2 given the range (min, max) of number of
cells for each sample to be randomly generated from the uniform
distribution. Default c(50, 100).}

\item{NRange}{Integer vector. A vector of length 2 that specifies the lower
and upper bounds of the number of counts generated for each cell. Default
c(500, 1000).}

\item{G}{Integer. The total number of features to be simulated. Default 100.}

\item{K}{Integer. Number of cell populations. Default 5.}

\item{alpha}{Numeric. Concentration parameter for Theta. Adds a pseudocount
to each cell population in each sample. Default 1.}

\item{beta}{Numeric. Concentration parameter for Phi. Adds a pseudocount to
each feature in each cell population. Default 1.}

\item{seed}{Integer. Passed to \link[withr]{with_seed}. For reproducibility,
a default value of 12345 is used. If NULL, no calls to
\link[withr]{with_seed} are made.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object with
 simulated count matrix stored in the "counts" assay slot. Function
 parameter settings are stored in the \link[S4Vectors]{metadata} slot.
 Columns \code{sample_label} and \code{cell_cluster} in
 \link[SummarizedExperiment]{colData} contain simulated sample labels and
 cell population clusters.
}
\description{
Generates a simulated counts matrix, cell subpopulation
 clusters, and sample labels according to the generative process of the
 celda_C model.
}
\examples{
sce <- simulateCells(model = "celda_C", K = 10)
}
\seealso{
`celda_G()` for simulating feature modules and `celda_CG()` for
 simulating feature modules and cell populations.
}
