% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{topRank}
\alias{topRank}
\title{Identify features with the highest influence on clustering}
\usage{
topRank(fm, n = 25, margin = 2, threshold = 0, decreasing = TRUE)
}
\arguments{
\item{fm}{factorized matrix.}

\item{n}{Maximum number of items returned for each entry.}

\item{margin}{1 for rows, 2 for columns.}

\item{threshold}{only include entries in the matrix that is greader than the threshold.}

\item{decreasing}{Logical; specifying if rank should be decreasing. Default to be TRUE.}
}
\description{
Identify features with the highest influence on clustering
}
