% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda.R
\name{validate_args}
\alias{validate_args}
\title{Sanity check arguments to celda() to ensure a smooth run.}
\usage{
validate_args(counts, model, sample.label, nchains, cores, seed)
}
\description{
Sanity check arguments to celda() to ensure a smooth run.
}

