% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda.R
\name{validate_counts}
\alias{validate_counts}
\title{Perform some simple checks on the counts matrix, to ensure celda won't choke.
Currently only checks that none of the rows or columns in the count matrix are all 0.}
\usage{
validate_counts(counts)
}
\description{
Perform some simple checks on the counts matrix, to ensure celda won't choke.
Currently only checks that none of the rows or columns in the count matrix are all 0.
}

