% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{validateRunParams}
\alias{validateRunParams}
\title{Check whether the given celda_list's run.params attribute ordering matches it's res.list models}
\usage{
validateRunParams(celda.list)
}
\arguments{
\item{celda.list}{A celda_list model as generated by celda()}
}
\value{
TRUE if the celda.list's run.params K/L attributes are ordered similarly to the res.list models' K/L values, FALSE otherwise.
}
\description{
Check whether the given celda_list's run.params attribute ordering matches it's res.list models
}
