% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{visualize_model_performance.celda_C}
\alias{visualize_model_performance.celda_C}
\title{visualize_model_performance for celda Cell clustering function}
\usage{
\method{visualize_model_performance}{celda_C}(celda.list,
  method = "perplexity", title = "Model Performance (All Chains)",
  log = F)
}
\arguments{
\item{celda.list}{A celda_list object returned from celda()}

\item{method}{One of "perplexity", "loglik"}

\item{title}{Title for the plot}

\item{log}{Currently not working for celda_C objects}
}
\description{
visualize_model_performance for celda Cell clustering function
}
