% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{visualize_model_performance.celda_G}
\alias{visualize_model_performance.celda_G}
\title{visualize_model_performance for the celda Gene function}
\usage{
\method{visualize_model_performance}{celda_G}(celda.list,
  method = "perplexity", title = "Model Performance (All Chains)",
  log = F)
}
\arguments{
\item{celda.list}{A celda_list object returned from celda()}

\item{method}{One of "perplexity" or "loglik"}

\item{title}{Title for the plot}

\item{log}{Currently not working for celda.G objects}
}
\description{
visualize_model_performance for the celda Gene function
}
