% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_performance.R
\name{visualize_silhouette_widths}
\alias{visualize_silhouette_widths}
\title{Visualize the average cluster silhouette width for all models in a celda_list.}
\usage{
visualize_silhouette_widths(celda.res, counts, normalize.counts = T,
  dist.metric = "euclidian",
  title = "Average Silhouette Width By Cluster Size")
}
\arguments{
\item{celda.res}{A celda_list returned from celda()}

\item{counts}{The counts matrix used to generate celda.res}

\item{normalize.counts}{Whether to normalize the counts matrix, defaults to TRUE}

\item{dist.metric}{The distance metric to use in calculating the dissimilarity matrix. Choices are "euclidian", "cosine". Defaults to "euclidian".}

\item{title}{Plot title to use when visualizing silhouette width distributions}
}
\description{
Returns a ggplot of boxplots, breaking out silhouette widths by each value of K/L
(or combination thereof) present in the celda_list, across all chains.
}
