% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{visualizeModelPerformance.celda_CG}
\alias{visualizeModelPerformance.celda_CG}
\title{Visualize the performance of celda_CG models grouped by L and K}
\usage{
\method{visualizeModelPerformance}{celda_CG}(celda.list, counts,
  method = "perplexity", resample = 1,
  title = "Model Performance (All Chains)", log = FALSE)
}
\arguments{
\item{celda.list}{A list of celda_CG objects returned from celda function}

\item{counts}{The counts used to generate the celda.list results}

\item{method}{One of "perplexity" or "loglik"}

\item{resample}{Number of resamplings to evaluate for perplexity, if method = "perplexity"}

\item{title}{Title for the visualizeModelPerformance}

\item{log}{Set log to TRUE to visualize the log(perplexity) of Celda_CG objects.}
}
\description{
Plot the performance of a list of celda_CG models returned 
from running celda function. For each number of gene clusters 
L (cell clusters K), plot the performance of each number of cell
clusters K (gene clusters L).
}
