% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{visualizeModelPerformance.celda_G}
\alias{visualizeModelPerformance.celda_G}
\title{visualizeModelPerformance for the celda Gene function}
\usage{
\method{visualizeModelPerformance}{celda_G}(celda.list, counts,
  method = "perplexity", title = "Model Performance (All Chains)",
  log = F)
}
\arguments{
\item{celda.list}{A celda_list object returned from celda()}

\item{counts}{The counts used to generate the celda.list results}

\item{method}{One of "perplexity" or "loglik"}

\item{title}{Title for the plot}

\item{log}{Currently not working for celda.G objects}
}
\description{
visualizeModelPerformance for the celda Gene function
}
