% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{visualizeModelPerformance}
\alias{visualizeModelPerformance}
\title{Visualize various performance metrics as a function of K / L to aid parameter choice.}
\usage{
visualizeModelPerformance(celda.list, method, title, log = FALSE)
}
\arguments{
\item{celda.list}{A celda_list object as returned from *celda()*}

\item{method}{Which performance metric to visualize. One of ("perplexity", "harmonic", "loglik"). "perplexity" calculates the inverse of the geometric mean of the log likelihoods from each iteration of Gibbs sampling. "harmonic" calculates the marginal likelihood has the harmonic mean of the likelihoods. "loglik" plots the highest log-likelihood during Gibbs iteration.}

\item{title}{Title for the visualizeModelPerformance}

\item{log}{Set log to TRUE to visualize the log(perplexity) of Celda_CG objects. Does not work for "harmonic" metric}
}
\value{
A ggplot object containing the requested plot(s)
}
\description{
Visualize various performance metrics as a function of K / L to aid parameter choice.
}
