% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{visualizePerplexity}
\alias{visualizePerplexity}
\title{Visualize perplexity as a function of cell/gene clustering}
\usage{
visualizePerplexity(perplexity.df, title = "")
}
\arguments{
\item{perplexity.df}{A data.frame as described above, or as returned by calculatePerplexityWithResampling()}

\item{title}{A subtitle for the plot}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters
}
\description{
This function plots perplexity as a function of the cell/gene (K/L) clusters as generated by celda().
It accepts a perplexity dataframe as generated by calculatePerplexityWithResampling().
}
\details{
The provided #' data frame should have columns for at least one of K or L 
(the clusterings provided by celda()), and a column for the perplexity calculated for 
the corresponding chain.
}
