% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_heatmap.R
\name{absoluteProbabilityHeatmap}
\alias{absoluteProbabilityHeatmap}
\title{Renders a heatmap based on a population matrix from the factorized counts matrix.}
\usage{
absoluteProbabilityHeatmap(counts, celda.mod, main = NA)
}
\arguments{
\item{counts}{A count matrix where rows are genes and columns are cells, used to generate celda model.}

\item{celda.mod}{celda model of class "celda_CG".}

\item{main}{The title of the plot; default = NA.}
}
\description{
Renders a heatmap based on a population matrix from the factorized counts matrix.
}
