% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{bestLogLikelihood,celdaModel-method}
\alias{bestLogLikelihood,celdaModel-method}
\title{Get the log-likelihood}
\usage{
\S4method{bestLogLikelihood}{celdaModel}(celdaMod)
}
\arguments{
\item{celdaMod}{A celdaModel object of class celda_C, celda_G, or celda_CG.}
}
\value{
Numeric. The log-likelihood at the final step of Gibbs sampling used
 to generate the model.
}
\description{
Retrieves the final log-likelihood from all iterations of Gibbs
 sampling used to generate a celdaModel.
}
\examples{
data(celdaCGMod)
bestLogLikelihood(celdaCGMod)
}
