% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{bestLogLikelihood}
\alias{bestLogLikelihood}
\alias{bestLogLikelihood,celdaModel-method}
\title{Get the log-likelihood}
\usage{
bestLogLikelihood(x)

\S4method{bestLogLikelihood}{celdaModel}(x)

\S4method{bestLogLikelihood}{celdaModel}(x)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object
returned by \link{celda_C}, \link{celda_G}, or \link{celda_CG}, or a celda
model object.}
}
\value{
Numeric. The log-likelihood at the final step of Gibbs sampling used
 to generate the model.
}
\description{
Retrieves the final log-likelihood from all iterations of Gibbs
 sampling used to generate a celdaModel.
}
\examples{
data(celdaCGMod)
bestLogLikelihood(celdaCGMod)
data(sceCeldaCG)
bestLogLikelihood(sceCeldaCG)
data(celdaCGMod)
bestLogLikelihood(celdaCGMod)
}
