% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{bestLogLikelihood}
\alias{bestLogLikelihood}
\title{Get the log-likelihood}
\usage{
bestLogLikelihood(celda.mod)
}
\value{
Numeric. The log-likelihood at the final step of Gibbs sampling used to generate the model.
}
\description{
Retrieves the final log-likelihood from all iterations of Gibbs sampling used to generate a celda model.
}
\examples{
bestLogLikelihood(celda.CG.mod)
}
