% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_perplexity}
\alias{calculate_perplexity}
\title{Calculate the perplexity from a single celda chain.
Perplexity is defined as the inverse of the geometric mean of the}
\usage{
calculate_perplexity(completeLogLik)
}
\arguments{
\item{completeLogLik}{The complete Gibbs sampling history of log-likelihoods for a single celda chain}
}
\value{
The perplexity for the provided chain as an mpfr number
}
\description{
Calculate the perplexity from a single celda chain.
Perplexity is defined as the inverse of the geometric mean of the
}
