% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{calculateLoglikFromVariables.celda_C}
\alias{calculateLoglikFromVariables.celda_C}
\title{Calculate the celda_C log likelihood for user-provided cluster assignments}
\usage{
calculateLoglikFromVariables.celda_C(counts, sample.label, z, K, alpha, beta)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{sample.label}{Vector or factor. Denotes the sample label for each cell (column) in the count matrix.}

\item{z}{Numeric vector. Denotes cell population labels.}

\item{K}{Integer. Number of cell populations.}

\item{alpha}{Numeric. Concentration parameter for Theta. Adds a pseudocount to each cell population in each sample. Default 1.}

\item{beta}{Numeric. Concentration parameter for Phi. Adds a pseudocount to each feature in each cell population. Default 1.}

\item{...}{Additional parameters.}
}
\description{
Calculate the celda_C log likelihood for user-provided cluster assignments
}
