% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{calculateLoglikFromVariables.celda_CG}
\alias{calculateLoglikFromVariables.celda_CG}
\title{Calculate log lileklihood for the celda Cell and Gene clustering model, given a set of cell / gene cluster assignments}
\usage{
calculateLoglikFromVariables.celda_CG(counts, sample.label, z, y, K, L, alpha,
  beta, delta, gamma)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{sample.label}{Vector or factor. Denotes the sample label for each cell (column) in the count matrix.}

\item{z}{Numeric vector. Denotes cell population labels.}

\item{y}{Numeric vector. Denotes feature module labels.}

\item{K}{Integer. Number of cell populations.}

\item{L}{Integer. Number of feature modules.}

\item{alpha}{Numeric. Concentration parameter for Theta. Adds a pseudocount to each cell population in each sample. Default 1.}

\item{beta}{Numeric. Concentration parameter for Phi. Adds a pseudocount to each feature module in each cell population. Default 1.}

\item{delta}{Numeric. Concentration parameter for Psi. Adds a pseudocount to each feature in each module. Default 1.}

\item{gamma}{Numeric. Concentration parameter for Eta. Adds a pseudocount to the number of features in each module. Default 1.}

\item{...}{Additional parameters.}
}
\description{
Calculate log lileklihood for the celda Cell and Gene clustering model, given a set of cell / gene cluster assignments
}
