% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{calculateLoglikFromVariables.celda_G}
\alias{calculateLoglikFromVariables.celda_G}
\title{Calculate the celda_G log likelihood for user-provided cluster assignments}
\usage{
\method{calculateLoglikFromVariables}{celda_G}(counts, y, L, beta, delta, gamma,
  ...)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{y}{A numeric vector of gene cluster assignments}

\item{L}{The number of clusters being considered}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state}

\item{...}{Additional parameters}
}
\value{
The log likelihood of the provided cluster assignment, as calculated by the celda_G likelihood function
}
\description{
This function calculates the log likelihood of each clustering of genes generated
over multiple iterations of Gibbs sampling.
}
\keyword{likelihood}
\keyword{log}
