% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{calculateLoglikFromVariables}
\alias{calculateLoglikFromVariables}
\title{Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.}
\usage{
calculateLoglikFromVariables(model, counts, ...)
}
\arguments{
\item{model}{Model to use for calculating log-likelihood of assignments; one of ("celda_C", "celda_CG", "celda_G")}

\item{counts}{The counts matrix used to generate the provided cluster assignments}

\item{...}{extra parameters passed onto calculate_loglik_from_variables}
}
\value{
The log-likelihood of the provided cluster assignment for the provided counts matrix.
}
\description{
Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.
}
