% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{calculatePerplexity}
\alias{calculatePerplexity}
\title{Calculate the perplexity from a single celda model}
\usage{
calculatePerplexity(counts, celda.mod, new.counts = NULL)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C", "celda_G" or "celda_CG".}

\item{new.counts}{A new counts matrix used to calculate perplexity. If NULL, perplexity will be calculated for the 'counts' matrix. Default NULL.}
}
\value{
Numeric. The perplexity for the provided count data and model.
}
\description{
Perplexity can be seen as a measure of how well a provided set of 
cluster assignments fit the data being clustered.
}
\examples{
celda.sim = simulateCells(model="celda_CG", K=5, L=10, G=100)
celda.mod = celda_CG(celda.sim$counts, sample.label=celda.sim$sample.label,
                     K=celda.sim$K, L=celda.sim$L, nchain=1)
perplexity = calculatePerplexity(celda.sim$counts, celda.mod)
}
