% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{calculatePerplexity}
\alias{calculatePerplexity}
\title{Calculate the perplexity from a single celda run}
\usage{
calculatePerplexity(counts, celda.mod, validate.counts = TRUE)
}
\arguments{
\item{counts}{The count matrix modeled in the celdaRun parameter}

\item{celda.mod}{A single celda run (usually from the _res.list_ property of a celda_list)}

\item{validate.counts}{Whether to verify that the counts matrix provided was used to generate the results in celda.mod. Defaults to TRUE.}
}
\value{
The perplexity for the provided data and model
}
\description{
Perplexity can be seen as a measure of how well a provided set of 
cluster assignments fit the count data being modeled.
}
\details{
Perplexity is defined in LDA as the exp of the log likelihood of the model
divided by the total amount of word tokens. The corresponding perplexity
for the celda models are derived in their respective model description 
documents. These documents will not be made publicly available until 
after publication of celda.
}
