% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{calculatePerplexityWithResampling}
\alias{calculatePerplexityWithResampling}
\title{Calculate and visualize perplexity of all models in a celda_list, with resampling}
\usage{
calculatePerplexityWithResampling(celda.list, counts, resample, title = "",
  validate.counts = TRUE)
}
\arguments{
\item{celda.list}{A celda_list object as returned from *celda()*}

\item{counts}{The counts matrix used to generate the provided celda.list}

\item{resample}{The number of times to resample the counts matrix for evaluating perplexity. Works with method="perplexity."}

\item{title}{Title to be appended to the perplexity plot title. Default is "".}

\item{validate.counts}{Whether to verify that the counts matrix provided was used to generate the results in celda.mod. Defaults to TRUE.}
}
\value{
A list with a data frame summarizing all of the calculated perplexities, and a ggplot2 object visualizing them.
}
\description{
Calculates the perplexity (a measure of how well celda's cluster assignments fit the data) for every celda model in a celda_list generated by the celda() function. Perplexity of the cluster assignments given the provided count matrix,
as well as resamplings of that count matrix, provide a distribution of perplexities and a better sense of the quality of a given K/L choice.
}
