% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{calculateTsne}
\alias{calculateTsne}
\title{Run the t-SNE algorithm for dimensionality reduction}
\usage{
calculateTsne(norm, perplexity = 20, max.iter = 2500, seed = 12345,
  do.pca = FALSE, initial.dims = 20)
}
\arguments{
\item{norm}{Normalized count matrix.}

\item{perplexity}{Numeric vector. Determines perplexity for tsne. Default 20.}

\item{max.iter}{Numeric vector. Determines iterations for tsne. Default 1000.}

\item{seed}{Integer. Seed for random number generation. Defaults to 12345.}

\item{do.pca}{Logical. Whether to perform dimensionality reduction with PCA before tSNE.}

\item{initial.dims}{Integer. Number of dimensions from PCA to use as input in tSNE.}
}
\description{
Run the t-SNE algorithm for dimensionality reduction
}
\examples{
celda.sim = simulateCells("celda_CG")
celda.mod = celda_CG(celda.sim$counts, K=celda.sim$K, L=celda.sim$L,
                     nchains=1, max.iter=1)
norm = normalizeCounts(celda.sim$counts, normalize="proportion", transformation.fun=sqrt)
tsne = calculateTsne(norm = norm, max.iter = 10)
}
