% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{calculateTsne}
\alias{calculateTsne}
\title{Uses Rtsne package to run tSNE.}
\usage{
calculateTsne(norm, perplexity = 20, max.iter = 2500,
  distance = c("hellinger", "euclidean", "cosine", "spearman"),
  seed = 12345, do.pca = FALSE, initial.dims = 20)
}
\arguments{
\item{norm}{Normalized count matrix}

\item{perplexity}{Numeric vector; determines perplexity for tsne. Default 20.}

\item{max.iter}{Numeric vector; determines iterations for tsne. Default 1000.}

\item{distance}{Character vector; determines which distance metric to use for tsne. Options: cosine, hellinger, spearman or euclidean.}

\item{seed}{Seed for random number generation. Defaults to 12345.}

\item{do.pca}{Perform dimensionality reduction with PCA before tSNE.}

\item{initial.dims}{Number of dimensions from PCA to use as input in tSNE.}
}
\description{
Uses Rtsne package to run tSNE.
}
