% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{cC.decomposeCounts}
\alias{cC.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cC.decomposeCounts(counts, s, z, K)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{s}{Integer vector. Contains the sample label for each cell (column) in the count matrix.}

\item{z}{A numeric vector of cluster assignments}

\item{K}{Integer. Number of cell populations.}
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
