% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{cC.decomposeCounts}
\alias{cC.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cC.decomposeCounts(counts, s, z, K)
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
