% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{cCG.decomposeCounts}
\alias{cCG.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cCG.decomposeCounts(counts, s, z, y, K, L)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{s}{Integer vector. Contains the sample label for each cell (column) in the count matrix.}

\item{z}{Numeric vector. Denotes cell population labels.}

\item{y}{Numeric vector. Denotes feature module labels.}

\item{K}{Integer. Number of cell populations.}

\item{L}{Integer. Number of feature modules.}
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
