% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{cCG.decomposeCounts}
\alias{cCG.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cCG.decomposeCounts(counts, s, z, y, K, L)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{s}{An integer vector indicating the sample label for each cell (column) in the count matrix}

\item{z}{A numeric vector of cell cluster assignments}

\item{y}{A numeric vector of gene cluster assignments}

\item{K}{The number of cell clusters}

\item{L}{The number of gene clusters}
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
