% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celda.CG.grid.search.res}
\alias{celda.CG.grid.search.res}
\title{celda.CG.grid.search.res}
\format{An object as returned from celdaGridSearch()}
\usage{
celda.CG.grid.search.res
}
\description{
Example results of celdaGridSearch on celda.CG.sim
}
\examples{
cgs = celdaGridSearch(celda.CG.sim$counts, model="celda_CG", 
                      params.test=list(K=4:6, L=9:11), 
                      params.fixed=list(sample.label=celda.CG.sim$sample.label),
                      best.only=TRUE, nchains=1)
}
\keyword{datasets}
