% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda.R
\name{celda}
\alias{celda}
\title{Run the celda Bayesian hierarchical model on a matrix of counts.}
\usage{
celda(counts, model, sample.label = NULL, nchains = 1, cores = 1,
  seed = 12345, ...)
}
\arguments{
\item{counts}{A count matrix}

\item{model}{Which celda sub-model to run. Options include "celda_C" (cell clustering), "celda_G" (gene clustering), "celda_CG" (gene and cell clustering)}

\item{sample.label}{A numeric vector indicating the sample for each cell (column) in the count matrix}

\item{nchains}{The number of chains of Gibbs sampling to run for every combination of K/L parameters}

\item{cores}{The number of cores to use to speed up Gibbs sampling}

\item{seed}{The base seed for random number generation. Each chain celda runs with have a seed index off of this one.}

\item{K}{An integer or range of integers indicating the desired number of cell clusters (for celda_C / celda_CG models)}

\item{L}{An integer or range of integers indicating the desired number of gene clusters (for celda_G / celda_CG models)}
}
\description{
Run the celda Bayesian hierarchical model on a matrix of counts.
}

