% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{celda}
\alias{celda}
\title{Deprecation warning for old grid search function}
\usage{
celda(...)
}
\arguments{
\item{...}{Additional parameters.}
}
\description{
Deprecation warning for old grid search function
}
