% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celda_C}
\alias{celda_C}
\title{celda Cell Clustering Model}
\usage{
celda_C(counts, sample.label = NULL, K, alpha = 1, beta = 1,
  count.checksum = NULL, max.iter = 50, seed = 12345, split.on.iter = 5,
  num.splits = 5, z.init = NULL, logfile = NULL, ...)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{sample.label}{A vector indicating the sample for each cell (column) in the count matrix}

\item{K}{An integer or range of integers indicating the desired number of cell clusters (for celda_C / celda_CG models)}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution}

\item{beta}{Non-zero concentration parameter for gene Dirichlet distribution}

\item{count.checksum}{An MD5 checksum for the provided counts matrix}

\item{max.iter}{Maximum iterations of Gibbs sampling to perform. Defaults to 25}

\item{seed}{Parameter to set.seed() for random number generation}

\item{split.on.iter}{On every split.on.iter iteration, a heuristic will be applied to determine if a cell cluster should be reassigned and another cell cluster should be split into two clusters. Default to be 5.}

\item{num.splits}{Maximum number of times to perform the heuristic described in split.on.iter. Default 5.}

\item{z.init}{Initial values of z. If NULL, z will be randomly sampled. Default NULL.}

\item{logfile}{If NULL, messages will be displayed as normal. If set to a file name, messages will be redirected messages to the file. Default NULL.}

\item{...}{additonal parameters}
}
\value{
An object of class celda_C with clustering results and Gibbs sampling statistics
}
\description{
celda Cell Clustering Model
}
