% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celda_C}
\alias{celda_C}
\title{celda Cell Clustering Model}
\usage{
celda_C(counts, sample.label = NULL, K, alpha = 1, beta = 1,
  algorithm = c("EM", "Gibbs"), stop.iter = 10, max.iter = 200,
  split.on.iter = 10, split.on.last = TRUE, seed = 12345, nchains = 3,
  count.checksum = NULL, z.init = NULL, logfile = NULL)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{sample.label}{A vector indicating the sample for each cell (column) in the count matrix}

\item{K}{An integer or range of integers indicating the desired number of cell clusters (for celda_C / celda_CG models)}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution}

\item{beta}{Non-zero concentration parameter for gene Dirichlet distribution}

\item{algorithm}{Use 'EM' or 'Gibbs' sampling for clustering of cells into subpopulations. EM is much faster for larger datasets. Default 'EM'.}

\item{stop.iter}{Number of iterations without improvement in the log likelihood to stop the inference algorithm. Default 10.}

\item{max.iter}{Maximum iterations of inference algorithm to perform regardless of convergence. Default 200.}

\item{split.on.iter}{On every 'split.on.iter' iteration, a heuristic will be applied to determine if a gene/cell cluster should be reassigned and another gene/cell cluster should be split into two clusters. Default 10.}

\item{split.on.last}{After the the chain has converged according to 'stop.iter', a heuristic will be applied to determine if a gene/cell cluster should be reassigned and another gene/cell cluster should be split into two clusters. If a split occurs, then 'stop.iter' will be reset. Default TRUE.}

\item{seed}{Parameter to set.seed() for random number generation. Default 12345.}

\item{nchains}{Number of random z initializations. Default 3.}

\item{count.checksum}{An MD5 checksum for the provided counts matrix. Default NULL.}

\item{z.init}{Initial values of z. If NULL, z will be randomly sampled. Default NULL.}

\item{logfile}{If NULL, messages will be displayed as normal. If set to a file name, messages will be redirected messages to the file. Default NULL.}
}
\value{
An object of class celda_C with clustering results and Gibbs sampling statistics
}
\description{
celda Cell Clustering Model
}
