% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{celda_CG}
\alias{celda_CG}
\title{celda Cell and Gene Clustering Model}
\usage{
celda_CG(counts, sample.label = NULL, K, L, alpha = 1, beta = 1,
  delta = 1, gamma = 1, algorithm = c("EM", "Gibbs"), stop.iter = 10,
  max.iter = 200, split.on.iter = 10, split.on.last = TRUE,
  seed = 12345, nchains = 3, count.checksum = NULL, z.init = NULL,
  y.init = NULL, logfile = NULL)
}
\arguments{
\item{counts}{A numeric count matrix.}

\item{sample.label}{A vector indicating the sample for each cell in the count matrix}

\item{K}{The number of cell populations}

\item{L}{The number of gene clusters being considered}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell. Default 1.}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state. Default 1.}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state. Default 1.}

\item{algorithm}{Use 'EM' or 'Gibbs' sampling for clustering of cells into subpopulations. EM is much faster for larger datasets. Default 'EM'.}

\item{stop.iter}{Number of iterations without improvement in the log likelihood to stop the Gibbs sampler. Default 10.}

\item{max.iter}{Maximum iterations of Gibbs sampling to perform regardless of convergence. Default 200.}

\item{split.on.iter}{On every 'split.on.iter' iteration, a heuristic will be applied to determine if a gene/cell cluster should be reassigned and another gene/cell cluster should be split into two clusters. Default 10.}

\item{split.on.last}{After the the chain has converged according to 'stop.iter', a heuristic will be applied to determine if a gene/cell cluster should be reassigned and another gene/cell cluster should be split into two clusters. If a split occurs, then 'stop.iter' will be reset. Default TRUE.}

\item{seed}{Parameter to set.seed() for random number generation. Default 12345.}

\item{nchains}{Number of random z/y initializations. Default 3.}

\item{count.checksum}{An MD5 checksum for the provided counts matrix. Default NULL.}

\item{z.init}{Initial values of z. If NULL, z will be randomly sampled. Default NULL.}

\item{y.init}{Initial values of y. If NULL, y will be randomly sampled. Default NULL.}

\item{logfile}{The name of the logfile to redirect messages to.}
}
\description{
celda Cell and Gene Clustering Model
}
