% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celda_G}
\alias{celda_G}
\title{Cluster Genes from Single Cell Sequencing Data}
\usage{
celda_G(counts, L, beta = 1, delta = 1, gamma = 1, max.iter = 25,
  seed = 12345, best = TRUE, y.split.on.iter = 3, y.num.splits = 3,
  thread = 1, save.prob = FALSE, save.history = FALSE, ...)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{max.iter}{Maximum iterations of Gibbs sampling to perform. Defaults to 25.}

\item{seed}{Parameter to set.seed() for random number generation}

\item{best}{Whether to return the cluster assignment with the highest log-likelihood. Defaults to TRUE. Returns last generated cluster assignment when FALSE.}

\item{thread}{The thread index, used for logging purposes}

\item{save.prob}{Logical; whether to return the history of cluster assignment probabilities. Defaults to FALSE}

\item{save.history}{Logical; whether to return the history of cluster assignments. Defaults to FALSE}

\item{k}{The number of clusters to generate}

\item{a}{Vector of non-zero concentration parameters for sample <-> cluster assignment Dirichlet distribution}

\item{b}{Vector of non-zero concentration parameters for cluster <-> gene assignment Dirichlet distribution}

\item{g}{Number of cell states ("topics")}

\item{min.cell}{Desired minimum number of cells per cluster}

\item{kick}{Whether to randomize cluster assignments when a cluster has fewer than min.cell cells assigned to it during Gibbs sampling. (TODO param currently unused?)}
}
\description{
geneCluster provides cluster assignments for all genes in a provided single-cell 
sequencing count matrix, using the celda Bayesian hierarchical model.
}
\examples{
TODO
}
\keyword{LDA}
\keyword{clustering}
\keyword{gene}
\keyword{gibbs}

