% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celda_G}
\alias{celda_G}
\title{celda Gene Clustering Model}
\usage{
celda_G(counts, L, beta = 1, delta = 1, gamma = 1, max.iter = 50,
  count.checksum = NULL, seed = 12345, split.on.iter = 5,
  num.splits = 5, y.init = NULL, logfile = NULL, ...)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{L}{The number of clusters to generate}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell.}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state. Default to 1.}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state.}

\item{max.iter}{Maximum iterations of Gibbs sampling to perform. Defaults to 25.}

\item{count.checksum}{An MD5 checksum for the provided counts matrix}

\item{seed}{Parameter to set.seed() for random number generation.}

\item{split.on.iter}{On every split.on.iter iteration, a heuristic will be applied to determine if a gene cluster should be reassigned and another gene cluster should be split into two clusters. Default to be 5.}

\item{num.splits}{Maximum number of times to perform the heuristic described in split.on.iter. Default 5.}

\item{y.init}{Initial values of y. If NULL, y will be randomly sampled. Default NULL.}

\item{logfile}{The name of the logfile to redirect messages to.}

\item{...}{Additional parameters}
}
\description{
Provides cluster assignments for all genes in a provided single-cell 
sequencing count matrix, using the celda Bayesian hierarchical model.
}
\keyword{LDA}
\keyword{clustering}
\keyword{gene}
\keyword{gibbs}
