% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celda_G}
\alias{celda_G}
\title{celda Gene Clustering Model}
\usage{
celda_G(counts, L, beta = 1, delta = 1, gamma = 1, max.iter = 25,
  count.checksum = NULL, seed = 12345, y.split.on.iter = 3,
  y.num.splits = 3, logfile = NULL, ...)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{L}{The number of clusters to generate}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell.}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state. Default to 1.}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state.}

\item{max.iter}{Maximum iterations of Gibbs sampling to perform. Defaults to 25.}

\item{count.checksum}{An MD5 checksum for the provided counts matrix}

\item{seed}{Parameter to set.seed() for random number generation.}

\item{y.split.on.iter}{On every y.split.on.iter iteration, a heuristic will be applied using hierarchical clustering to determine if a gene cluster should be merged with another gene cluster and a third gene cluster should be split into two clusters. This helps avoid local optimum during the initialization. Default to be 3.}

\item{y.num.splits}{Maximum number of times to perform the heuristic described in y.split.on.iter.}

\item{logfile}{The name of the logfile to redirect messages to.}

\item{...}{Additional parameters}
}
\description{
Provides cluster assignments for all genes in a provided single-cell 
sequencing count matrix, using the celda Bayesian hierarchical model.
}
\keyword{LDA}
\keyword{clustering}
\keyword{gene}
\keyword{gibbs}
