% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celda_G}
\alias{celda_G}
\title{celda Gene Clustering Model}
\usage{
celda_G(counts, L, beta = 1, delta = 1, gamma = 1, stop.iter = 10,
  max.iter = 200, split.on.iter = 10, split.on.last = TRUE,
  seed = 12345, nchains = 3, count.checksum = NULL, y.init = NULL,
  logfile = NULL)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{L}{The number of clusters to generate}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell.}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state. Default to 1.}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state.}

\item{stop.iter}{Number of iterations without improvement in the log likelihood to stop the Gibbs sampler. Default 10.}

\item{max.iter}{Maximum iterations of Gibbs sampling to perform regardless of convergence. Default 200.}

\item{split.on.iter}{On every 'split.on.iter' iteration, a heuristic will be applied to determine if a gene/cell cluster should be reassigned and another gene/cell cluster should be split into two clusters. Default 10.}

\item{split.on.last}{After the the chain has converged according to 'stop.iter', a heuristic will be applied to determine if a gene/cell cluster should be reassigned and another gene/cell cluster should be split into two clusters. If a split occurs, then 'stop.iter' will be reset. Default TRUE.}

\item{seed}{Parameter to set.seed() for random number generation. Default 12345.}

\item{nchains}{Number of random y initializations. Default 3.}

\item{count.checksum}{An MD5 checksum for the provided counts matrix. Default NULL.}

\item{y.init}{Initial values of y. If NULL, y will be randomly sampled. Default NULL.}

\item{logfile}{The name of the logfile to redirect messages to.}
}
\description{
Provides cluster assignments for all genes in a provided single-cell 
sequencing count matrix, using the celda Bayesian hierarchical model.
}
\keyword{LDA}
\keyword{clustering}
\keyword{gene}
\keyword{gibbs}
