% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_heatmap.R
\name{celda_heatmap}
\alias{celda_heatmap}
\title{plot the heatmap of the counts data}
\usage{
celda_heatmap(counts, K, z, L, y, col = "YlOrBr")
}
\arguments{
\item{counts}{the counts matrix}

\item{K}{The number of clusters being considered  (Question1)or: Total number of cell populations??}

\item{z}{A numeric vector of cluster assignments for cell}

\item{L}{Total number of transcriptional states}

\item{y}{A numeric vector of cluster assignments for gene}

\item{col}{vector of colors used in heatmap}

\item{cluster_gene}{##  has to set to be FALSE in pheatmap --> need to improve this ourself}

\item{cluster_cell}{##  has to set to be FALSE in pheatmap --> need to improve this ourself}
}
\description{
plot the heatmap of the counts data
}
