% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{celda_heatmap}
\alias{celda_heatmap}
\title{Render a stylable heatmap of count data based on celda clustering results.}
\usage{
celda_heatmap(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{A celda model object (of class "celda_C", "celda_G" or "celda_CG")}

\item{counts}{the counts matrix}

\item{...}{extra parameters passed onto celda_heatmap}
}
\description{
Render a stylable heatmap of count data based on celda clustering results.
}
