% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaCGGridSearchRes}
\alias{celdaCGGridSearchRes}
\title{celdaCGGridSearchRes}
\format{
An object as returned from old celdaGridSearch()
}
\usage{
celdaCGGridSearchRes
}
\description{
Example results of old celdaGridSearch on celdaCGSim
}
\keyword{datasets}
