% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaCGMod}
\alias{celdaCGMod}
\title{celdaCGmod}
\format{
A celda_CG object
#@examples
data(celdaCGSim)
celdaCGMod <- celda_CG(celdaCGSim$counts,
  K = celdaCGSim$K,
  L = celdaCGSim$L,
  nchains = 1)
}
\usage{
celdaCGMod
}
\description{
celda_CG model object generated from \code{celdaCGSim} using
 old \code{celda_CG} function.
}
\keyword{datasets}
