% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaCGMod}
\alias{celdaCGMod}
\title{celdaCGmod}
\format{A celda_CG object}
\usage{
celdaCGMod
}
\description{
celda_CG results generated from celdaCGSim
}
\examples{
celdaCGMod = celda_CG(celdaCGSim$counts,
    K = celdaCGSim$K,
    L = celdaCGSim$L,
    nchains = 1)
}
\keyword{datasets}
