% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celdaCGSim}
\alias{celdaCGSim}
\title{celdaCGSim}
\format{
A list of counts and properties as returned from old simulateCells().
}
\usage{
celdaCGSim
}
\description{
An deprecated example of simulated count matrix from the
 celda_CG model.
}
\keyword{datasets}
