% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{celdaGridSearch}
\alias{celdaGridSearch}
\title{Run the celda Bayesian hierarchical model on a matrix of counts.}
\usage{
celdaGridSearch(counts, model, sample.label = NULL, K = NULL, L = NULL,
  alpha = 1, beta = 1, delta = 1, gamma = 1, max.iter = 200,
  z.init = NULL, y.init = NULL, stop.iter = 10, split.on.iter = 10,
  nchains = 1, bestChainsOnly = TRUE, cores = 1, seed = 12345,
  verbose = FALSE, logfile_prefix = "Celda")
}
\arguments{
\item{counts}{A count matrix}

\item{model}{Which celda sub-model to run. Options include "celda_C" (cell clustering), "celda_G" (gene clustering), "celda_CG" (gene and cell clustering)}

\item{sample.label}{A numeric vector, character vector, or factor indicating the originating sample for each cell (column) in the count matrix. By default, every cell will be assumed to be from an independent sample.}

\item{K}{Number of desired cell subpopulation clusters. Required for celda_C and celda_CG models.}

\item{L}{Number of desired gene clusters. Required for celda_G and celda_CG models.}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution (celda_C / celda_CG only)}

\item{beta}{Non-zero concentration parameter for gene Dirichlet distribution}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state (celda_G / celda_CG only)}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state (celda_G / celda_CG only)}

\item{max.iter}{The maximum number of iterations}

\item{z.init}{Initial values of z. If NULL, z will be randomly sampled. Default NULL. (celda_C / celda_CG only)}

\item{y.init}{Initial values of y. If NULL, y will be randomly sampled. Default NULL. (celda_G / celda_CG only)}

\item{stop.iter}{Number of iterations without improvement in the log likelihood to stop the Gibbs sampler. Default 10.}

\item{split.on.iter}{On every 'split.on.iter' iteration, a heuristic will be applied to determine if a gene/cell cluster should be reassigned and another gene/cell cluster should be split into two clusters. Default 10.}

\item{nchains}{The number of chains of Gibbs sampling to run for every combination of K/L parameters. Defaults to 1 (celda_G / celda_CG only)}

\item{bestChainsOnly}{Return only the best chain (by final log-likelihood) per K/L combination.}

\item{cores}{The number of cores to use for parallell Gibbs sampling. Defaults to 1.}

\item{seed}{The base seed for random number generation}

\item{verbose}{Print log messages during celda chain execution}

\item{logfile_prefix}{Prefix for log files from worker threads and main process.}
}
\value{
Object of class "celda_list", which contains results for all model parameter combinations and summaries of the run parameters
}
\description{
Yields assigments of genes/cells to clusters, depending on the provided model type.
}
