% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celdaHeatmap.celda_C}
\alias{celdaHeatmap.celda_C}
\title{celdaHeatmap for celda Cell clustering function}
\usage{
\method{celdaHeatmap}{celda_C}(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C".}

\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{...}{Additional parameters.}
}
\description{
celdaHeatmap for celda Cell clustering function
}
