% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celdaHeatmap.celda_C}
\alias{celdaHeatmap.celda_C}
\title{celdaHeatmap for celda Cell clustering function}
\usage{
\method{celdaHeatmap}{celda_C}(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_C"}

\item{counts}{A numeric count matrix}

\item{...}{extra parameters passed onto the render_celda_heatmap}
}
\description{
celdaHeatmap for celda Cell clustering function
}
