% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celdaHeatmap.celda_G}
\alias{celdaHeatmap.celda_G}
\title{celdaHeatmap for celda Gene clustering model}
\usage{
\method{celdaHeatmap}{celda_G}(counts, celda.mod, nfeatures = 25, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_G".}

\item{nfeatures}{Integer. Maximum number of features to select for each module. Default 25.}

\item{...}{Additional parameters.}
}
\value{
list A list containing dendrogram information and the heatmap grob
}
\description{
celdaHeatmap for celda Gene clustering model
}
\examples{
celda.sim = simulateCells("celda_G")
celda.mod = celda_G(celda.sim$counts, L=celda.sim$L, nchains=1, max.iter=1)
celdaHeatmap(celda.sim$counts, celda.mod)
}
