% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celdaHeatmap.celda_G}
\alias{celdaHeatmap.celda_G}
\title{celdaHeatmap for celda Gene clustering model}
\usage{
\method{celdaHeatmap}{celda_G}(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_G".}

\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{...}{Additional parameters.}
}
\description{
celdaHeatmap for celda Gene clustering model
}
