% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{celdaHeatmap}
\alias{celdaHeatmap}
\title{Render a stylable heatmap of count data based on celda clustering results.}
\usage{
celdaHeatmap(celda.mod, counts, ...)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}

\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{...}{Additional parameters.}
}
\description{
Render a stylable heatmap of count data based on celda clustering results.
}
