% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{celdaHeatmap}
\alias{celdaHeatmap}
\title{Render a stylable heatmap of count data based on celda clustering results.}
\usage{
celdaHeatmap(counts, celda.mod, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}

\item{...}{Additional parameters.}
}
\value{
list A list containing dendrogram information and the heatmap grob
}
\description{
Render a stylable heatmap of count data based on celda clustering results.
}
\examples{
celda.sim = simulateCells("celda_CG")
celda.mod = celda_CG(celda.sim$counts, K=celda.sim$K, L=celda.sim$L,
                     nchains=1, max.iter=1)
celdaHeatmap(celda.sim$counts, celda.mod)
}
