% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaHeatmap}
\alias{celdaHeatmap}
\title{Render a stylable heatmap of count data based on celda clustering results.}
\usage{
celdaHeatmap(counts, celda.mod, feature.ix, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{A celda model object of class "celda_C", "celda_G", or "celda_CG".}

\item{feature.ix}{Integer vector. Select features for display in heatmap. If NULL, no subsetting will be performed. Default NULL.}

\item{...}{Additional parameters.}
}
\value{
list A list containing dendrogram information and the heatmap grob
}
\description{
Render a stylable heatmap of count data based on celda clustering results.
}
\examples{
celdaHeatmap(celda.CG.sim$counts, celda.CG.mod)
}
