% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaHeatmap}
\alias{celdaHeatmap}
\title{Plot celda Heatmap}
\usage{
celdaHeatmap(counts, celdaMod, featureIx, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{A celdaModel object of class "celda_C", "celda_G", or
"celda_CG".}

\item{featureIx}{Integer vector. Select features for display in heatmap. If
NULL, no subsetting will be performed. Default NULL.}

\item{...}{Additional parameters.}
}
\value{
list A list containing dendrogram information and the heatmap grob
}
\description{
Render a stylable heatmap of count data based on celda
 clustering results.
}
\examples{
celdaHeatmap(celdaCGSim$counts, celdaCGMod)
}
