% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decontx.R
\name{celdaMarkerPlot}
\alias{celdaMarkerPlot}
\title{Barplot that shows the percentage of cells representing given marker genes}
\usage{
celdaMarkerPlot(
  counts,
  z,
  geneMarkers,
  threshold = 1,
  color = "red3",
  textLabelSize = 3,
  precision = 2
)
}
\arguments{
\item{counts}{Count matrix.}

\item{z}{A vector. Cell cluster label, same length as the column numbers of the count matrix}

\item{geneMarkers}{A dataframe of marker genes which corresponding to their cell types}

\item{threshold}{Default 1.}

\item{color, }{default as "red3"}

\item{textLabelSize, }{default as 3}

\item{precision, }{default as 2. Precision of percentage of cells showing the marker gene shown on the barplot}
}
\description{
Barplot that shows the percentage of cells representing given marker genes
}
