% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaModel}
\alias{celdaModel}
\alias{celdaModel,SingleCellExperiment-method}
\title{Get celda model from a celda
 \link[SingleCellExperiment]{SingleCellExperiment} object}
\usage{
celdaModel(sce)

\S4method{celdaModel}{SingleCellExperiment}(sce)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
returned by \link{celda_C}, \link{celda_G}, or \link{celda_CG}.}
}
\value{
Character. The celda model. Can be one of "celda_C", "celda_G", or
 "celda_CG".
}
\description{
Return the celda model for \code{sce} returned by
 \link{celda_C}, \link{celda_G} or \link{celda_CG}.
}
\examples{
data(sceCelda_CG)
celdaModel(sceCelda_CG)
}
