% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{celdaModel}
\alias{celdaModel}
\alias{celdaModel,SingleCellExperiment-method}
\title{Get celda model from a celda
 \link[SingleCellExperiment]{SingleCellExperiment} object}
\usage{
celdaModel(sce, altExpName = "featureSubset")

\S4method{celdaModel}{SingleCellExperiment}(sce, altExpName = "featureSubset")
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
returned by \link{celda_C}, \link{celda_G}, or \link{celda_CG}.}

\item{altExpName}{The name for the \link{altExp} slot
to use. Default "featureSubset".}
}
\value{
Character. The celda model. Can be one of "celda_C", "celda_G", or
 "celda_CG".
}
\description{
Return the celda model for \code{sce} returned by
 \link{celda_C}, \link{celda_G} or \link{celda_CG}.
}
\examples{
data(sceCeldaCG)
celdaModel(sceCeldaCG)
}
