% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{celdaModules}
\alias{celdaModules}
\alias{celdaModules,SingleCellExperiment-method}
\alias{celdaModules<-}
\alias{celdaModules<-,SingleCellExperiment-method}
\title{Get or set the feature module labels from a celda
 \linkS4class{SingleCellExperiment} object.}
\usage{
celdaModules(sce, altExpName)

\S4method{celdaModules}{SingleCellExperiment}(sce, altExpName = "featureSubset")

celdaModules(sce, altExpName) <- value

\S4method{celdaModules}{SingleCellExperiment}(sce, altExpName = "featureSubset") <- value
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object returned by
\link{celda_G}, or \link{celda_CG}, with the matrix
located in the \code{useAssay} assay slot.
Rows represent features and columns represent cells.}

\item{altExpName}{The name for the \link[SingleCellExperiment]{altExp} slot
to use. Default "featureSubset".}

\item{value}{Character vector of feature module labels for replacements.
Works only if \code{x} is a \linkS4class{SingleCellExperiment} object.}
}
\value{
Character vector. Contains feature module labels for each
 feature in x.
}
\description{
Return or set the feature module cluster labels determined
 by \link{celda_G} or \link{celda_CG} models.
}
\examples{
data(sceCeldaCG)
celdaModules(sceCeldaCG)
}
