% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaPerplexity}
\alias{celdaPerplexity}
\title{Get perplexity for every model in a celdaList}
\usage{
celdaPerplexity(celdaMod)
}
\arguments{
\item{celdaMod}{A celdaModel object of class "celda_C", "celda_G", or
"celda_CG".}
}
\value{
List. Contains one celdaModel object for each of the parameters
 specified in the `runParams()` of the provided celda list.
}
\description{
Returns perplexity for each model in a celdaList as calculated
 by `perplexity().`
}
\examples{
data(celdaCGGridSearchRes)
celdaCGGridModelPerplexities <- celdaPerplexity(celdaCGGridSearchRes)
}
