% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\docType{methods}
\name{celdaProbabilityMap,celda_C-method}
\alias{celdaProbabilityMap,celda_C-method}
\title{Probability map for a celda_C model}
\usage{
\S4method{celdaProbabilityMap}{celda_C}(counts, celda.mod,
  level = c("sample"), ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class `celda_C`.}

\item{level}{Character. 'sample' will display the absolute probabilities and relative normalized abundance of each cell population in each sample. Default 'sample'.}

\item{...}{Additional parameters.}
}
\value{
A grob containing the specified plots
}
\description{
Renders probability and relative expression heatmaps to visualize the relationship between cell populations and samples.
}
\examples{
celdaProbabilityMap(celda.C.sim$counts, celda.C.mod)
}
\seealso{
`celda_C()` for clustering cells
}
