% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{celdaProbabilityMap}
\alias{celdaProbabilityMap}
\title{Generate factorized matrices showing each feature's influence on cell / gene clustering}
\usage{
celdaProbabilityMap(counts, celda.mod, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C" or "celda_CG".}
}
\description{
Generate factorized matrices showing each feature's influence on cell / gene clustering
}
