% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaProbabilityMap.R
\name{celdaProbabilityMap}
\alias{celdaProbabilityMap}
\alias{celdaProbabilityMap,SingleCellExperiment-method}
\title{Probability map for a celda model}
\usage{
celdaProbabilityMap(sce, ...)

\S4method{celdaProbabilityMap}{SingleCellExperiment}(
  sce,
  useAssay = "counts",
  altExpName = "featureSubset",
  level = c("cellPopulation", "sample")
)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
returned by \link{celda_C}, \link{celda_G}, or \link{celda_CG}.}

\item{...}{Additional parameters.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use. Default "counts".}

\item{altExpName}{The name for the \link[SingleCellExperiment]{altExp} slot
to use. Default "featureSubset".}

\item{level}{Character. One of "cellPopulation" or "Sample".
"cellPopulation" will display the absolute probabilities and relative
normalized expression of each module in each cell population.
\strong{\code{level = "cellPopulation"} only works for celda_CG \code{sce}
objects}. "sample" will display the absolute probabilities and relative
normalized abundance of each cell population in each sample. Default
"cellPopulation".}
}
\value{
A grob containing the specified plots
}
\description{
Renders probability and relative expression heatmaps to
 visualize the relationship between features and cell populations (or cell
 populations and samples).
}
\examples{
data(sceCeldaCG)
celdaProbabilityMap(sceCeldaCG)
}
\seealso{
\link{celda_C} for clustering cells. \link{celda_CG} for
 clustering features and cells
}
