% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaProbabilityMap}
\alias{celdaProbabilityMap}
\title{Renders probability and relative expression heatmaps to visualize the relationship between feature modules and cell populations.}
\usage{
celdaProbabilityMap(counts, celda.mod, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C" or "celda_CG".}

\item{...}{Additional parameters.}
}
\value{
A grob containing the specified plots
}
\description{
It is often useful to visualize to what degree each feature influences each 
cell cluster. This can also be useful for identifying features which may
be redundant or unassociated with cell clustering.
}
\examples{
celdaProbabilityMap(celda.CG.sim$counts, celda.CG.mod)
}
