% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\docType{methods}
\name{celdaTsne,celda_CG-method}
\alias{celdaTsne,celda_CG-method}
\title{tSNE for celda_CG}
\usage{
\S4method{celdaTsne}{celda_CG}(counts, celda.mod, max.cells = 25000,
  min.cluster.size = 100, initial.dims = 20, modules = NULL,
  perplexity = 20, max.iter = 2500, seed = 12345, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class `celda_CG`.}

\item{max.cells}{Integer. Maximum number of cells to plot. Cells will be randomly subsampled if ncol(counts) > max.cells. Larger numbers of cells requires more memory. Default 25000.}

\item{min.cluster.size}{Integer. Do not subsample cell clusters below this threshold. Default 100.}

\item{modules}{Integer vector. Determines which features modules to use for tSNE. If NULL, all modules will be used. Default NULL.}

\item{perplexity}{Numeric. Perplexity parameter for tSNE. Default 20.}

\item{max.iter}{Integer. Maximum number of iterations in tSNE generation. Default 2500.}

\item{seed}{Integer. Passed to `set.seed()`. Default 12345.}

\item{...}{Additional parameters.}
}
\value{
A two column matrix of t-SNE coordinates
}
\description{
Embeds cells in two dimensions using tSNE based on a `celda_CG` model. tSNE is run on module probabilities to reduce the number of features instead of using PCA. Module probabilities square-root trasformed before applying tSNE.
}
\examples{
tsne.res = celdaTsne(celda.CG.sim$counts, celda.CG.mod)
}
\seealso{
`celda_CG()` for clustering features and cells  and `celdaHeatmap()` for displaying expression
}
