% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celdaTsne.celda_C}
\alias{celdaTsne.celda_C}
\title{Embeds cells in two dimensions using tSNE based on celda_C results.}
\usage{
\method{celdaTsne}{celda_C}(counts, celda.mod, max.cells = 10000,
  min.cluster.size = 100, initial.dims = 20, perplexity = 20,
  max.iter = 2500, seed = 12345, ...)
}
\arguments{
\item{counts}{Counts matrix, should have cell name for column name and gene name for row name.}

\item{celda.mod}{Celda model to use for tsne.}

\item{max.cells}{Integer; Maximum number of cells to plot. Cells will be randomly subsampled if ncol(conts) > max.cells. Larger numbers of cells requires more memory. Default 10000.}

\item{min.cluster.size}{Integer; Do not subsample cell clusters below this threshold. Default 100.}

\item{initial.dims}{PCA will be used to reduce the dimentionality of the dataset. The top 'initial.dims' principal components will be used for tSNE.}

\item{perplexity}{Numeric vector; determines perplexity for tSNE. Default 20.}

\item{max.iter}{Numeric vector; determines iterations for tsne. Default 1000.}

\item{seed}{Seed for random number generation. Default 12345.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Embeds cells in two dimensions using tSNE based on celda_C results.
}
