% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celdaTsne.celda_G}
\alias{celdaTsne.celda_G}
\title{Embeds cells in two dimensions using tSNE based on celda_CG results.}
\usage{
\method{celdaTsne}{celda_G}(counts, celda.mod, states = NULL,
  perplexity = 20, max.iter = 2500, distance = "hellinger",
  seed = 12345, ...)
}
\arguments{
\item{counts}{Counts matrix, should have cell name for column name and gene name for row name.}

\item{celda.mod}{Celda model to use for tsne.}

\item{states}{Numeric vector; determines which gene states to use for tSNE. If NULL, all states will be used. Default NULL.}

\item{perplexity}{Numeric vector; determines perplexity for tSNE. Default 20.}

\item{max.iter}{Numeric vector; determines iterations for tsne. Default 1000.}

\item{distance}{Character vector; determines which distance metric to use for tSNE. One of 'hellinger', 'cosine', 'spearman'.}

\item{seed}{Seed for random number generation. Default 12345.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Embeds cells in two dimensions using tSNE based on celda_CG results.
}
