% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{celdaTsne.celda_G}
\alias{celdaTsne.celda_G}
\title{Embeds cells in two dimensions using tSNE based on celda_CG results.}
\usage{
\method{celdaTsne}{celda_G}(counts, celda.mod, max.cells = 10000,
  modules = NULL, perplexity = 20, max.iter = 2500, seed = 12345, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_G".}

\item{max.cells}{Integer. Maximum number of cells to plot. Cells will be randomly subsampled if ncol(conts) > max.cells. Larger numbers of cells requires more memory. Default 10000.}

\item{modules}{Integer vector. Determines which feature modules to use for tSNE. If NULL, all modules will be used. Default NULL.}

\item{perplexity}{Numeric. Perplexity parameter for tSNE. Default 20.}

\item{max.iter}{Integer. Maximum number of iterations in tSNE generation. Default 2500.}

\item{seed}{Integer. Passed to set.seed(). Default 12345.}

\item{...}{Additional parameters.}
}
\description{
Embeds cells in two dimensions using tSNE based on celda_CG results.
}
