% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{celdaTsne}
\alias{celdaTsne}
\title{Runs tSNE via Rtsne based on the CELDA model and specified cell states.}
\usage{
celdaTsne(counts, celda.mod, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda model. Options available in `celda::available.models`.}

\item{...}{Unused parameters.}
}
\description{
Runs tSNE via Rtsne based on the CELDA model and specified cell states.
}
