% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/celdatSNE.R
\name{celdaTsne}
\alias{celdaTsne}
\alias{celdaTsne,SingleCellExperiment-method}
\title{Embeds cells in two dimensions using tSNE based on celda_CG results.}
\usage{
celdaTsne(sce, ...)

celdaTsne(sce, ...)

\S4method{celdaTsne}{SingleCellExperiment}(
  sce,
  useAssay = "counts",
  maxCells = NULL,
  minClusterSize = 100,
  initialDims = 20,
  modules = NULL,
  perplexity = 20,
  maxIter = 2500,
  seed = 12345
)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object
returned by \link{celda_C}, \link{celda_G}, or \link{celda_CG}.}

\item{...}{Additional parameters.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use. Default "counts".}

\item{maxCells}{Integer. Maximum number of cells to plot. Cells will be
randomly subsampled if \code{ncol(counts) > maxCells}. Larger numbers of
cells requires more memory. If \code{NULL}, no subsampling will be
performed. Default \code{NULL}.}

\item{minClusterSize}{Integer. Do not subsample cell clusters below this
threshold. Default 100.}

\item{initialDims}{Integer. PCA will be used to reduce the dimentionality
of the dataset. The top 'initialDims' principal components will be used
for tSNE. Default 20.}

\item{modules}{Integer vector. Determines which feature modules to use for
tSNE. If \code{NULL}, all modules will be used. Default \code{NULL}.}

\item{perplexity}{Numeric. Perplexity parameter for tSNE. Default 20.}

\item{maxIter}{Integer. Maximum number of iterations in tSNE generation.
Default 2500.}

\item{seed}{Integer. Passed to \link[withr]{with_seed}. For reproducibility,
a default value of 12345 is used. If NULL, no calls to
\link[withr]{with_seed} are made.}

\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Celda object of class `celda_CG`.}
}
\value{
Numeric Matrix of dimension `ncol(counts)` x 2, colums representing
 the "X" and "Y" coordinates in the data's t-SNE represetation.

\code{sce} with t-SNE coordinates
 (columns "celda_tSNE1" & "celda_tSNE2") added to
 \code{\link[SummarizedExperiment]{colData}(sce)}.
}
\description{
Embeds cells in two dimensions using \link[Rtsne]{Rtsne} based
 on a celda model. For celda_C \code{sce} objects, PCA on the normalized
 counts is used to reduce the number of features before applying t-SNE. For
 celda_CG and celda_G \code{sce} objects, tSNE is run on module
 probabilities to reduce the number of features instead of using PCA.
 Module probabilities are square-root transformed before applying tSNE.
}
\examples{
data(celdaCGSim, celdaCGMod)
tsneRes <- celdaTsne(celdaCGSim$counts, celdaCGMod)
data(sceCeldaCG)
tsneRes <- celdaTsne(sceCeldaCG)
}
