% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaTsne}
\alias{celdaTsne}
\title{Embeds cells in two dimensions using tSNE based on celda_CG results.}
\usage{
celdaTsne(counts, celda.mod, max.cells = 25000, min.cluster.size = 100,
  initial.dims = 20, modules = NULL, perplexity = 20,
  max.iter = 2500, seed = 12345, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda model. Options available in `celda::available.models`.}

\item{...}{Additional parameters.}
}
\value{
Numeric Matrix of dimension `ncol(counts)` x 2, colums representing the "X" and "Y" coordinates in the data's t-SNE represetation.
}
\description{
Embeds cells in two dimensions using tSNE based on celda_CG results.
}
\examples{
tsne.res = celdaTsne(celda.CG.sim$counts, celda.CG.mod)
}
