% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{celdaTsne}
\alias{celdaTsne}
\title{Runs tSNE via Rtsne based on the CELDA model and specified cell states.}
\usage{
celdaTsne(counts, celda.mod, ...)
}
\arguments{
\item{counts}{Counts matrix, should have cell name for column name and gene name for row name.}

\item{celda.mod}{Celda model to use for tsne. class "celda_C","celda_G" or "celda_CG".}

\item{...}{Other arguments to be passed to model-specific tSNE functions. Use methods("celdaTsne") for list of 'celdaTsne' functions.}
}
\description{
Runs tSNE via Rtsne based on the CELDA model and specified cell states.
}
