% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{celdaUmap,celda_C-method}
\alias{celdaUmap,celda_C-method}
\title{umap for celda_C}
\usage{
\S4method{celdaUmap}{celda_C}(
  counts,
  celdaMod,
  maxCells = NULL,
  minClusterSize = 100,
  seed = 12345,
  nNeighbors = 30,
  minDist = 0.75,
  spread = 1,
  pca = TRUE,
  initialDims = 50,
  cores = 1,
  ...
)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Celda object of class `celda_C`.}

\item{maxCells}{Integer. Maximum number of cells to plot. Cells will be
randomly subsampled if ncol(counts) > maxCells. Larger numbers of cells
requires more memory. If NULL, no subsampling will be performed.
Default NULL.}

\item{minClusterSize}{Integer. Do not subsample cell clusters below this
threshold. Default 100.}

\item{seed}{Integer. Passed to \link[withr]{with_seed}. For reproducibility,
a default value of 12345 is used. If NULL, no calls to
\link[withr]{with_seed} are made.}

\item{nNeighbors}{The size of local neighborhood used for
manifold approximation. Larger values result in more global
views of the manifold, while smaller values result in more
local data being preserved. Default 30.
See `?uwot::umap` for more information.}

\item{minDist}{The effective minimum distance between embedded points.
Smaller values will result in a more clustered/clumped
embedding where nearby points on the manifold are drawn
closer together, while larger values will result on a more
even dispersal of points. Default 0.2.
See `?uwot::umap` for more information.}

\item{spread}{The effective scale of embedded points. In combination with
‘min_dist’, this determines how clustered/clumped the
embedded points are. Default 1. See `?uwot::umap` for more information.}

\item{pca}{Logical. Whether to perform
dimensionality reduction with PCA before UMAP.}

\item{initialDims}{Integer. Number of dimensions from PCA to use as
input in UMAP. Default 50.}

\item{cores}{Number of threads to use. Default 1.}

\item{...}{Other parameters to pass to `uwot::umap`.}
}
\value{
A two column matrix of UMAP coordinates
}
\description{
Embeds cells in two dimensions using umap based on a `celda_C`
 model. PCA on the normalized counts is used to reduce the number of features
 before applying umap.
}
\examples{
data(celdaCSim, celdaCMod)
umapRes <- celdaUmap(celdaCSim$counts, celdaCMod)
}
\seealso{
`celda_C()` for clustering cells and `celdaHeatmap()` for displaying
 expression.
}
