% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaUmap}
\alias{celdaUmap}
\title{Embeds cells in two dimensions using umap.}
\usage{
celdaUmap(counts, celda.mod, max.cells = 25000, min.cluster.size = 100,
  modules = NULL, umap.config = umap::umap.defaults)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class `celda_CG`.}

\item{max.cells}{Integer. Maximum number of cells to plot. Cells will be randomly subsampled if ncol(counts) > max.cells. Larger numbers of cells requires more memory. Default 25000.}

\item{min.cluster.size}{Integer. Do not subsample cell clusters below this threshold. Default 100.}

\item{modules}{Integer vector. Determines which features modules to use for tSNE. If NULL, all modules will be used. Default NULL.}

\item{perplexity}{Numeric. Perplexity parameter for tSNE. Default 20.}

\item{max.iter}{Integer. Maximum number of iterations in tSNE generation. Default 2500.}

\item{seed}{Integer. Passed to `set.seed()`. Default 12345. If NULL, no calls to `set.seed()` are made.}

\item{...}{Additional parameters.}

\item{...}{Additional parameters.}
}
\value{
Numeric Matrix of dimension `ncol(counts)` x 2, colums representing the "X" and "Y" coordinates in the data's t-SNE represetation.
}
\description{
Embeds cells in two dimensions using umap.
}
\examples{
tsne.res = celdaTsne(celda.CG.sim$counts, celda.CG.mod)
}
