% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaUmap}
\alias{celdaUmap}
\title{Embeds cells in two dimensions using umap.}
\usage{
celdaUmap(counts, celdaMod, maxCells = 25000, minClusterSize = 100,
  modules = NULL, umapConfig = umap::umap.defaults)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Celda object of class `celda_CG`.}

\item{maxCells}{Integer. Maximum number of cells to plot. Cells will be
randomly subsampled if ncol(counts) > maxCells. Larger numbers of cells
requires more memory. Default 25000.}

\item{minClusterSize}{Integer. Do not subsample cell clusters below this
threshold. Default 100.}

\item{modules}{Integer vector. Determines which features modules to use for
tSNE. If NULL, all modules will be used. Default NULL.}

\item{umapConfig}{An object of class "umapConfig" specifying parameters to
the UMAP algorithm.}

\item{...}{Additional parameters.}
}
\value{
Numeric Matrix of dimension `ncol(counts)` x 2, colums representing
 the "X" and "Y" coordinates in the data's t-SNE represetation.
}
\description{
Embeds cells in two dimensions using umap.
}
\examples{
tsneRes <- celdaUmap(celdaCGSim$counts, celdaCGMod)
}
