% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{celdaUmap}
\alias{celdaUmap}
\title{Embeds cells in two dimensions using umap.}
\usage{
celdaUmap(counts, celdaMod, maxCells = NULL, minClusterSize = 100,
  modules = NULL, seed = 12345, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Celda object of class `celda_CG`.}

\item{maxCells}{Integer. Maximum number of cells to plot. Cells will be
randomly subsampled if ncol(counts) > maxCells. Larger numbers of cells
requires more memory. Default 25000.}

\item{minClusterSize}{Integer. Do not subsample cell clusters below this
threshold. Default 100.}

\item{modules}{Integer vector. Determines which features modules to use for
tSNE. If NULL, all modules will be used. Default NULL.}

\item{seed}{Integer. Passed to \link[withr]{with_seed}. For reproducibility,
a default value of 12345 is used. If NULL, no calls to
\link[withr]{with_seed} are made.
the UMAP algorithm.}

\item{...}{Additional parameters to `uwot::umap`}
}
\value{
A two column matrix of UMAP coordinates#' @examples
data(celdaCGSim, celdaCGMod)
umapRes <- celdaUmap(celdaCGSim$counts, celdaCGMod)
}
\description{
Embeds cells in two dimensions using umap.
}
