% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{cG.calcGibbsProbY}
\alias{cG.calcGibbsProbY}
\title{Calculate Log Likelihood For Single Set of Cluster Assignments (Gene Clustering)}
\usage{
cG.calcGibbsProbY(n.TS.by.C, n.by.TS, nG.by.TS, nG.in.Y, L, beta, delta, gamma)
}
\arguments{
\item{beta}{Vector of non-zero concentration parameters for cluster <-> gene assignment Dirichlet distribution}

\item{gamma}{The number of cell states ("topics")}

\item{ix}{The index of the cell being assigned a cluster during the current iteration of Gibbs sampling}

\item{counts}{A numeric count matrix}

\item{z}{A numeric vector of cluster assignments}

\item{k}{The number of clusters being considered}

\item{alpha}{Vector of non-zero concentration parameters for sample <-> cluster assignment Dirichlet distribution}
}
\description{
This function calculates the log-likelihood of a given set of cluster assigments for the samples
represented in the provided count matrix.
}
\examples{
TODO
}
\keyword{likelihood}
\keyword{log}
