% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{cG.decomposeCounts}
\alias{cG.decomposeCounts}
\title{Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation}
\usage{
cG.decomposeCounts(counts, y, L)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{y}{Numeric vector. Denotes feature module labels.}

\item{L}{Integer. Number of feature modules.}
}
\description{
Takes raw counts matrix and converts it to a series of matrices needed for log likelihood calculation
}
