% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{clusterProbabilities}
\alias{clusterProbabilities}
\title{Get the probability of the cluster assignments generated during a celda run.}
\usage{
clusterProbabilities(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get the probability of the cluster assignments generated during a celda run.
}

