% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{clusterProbability.celda_C}
\alias{clusterProbability.celda_C}
\title{Calculates the conditional probability of each cell belong to each cluster given all other cluster assignments}
\usage{
clusterProbability.celda_C(counts, celda.mod, log = FALSE)
}
\arguments{
\item{counts}{The original count matrix used in the model}

\item{celda.mod}{A model returned from the 'celda_C' function}

\item{log}{If FALSE, then the normalized conditional probabilities will be returned. If TRUE, then the unnormalized log probabilities will be returned.}
}
\value{
A list containging a matrix for the conditional cell cluster probabilities.
}
\description{
Calculates the conditional probability of each cell belong to each cluster given all other cluster assignments
}
