% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{clusterProbability}
\alias{clusterProbability}
\title{Get cluster probability}
\usage{
clusterProbability(counts, celdaMod, log = FALSE, ...)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate `celdaMod`.}

\item{celdaMod}{celdaModel. Options available in `celda::availableModels`.}

\item{log}{Logical. If FALSE, then the normalized conditional probabilities
will be returned. If TRUE, then the unnormalized log probabilities will be
returned. Default FALSE.}

\item{...}{Additional parameters.}
}
\value{
A numeric vector of the cluster assignment probabilties
}
\description{
Get the probability of the cluster assignments generated during
 a celda run.
}
\examples{
clusterProb <- clusterProbability(celdaCGSim$counts, celdaCGMod)
}
