% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{clusters,celdaModel-method}
\alias{clusters,celdaModel-method}
\title{Get clustering outcomes from a celdaModel}
\usage{
\S4method{clusters}{celdaModel}(celdaMod)
}
\arguments{
\item{celdaMod}{celdaModel. Options available in `celda::availableModels`.}
}
\value{
List. Contains z (for celda_C and celdaCGModels) and/or y
 (for celda_G and celdaCGModels)
}
\description{
Returns the z / y results corresponding to the cell / gene
 cluster labels determined by the provided celdaModel.
}
\examples{
data(celdaCGMod)
clusters(celdaCGMod)
}
