% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{clusters}
\alias{clusters}
\alias{clusters,SingleCellExperiment-method}
\alias{clusters,celdaModel-method}
\alias{clusters<-}
\alias{clusters<-,SingleCellExperiment-method}
\title{Get or set the cell cluster labels from a celda
 \link[SingleCellExperiment]{SingleCellExperiment} object or celda model
 object.}
\usage{
clusters(x)

\S4method{clusters}{SingleCellExperiment}(x)

\S4method{clusters}{celdaModel}(x)

clusters(x) <- value

\S4method{clusters}{SingleCellExperiment}(x) <- value
}
\arguments{
\item{x}{Can be one of
\itemize{
\item A \linkS4class{SingleCellExperiment} object returned by
\link{celda_C}, or \link{celda_CG}, with the matrix
located in the \code{useAssay} assay slot.
Rows represent features and columns represent cells.
\item Celda model object.}}
}
\value{
One of
\itemize{
 \item Character vector if \code{x} is a
 \linkS4class{SingleCellExperiment} object.
 Contains cell cluster labels for each cell in x.
 \item List if \code{x} is a celda model object. Contains cell cluster
 labels (for celda_C and celdaCG
 Models) and/or feature module labels (for celda_G and celdaCG Models).}
}
\description{
Return or set the cell cluster labels determined
 by \link{celda_C} or \link{celda_CG} models.
}
\examples{
data(sceCeldaCG)
clusters(sceCeldaCG)
data(celdaCGMod)
clusters(celdaCGMod)
}
