% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{compareCountMatrix}
\alias{compareCountMatrix}
\title{Check whether a count matrix was the one used in a given celda run}
\usage{
compareCountMatrix(counts, celda.mod, error.on.mismatch = TRUE)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{celda.mod}{Celda model. Options available in `celda::available.models`.}

\item{error.on.mismatch}{Logical. Whether to stop execution in the event of a count matrix mismatch. Default TRUE.}
}
\value{
TRUE if provided count matrix matches the one used in the celda run, FALSE otherwise. Error on FALSE if error.on.mismatch is TRUE.
}
\description{
Ensures that the provided celda object was generated from a counts matrix
with similar dimensions to the one provided.
}
\details{
Then, compare the MD5 checksum of a provided count.matrix to the count matrix
checksum on a celda_list object, to see if they're the same.
}
\examples{
 compareCountMatrix(celda.CG.sim$counts, celda.CG.mod, 
                    error.on.mismatch=FALSE)
}
