% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{compareCountMatrix}
\alias{compareCountMatrix}
\title{Check whether a count matrix was the one used in a given celda run}
\usage{
compareCountMatrix(count.matrix, celda.obj, error.on.mismatch = TRUE)
}
\arguments{
\item{count.matrix}{A numeric matrix of counts}

\item{celda.obj}{A single celda model as generated by celda(), of class "celda_C", "celda_G", "celda_CG"}

\item{error.on.mismatch}{Whether to stop execution in the event of a count matrix mismatch. Defaults to TRUE.}
}
\value{
TRUE if provided count matrix matches the one used in the celda run, FALSE otherwise. Error on FALSE if error.on.mismatch is TRUE.
}
\description{
Ensures that the provided celda object was generated from a counts matrix
with similar dimensions to the one provided.
}
\details{
Then, compare the MD5 checksum of a provided count.matrix to the count matrix
checksum on a celda_list object, to see if they're the same.
}
