% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{completeClusterHistory}
\alias{completeClusterHistory}
\title{Get the complete history of gene / cell / gene & cell cluster assignments 
generated during a celda run, dependent on the model provided.}
\usage{
completeClusterHistory(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get the complete history of gene / cell / gene & cell cluster assignments 
generated during a celda run, dependent on the model provided.
}

