% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{completeLogLikelihood}
\alias{completeLogLikelihood}
\title{Get the complete log likelihood for each iteration of Gibbs sampling
for each chain generated during a celda run.}
\usage{
completeLogLikelihood(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get the complete log likelihood for each iteration of Gibbs sampling
for each chain generated during a celda run.
}
