% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{completeLogLikelihood}
\alias{completeLogLikelihood}
\title{Get the complete log likelihood for a given celda model.}
\usage{
completeLogLikelihood(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}
}
\description{
Get the complete log likelihood for a given celda model.
}
