% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contaminationSim}
\alias{contaminationSim}
\title{contaminationSim}
\format{
A list
}
\usage{
contaminationSim
}
\description{
Generated by simulateContaminatedMatrix
}
\details{
A toy contamination data generated by simulateContaminatedMatrix
}
\keyword{datasets}
