% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{DecontX}
\alias{DecontX}
\title{This function updates decontamination on dataset with multiple batches}
\usage{
DecontX(counts, z = NULL, batch = NULL, max.iter = 200,
  beta = 1e-06, delta = 10, logfile = NULL, verbose = TRUE,
  seed = 1234567)
}
\arguments{
\item{counts}{Numeric/Integer matrix. Observed count matrix, rows represent features and columns represent cells.}

\item{z}{Integer vector. Cell population labels. Default NULL}

\item{batch}{Integer vector. Cell batch labels. Default NULL}

\item{max.iter}{Integer. Maximum iterations of EM algorithm. Default to be 200}

\item{beta}{Numeric. Concentration parameter for Phi. Default to be 1e-6}

\item{delta}{Numeric. Symmetric concentration parameter for Theta. Default to be 10}

\item{logfile}{Character. Messages will be redirected to a file named `logfile`. If NULL, messages will be printed to stdout.  Default NULL}

\item{verbose}{Logical. Whether to print log messages. Default TRUE}

\item{seed}{Integer. Passed to set.seed(). Default to be 1234567. If NULL, no calls to `set.seed()` are made.}
}
\value{
A list object which contains the decontaminated count matrix and related parameters
}
\description{
This function updates decontamination on dataset with multiple batches
}
\examples{
decon.c = DecontX( counts = contamination.sim$rmat + contamination.sim$cmat, z=contamination.sim$z, max.iter=3)
decon.bg = DecontX( counts=contamination.sim$rmat + contamination.sim$cmat, max.iter=3 ) 
}
