% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{decontX}
\alias{decontX}
\title{DecontX}
\usage{
decontX(
  counts,
  z = NULL,
  batch = NULL,
  maxIter = 500,
  delta = 10,
  convergence = 0.001,
  convergenceStep = 10,
  logfile = NULL,
  verbose = TRUE,
  varGenes = 5000,
  L = 50,
  dbscanEps = 1,
  seed = 12345
)
}
\arguments{
\item{counts}{Numeric matrix. Observed count matrix, rows represent
features and columns represent cells.}

\item{z}{Integer vector. Cell cluster labels. If NULL, Celda will be used
to reduce the dimensionality of the dataset to 'L' modules,
'\link[uwot]{umap}' from the 'uwot' package  will be used to further
reduce the dataset to 2 dimenions and the '\link[dbscan]{dbscan}'
function from the 'dbscan' package will be used to identify clusters
of broad cell types. Default NULL.}

\item{batch}{Integer vector. Batch labels for cells. If batch labels 
are supplied, DecontX is run on cells from each batch separately.
Default NULL.}

\item{maxIter}{Integer. Maximum iterations of the EM algorithm. Default 500.}

\item{delta}{Numeric. Symmetric Dirichlet concentration parameter
to initialize theta. Default 10.}

\item{convergence}{Numeric. The EM algorithm will be stopped if the maximum
difference in the contamination estimates between the previous 'convergenceCheck'
is less than this. Default 0.001.}

\item{logfile}{Character. Messages will be redirected to a file named
`logfile`. If NULL, messages will be printed to stdout.  Default NULL.}

\item{verbose}{Logical. Whether to print log messages. Default TRUE.}

\item{varGenes}{Integer. The number of variable genes to use in
Celda clustering. Variability is calcualted using '\link[scran]{modelGeneVar}'
function from the 'scran' package. 
Used only when z is not provided. Default 5000.}

\item{L}{Integer. Number of modules for Celda clustering. Used to reduce
the dimensionality of the dataset before applying UMAP and dbscan.
Used only when z is not provided. Default 50.}

\item{dbscanEps}{Numeric. The clustering resolution parameter
used in '\link[dbscan]{dbscan}' to estimate broad cell clusters.
Used only when z is not provided. Default 1.}

\item{seed}{Integer. Passed to \link[withr]{with_seed}. For reproducibility,
a default value of 12345 is used. If NULL, no calls to
\link[withr]{with_seed} are made.}

\item{iterLogLik}{Integer. Every iteration to calculate log likelihood
Default 10.}
}
\value{
'decontX_counts' contains the decontaminated count matrix.
'contamination' contains the per-cell contamination estimates.
'batchEstimates' contains the estimated probability distributions
for each batch. 'z' contains the cell cluster labels. 'runParams'
contains a list of arguments used in the function call.
}
\description{
Identifies contamination from factors such as ambient RNA
in single cell genomic datasets.
}
\examples{
 s <- simulateContaminatedMatrix()
 result <- decontX(s$observedCounts, s$z)
 contamination <- colSums(s$observedCounts - s$nativeCounts) /
                    colSums(s$observedCounts)
 plot(contamination, result$contamination)
}
