% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{decontX}
\alias{decontX}
\title{Decontaminate count matrix}
\usage{
decontX(counts, z = NULL, batch = NULL, maxIter = 200, delta = 10,
  logfile = NULL, verbose = TRUE, seed = 1234567)
}
\arguments{
\item{counts}{Numeric/Integer matrix. Observed count matrix, rows represent
features and columns represent cells.}

\item{z}{Integer vector. Cell population labels. Default NULL.}

\item{batch}{Integer vector. Cell batch labels. Default NULL.}

\item{maxIter}{Integer. Maximum iterations of EM algorithm. Default to be
200.}

\item{delta}{Numeric. Symmetric concentration parameter for Theta. Default
to be 10.}

\item{logfile}{Character. Messages will be redirected to a file named
`logfile`. If NULL, messages will be printed to stdout.  Default NULL.}

\item{verbose}{Logical. Whether to print log messages. Default TRUE.}

\item{seed}{Integer. Passed to set.seed(). Default to be 1234567. If NULL,
no calls to `set.seed()` are made.}
}
\value{
A list object which contains the decontaminated count matrix and
 related parameters.
}
\description{
This function updates decontamination on dataset with multiple
 batches.
}
\examples{
deconC <- decontX(counts = contaminationSim$rmat + contaminationSim$cmat,
 z = contaminationSim$z, maxIter = 3)
deconBg <- decontX(counts = contaminationSim$rmat + contaminationSim$cmat,
 maxIter = 3)
}
