% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMultInt}
\alias{eigenMatMultInt}
\title{Fast matrix multiplication for double x int}
\usage{
eigenMatMultInt(A, B)
}
\arguments{
\item{A}{a double matrix}

\item{B}{an integer matrix}
}
\description{
Fast matrix multiplication for double x int
}
