% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{factorizeMatrix.celda_CG}
\alias{factorizeMatrix.celda_CG}
\title{Generate factorized matrices showing each feature's influence on the celda_CG model clustering}
\usage{
\method{factorizeMatrix}{celda_CG}(counts, celda.mod, type = c("counts",
  "proportion", "posterior"))
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda model. Options are "celda_C" or "celda_CG". Celda object of class "celda_CG".}

\item{type}{Character vector. A vector containing one or more of "counts", "proportion", or "posterior". "counts" returns the raw number of counts for each factorized matrix. "proportions" returns the normalized probabilities for each factorized matrix, which are calculated by dividing the raw counts in each factorized matrix by the total counts in each column. "posterior" returns the posterior estimates. Default `c("counts", "proportion", "posterior")`.}
}
\value{
A list of factorized matrices, of the types requested by the user. NOTE: "population" state matrices are always returned in cell population (rows) x transcriptional states (cols).
}
\description{
Generate factorized matrices showing each feature's influence on the celda_CG model clustering
}
