% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{factorizeMatrix.celda_G}
\alias{factorizeMatrix.celda_G}
\title{Generate factorized matrices showing each feature's influence on the celda_G model clustering}
\usage{
\method{factorizeMatrix}{celda_G}(counts, celda.mod, type = c("counts",
  "proportion", "posterior"))
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_G".}

\item{type}{Character vector. A vector containing one or more of "counts", "proportion", or "posterior". "counts" returns the raw number of counts for each factorized matrix. "proportions" returns the normalized probabilities for each factorized matrix, which are calculated by dividing the raw counts in each factorized matrix by the total counts in each column. "posterior" returns the posterior estimates. Default `c("counts", "proportion", "posterior")`.}
}
\value{
A list of lists of the types of factorized matrices specified
}
\description{
Generate factorized matrices showing each feature's influence on the celda_G model clustering
}
\examples{
celda.sim = simulateCells("celda_G")
celda.mod = celda_G(celda.sim$counts, L=celda.sim$L)
factorized.matrices = factorizeMatrix(celda.sim$counts, celda.mod, "posterior")
}
