% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{factorizeMatrix}
\alias{factorizeMatrix}
\title{Generate factorized matrices showing each feature's influence on cell / gene clustering}
\usage{
factorizeMatrix(counts, celda.mod, type)
}
\arguments{
\item{counts}{A numeric count matrix}

\item{celda.mod}{An object from a celda_list's res.list property}

\item{type}{A character vector containing one or more of "counts", "proportions", or "posterior". "counts" returns the raw number of counts for each entry in each matrix. "proportions" returns the counts matrix where each vector is normalized to a probability distribution. "posterior" returns the posterior estimates which include the addition of the Dirichlet concentration parameter (essentially as a pseudocount).}
}
\description{
Generate factorized matrices showing each feature's influence on cell / gene clustering
}
