% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{featureModuleLookup}
\alias{featureModuleLookup}
\alias{featureModuleLookup,SingleCellExperiment-method}
\title{Obtain the gene module of a gene of interest}
\usage{
featureModuleLookup(sce, ...)

\S4method{featureModuleLookup}{SingleCellExperiment}(sce, feature, exactMatch = TRUE)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object returned by
\link{celda_G}, or \link{celda_CG}, with the matrix
located in the \code{useAssay} assay slot.
Rows represent features and columns represent cells.}

\item{feature}{Character vector. Identify feature modules for the specified
feature names. \code{feature} must match the rownames of \code{sce}.}

\item{exactMatch}{Logical. Whether to look for exactMatch of the gene name
within counts matrix. Default \code{TRUE}.}
}
\value{
List. Each entry corresponds to the feature module determined for
the provided features.
}
\description{
This function will output the corresponding feature module for
 a specified vector of genes from a celda_CG or celda_G celdaModel.
 \code{feature} must match the rownames of \code{sce}.
}
\examples{
data(sceCelda_CG)
module <- featureModuleLookup(sce = sceCelda_CG,
    feature = c("Gene_1", "Gene_XXX"))
}
