% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{featureModuleLookup}
\alias{featureModuleLookup}
\title{Obtain the gene module of a gene of interest}
\usage{
featureModuleLookup(counts, celdaMod, feature, exactMatch = TRUE)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells. This matrix should be the same as the one used to generate
`celdaMod`.}

\item{celdaMod}{Model of class "celda_G" or "celda_CG".}

\item{feature}{Character vector. Identify feature modules for the specified
feature names.}

\item{exactMatch}{Logical. Whether to look for exactMatch of the gene name
within counts matrix. Default TRUE.}
}
\value{
List. Each entry corresponds to the feature module determined for the
 provided features
}
\description{
This function will output the corresponding feature module for a
 specified list of genes from a celdaModel.
}
\examples{
featureModuleLookup(
    counts = celdaCGSim$counts,
    celdaMod = celdaCGMod, "Gene_1")
}
