% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{featureModuleLookup}
\alias{featureModuleLookup}
\title{Obtain the gene module of a gene of interest}
\usage{
featureModuleLookup(counts, celda.mod, feature)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Model of class "celda_G" or "celda_CG".}

\item{feature}{Character vector. Identify feature modules for the specified feature names.}
}
\description{
This function will output the corresponding feature module for a specified list of genes from a celda model.
}
