% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{finalClusterAssignment.celda_CG}
\alias{finalClusterAssignment.celda_CG}
\title{Get the final cell cluster assignments determined by a celda_CG model.}
\usage{
\method{finalClusterAssignment}{celda_CG}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_CG".}
}
\value{
List of integer vectors. The final cell and feature cluster assignments determined by the celda_CG model.
}
\description{
Get the final cell cluster assignments determined by a celda_CG model.
}
\examples{
celda.mod = celda_CG(celda::pbmc_select, K.to.test=10, 
                     L=50, max.iter=2, nchains=1)
final.clustering = finalClusterAssignment(celda.mod)
cell.clusters = final.clustering$z
feature.clusters = final.clustering$y
}
