% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{finalClusterAssignment.celda_G}
\alias{finalClusterAssignment.celda_G}
\title{finalClusterAssignment for celda Gene clustering function}
\usage{
\method{finalClusterAssignment}{celda_G}(celda.mod)
}
\arguments{
\item{celda.mod}{A celda model object of class "celda_G"}
}
\description{
finalClusterAssignment for celda Gene clustering function
}
