% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{finalClusterAssignment.celda_G}
\alias{finalClusterAssignment.celda_G}
\title{Get the final feature cluster assignments determined by a celda_G model.}
\usage{
finalClusterAssignment.celda_G(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_G".}
}
\value{
Integer vector. The final feature cluster assignments determined by the celda_G model.
}
\description{
Get the final feature cluster assignments determined by a celda_G model.
}
\examples{
celda.mod = celda_G(celda::pbmc_select, L=50)
final.clustering = finalClusterAssignment(celda.mod)
}
