% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{finalClusterAssignment}
\alias{finalClusterAssignment}
\title{Get the final gene / cell / gene & cell cluster assignments generated during
a celda run, dependent on the model provided.}
\usage{
finalClusterAssignment(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get the final gene / cell / gene & cell cluster assignments generated during
a celda run, dependent on the model provided.
}
