% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{finalClusterAssignment}
\alias{finalClusterAssignment}
\title{Get the final gene / cell / gene & cell cluster assignments generated during
a celda run, dependent on the model provided.}
\usage{
finalClusterAssignment(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}
}
\description{
Get the final gene / cell / gene & cell cluster assignments generated during
a celda run, dependent on the model provided.
}
