% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{finalLogLikelihood}
\alias{finalLogLikelihood}
\title{Get the final log likelihood for the cell/feature clustering determined by a celda model.}
\usage{
finalLogLikelihood(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}
}
\value{
Double. The final log likelihood associated with the clustering determined by the provided celda model.
}
\description{
Get the final log likelihood for the cell/feature clustering determined by a celda model.
}
\examples{
celda.mod = celda_CG(celda::pbmc_select, K=10, L=50)
final.loglik = finalLogLikelihood(celda.mod)
}
