% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{finalLogLikelihood}
\alias{finalLogLikelihood}
\title{Get the log likelihood from the final iteration of Gibbs sampling
for a given celda model.}
\usage{
finalLogLikelihood(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}
}
\description{
Get the log likelihood from the final iteration of Gibbs sampling
for a given celda model.
}
