% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetEnrich.R
\name{geneSetEnrich}
\alias{geneSetEnrich}
\title{Gene set enrichment}
\usage{
geneSetEnrich(counts, celdaModel, databases, fdr = 0.05)
}
\arguments{
\item{counts}{Integer count matrix. Rows represent genes and columns
represent cells. Row names of the matrix should be gene names.}

\item{celdaModel}{Celda object of class `celda_G` or `celda_CG`.}

\item{databases}{Character vector. Name of reference database. Available
databases can be viewed by running \code{enrichR::listEnrichrDbs()}.}

\item{fdr}{False discovery rate (FDR). Numeric. Cutoff value for adjusted
p-value, terms with FDR below this value are considered significantly
enriched.}
}
\value{
List of length 'L' where each member contains the significantly
 enriched terms for the corresponding module.
}
\description{
Identify and return significantly-enriched terms for each gene
 module in a Celda object. Performs gene set enrichment analysis for Celda
 identified modules using the enrichR package.
}
\examples{
library(M3DExampleData)
counts <- M3DExampleData::Mmus_example_list$data
# subset 100 genes for fast clustering
counts <- counts[1500:2000, ]
# cluster genes into 10 modules for quick demo
cm <- celda_G(counts = as.matrix(counts), L = 10, verbose = FALSE)
gse <- geneSetEnrich(counts,
  cm,
  databases = c("GO_Biological_Process_2018", "GO_Molecular_Function_2018")
)
}
\author{
Ahmed Youssef
}
