% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDecisions.R
\name{getDecisions}
\alias{getDecisions}
\title{Gets cluster estimates using rules generated by
 `celda::buildTreeHybrid`}
\usage{
getDecisions(rules, features)
}
\arguments{
\item{rules}{List object. The `rules` element from  `buildTreeHybrid`
output. Returns NA if cluster estimation was ambiguous.}

\item{features}{A L(features) by N(samples) numeric matrix.}
}
\value{
A character vector of label predicitions.
}
\description{
Get decisions for a matrix of features. Estimate cell
 cluster membership using feature matrix input.
}
\examples{
library(M3DExampleData)
counts <- M3DExampleData::Mmus_example_list$data
# Subset 500 genes for fast clustering
counts <- as.matrix(counts[1501:2000, ])
# Cluster genes ans samples each into 10 modules
cm <- celda_CG(counts = counts, L = 10, K = 5, verbose = FALSE)
# Get features matrix and cluster assignments
factorized <- factorizeMatrix(counts, cm)
features <- factorized$proportions$cell
class <- clusters(cm)$z
# Generate Decision Tree
DecTree <- buildTreeHybrid(features,
    class,
    oneoffMetric = "modified F1",
    threshold = 1,
    consecutiveOneoff = FALSE)

# Get sample estimates in training data
getDecisions(DecTree$rules, features)
}
