% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{getK.celda_C}
\alias{getK.celda_C}
\title{Get the K (number of cell clusters) parameter provided to a celda model.}
\usage{
\method{getK}{celda_C}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C"}
}
\value{
Integer. The K provided to the model during initialization.
}
\description{
Get the K (number of cell clusters) parameter provided to a celda model.
}
\examples{
celda.mod = celda_C(celda::pbmc_select, K=10, max.iter=2, nchains=1)
mod.k.value = getK(celda.mod)
}
