% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{getK.celda_CG}
\alias{getK.celda_CG}
\title{Get the K (number of cell clusters) parameter provided to a celda model.}
\usage{
\method{getK}{celda_CG}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_CG"}
}
\value{
Integer. The K provided to the model during initialization.
}
\description{
Get the K (number of cell clusters) parameter provided to a celda model.
}
\examples{
celda.mod = celda_CG(celda::pbmc_select, K.to.test=10, 
                     L=50, max.iter=2, nchains=1)
mod.k.value = getK(celda.mod)
}
