% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{getK.celda_G}
\alias{getK.celda_G}
\title{Get the K (number of cell clusters) parameter provided to a celda model.}
\usage{
\method{getK}{celda_G}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_G".}
}
\value{
NA. celda_G models do not utilize the K parameter.
}
\description{
Get the K (number of cell clusters) parameter provided to a celda model.
}
