% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{getK}
\alias{getK}
\title{Get the K value used for each chain in a celda run.}
\usage{
getK(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}
}
\description{
Get the K value used for each chain in a celda run.
}
