% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{getL.celda_CG}
\alias{getL.celda_CG}
\title{Get the L (number of feature clusters) parameter provided to a celda model.}
\usage{
\method{getL}{celda_CG}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_CG".}
}
\value{
Integer. The L provided to the model during initialization.
}
\description{
Get the L (number of feature clusters) parameter provided to a celda model.
}
\examples{
celda.mod = celda_CG(celda::pbmc_select, K=10, L=50, max.iter=2, nchains=1)
mod.l.value = getL(celda.mod)
}
