% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{getL.celda_G}
\alias{getL.celda_G}
\title{getL for celda Gene clustering model}
\usage{
\method{getL}{celda_G}(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_G".}
}
\description{
getL for celda Gene clustering model
}
