% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{getL}
\alias{getL}
\title{Get the L value used for each chain in a celda run.}
\usage{
getL(celda.mod)
}
\arguments{
\item{celda.mod}{Celda object of class "celda_C", "celda_G", or "celda_CG".}
}
\description{
Get the L value used for each chain in a celda run.
}
