% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettingClusters.R
\name{gettingClusters}
\alias{gettingClusters}
\title{Assists in selection of K/L parameter for downstream analysis.}
\usage{
gettingClusters(celda.list, counts, iterations)
}
\arguments{
\item{celda.list}{Object of class "celda_list". An object containing celda models returned from `celdaGridSearch()`.}

\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{iterations}{Numeric. Number of iterations to run the function. A higher number will generally output a smoother plot}
}
\description{
Assists in selection of K/L parameter for downstream analysis.
}
