% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.}
\usage{
<<<<<<< HEAD:man/calculateLoglikFromVariables.Rd
calculateLoglikFromVariables(counts, celda.mod, ...)
=======
logLikelihood(counts, model, ...)
>>>>>>> upstream-devel:man/logLikelihood.Rd
}
\arguments{
\item{counts}{The counts matrix used to generate the provided cluster assignments.}

<<<<<<< HEAD:man/calculateLoglikFromVariables.Rd
\item{celda.mod}{Celda model. Options available in `celda::available.models`.}
=======
\item{model}{Celda model. Options available in `celda::available.models`.}
>>>>>>> upstream-devel:man/logLikelihood.Rd

\item{...}{Additional parameters.}
}
\value{
The log-likelihood of the provided cluster assignment for the provided counts matrix.
}
\description{
Calculate a log-likelihood for a user-provided cluster assignment and count matrix, per the desired celda model.
}
\examples{
celda.sim = simulateCells(model="celda_CG")
loglik = logLikelihood(celda.sim$counts, model="celda_CG", 
                       sample.label=celda.sim$sample.label,
                       z=celda.sim$z, y=celda.sim$y,
                       K=celda.sim$K, L=celda.sim$L,
                       alpha=celda.sim$alpha, beta=celda.sim$beta,
                       gamma=celda.sim$gamma, delta=celda.sim$delta)
}
