% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Get final log-likelihood}
\usage{
logLikelihood(celda.mod)

logLikelihood(counts, model, ...)
}
\arguments{
\item{counts}{The counts matrix used to generate the provided cluster assignments.}

\item{model}{Celda model. Options available in `celda::available.models`.}

\item{...}{Additional parameters.}
}
\value{
Numeric. The log-likelihood at the final step of Gibbs sampling used to generate the model.

The log-likelihood of the provided cluster assignment for the provided counts matrix.
}
\description{
Retrieves the final log-likelihood from all iterations of Gibbs sampling used to generate a celda model.
}
\examples{
logLikelihood(celda.CG.mod)
loglik = logLikelihood(celda.CG.sim$counts, model="celda_CG", 
                       sample.label=celda.CG.sim$sample.label,
                       z=celda.CG.sim$z, y=celda.CG.sim$y,
                       K=celda.CG.sim$K, L=celda.CG.sim$L,
                       alpha=celda.CG.sim$alpha, beta=celda.CG.sim$beta,
                       gamma=celda.CG.sim$gamma, delta=celda.CG.sim$delta)
}
