% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{logLikelihood}
\alias{logLikelihood}
\alias{logLikelihood,SingleCellExperiment-method}
\title{Calculate the Log-likelihood of a celda model}
\usage{
logLikelihood(sce, ...)

\S4method{logLikelihood}{SingleCellExperiment}(sce, useAssay = "counts")
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object returned by
\link{celda_C}, \link{celda_G}, or \link{celda_CG}, with the matrix
located in the \code{useAssay} assay slot.
Rows represent features and columns represent cells.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use. Default "counts".}
}
\value{
The log-likelihood of the cluster assignment for the
 provided \linkS4class{SingleCellExperiment}.
}
\description{
Calculate the log-likelihood for user-provided cell population
 and feature module cluster
 assignments on the count matrix, per the desired celda model.
}
\examples{
data(sceCelda_C, sceCelda_CG)
loglikC <- logLikelihood(sceCelda_C)
loglikCG <- logLikelihood(sceCelda_CG)
}
\seealso{
`celda_C()` for clustering cells
}
