% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Calculate LogLikelihood}
\usage{
logLikelihood(counts, model, ...)
}
\arguments{
\item{counts}{The counts matrix used to generate the provided cluster
assignments.}

\item{model}{celdaModel. Options available in `celda::availableModels`.}

\item{...}{Additional parameters.}
}
\value{
The log-likelihood of the provided cluster assignment for the
 provided counts matrix.
}
\description{
Calculate a log-likelihood for a user-provided cluster
 assignment and count matrix, per the desired celdaModel.
}
\examples{
data(celdaCGSim)
loglik <- logLikelihood(celdaCGSim$counts,
    model = "celda_CG",
    sampleLabel = celdaCGSim$sampleLabel,
    z = celdaCGSim$z, y = celdaCGSim$y,
    K = celdaCGSim$K, L = celdaCGSim$L,
    alpha = celdaCGSim$alpha, beta = celdaCGSim$beta,
    gamma = celdaCGSim$gamma, delta = celdaCGSim$delta
)
}
