% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{logLikelihoodHistory,celdaModel-method}
\alias{logLikelihoodHistory,celdaModel-method}
\title{Get log-likelihood history}
\usage{
\S4method{logLikelihoodHistory}{celdaModel}(celdaMod)
}
\arguments{
\item{celdaMod}{celdaModel. Options available in `celda::availableModels`.}
}
\value{
Numeric. The log-likelihood at each step of Gibbs sampling used to
 generate the model.
}
\description{
Retrieves the complete log-likelihood from all iterations of
 Gibbs sampling used to generate a celdaModel.
}
\examples{
logLikelihoodHistory(celdaCGMod)
}
