% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{matrixNames,celdaModel-method}
\alias{matrixNames,celdaModel-method}
\title{Get feature, cell and sample names from a celdaModel}
\usage{
\S4method{matrixNames}{celdaModel}(celdaMod)
}
\arguments{
\item{celdaMod}{celdaModel. Options available in `celda::availableModels`.}
}
\value{
List. Contains row, column, and sample character vectors
 corresponding to the values provided when the celdaModel was generated.
}
\description{
Retrieves the row, column, and sample names used to generate a
 celdaModel.
}
\examples{
matrixNames(celdaCGMod)
}
