% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleHeatmap.R
\name{moduleHeatmap}
\alias{moduleHeatmap}
\alias{moduleHeatmap,SingleCellExperiment-method}
\alias{moduleHeatmap,matrix-method}
\title{Heatmap for featureModules}
\usage{
moduleHeatmap(x, ...)

\S4method{moduleHeatmap}{SingleCellExperiment}(
  x,
  useAssay = "counts",
  altExpName = "featureSubset",
  featureModule = 1,
  topCells = 100,
  topFeatures = NULL,
  normalizedCounts = NA,
  scaleRow = scale,
  showFeaturenames = TRUE
)

\S4method{moduleHeatmap}{matrix}(
  x,
  celdaMod,
  featureModule = 1,
  topCells = 100,
  topFeatures = NULL,
  normalizedCounts = NA,
  scaleRow = scale,
  showFeaturenames = TRUE
)
}
\arguments{
\item{x}{A numeric \link{matrix} of counts or a
\linkS4class{SingleCellExperiment}
with the matrix located in the assay slot under \code{useAssay}.
Rows represent features and columns represent cells.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use if \code{x} is a
\linkS4class{SingleCellExperiment} object. Default "counts".}

\item{altExpName}{The name for the \link[SingleCellExperiment]{altExp} slot
to use. Default "featureSubset".}

\item{featureModule}{Integer Vector. The featureModule(s) to display.
Multiple modules can be included in a vector.}

\item{topCells}{Integer. Number of cells with the highest and lowest
probabilities for this module to include in the heatmap. For example, if
`topCells` = 50, the 50 cells with the lowest probability and the 50 cells
with the highest probability for that featureModule will be included. If
NULL, all cells will be plotted. Default 100.}

\item{topFeatures}{Integer. Plot `topFeatures` with the highest probability
in the featureModule. If NULL, plot all features in the module. Default
NULL.}

\item{normalizedCounts}{Integer matrix. Rows represent features and columns
represent cells. This matrix should correspond to the one provided for
`counts`, but should be passed through. If NA, normalize `counts`.
Default NA.
`normalizeCounts(counts, "proportion", transformationFun=sqrt)`. Use of this
parameter is particularly useful for plotting many moduleHeatmaps, where
normalizing the counts matrix repeatedly would be too time consuming.}

\item{scaleRow}{Character. Which function to use to scale each individual
row. Set to NULL to disable. Occurs after normalization and log
transformation. For example, `scale` will Z-score transform each row.
Default `scale`.}

\item{showFeaturenames}{Logical. Wheter feature names should be displayed.
Default TRUE.}

\item{celdaMod}{Celda object of class \link{celda_G} or \link{celda_CG}. Used
only if \code{x} is a matrix object.}
}
\value{
A list containing row and column dendrograms as well as a gtable for
 grob plotting
}
\description{
Renders a heatmap for selected featureModules. Cells are
 ordered from those with the lowest probability of the module on the left to
 the highest probability on the right. If more than one module is used, then
 cells will be ordered by the probabilities of the first module only.
 Features are ordered from those with the highest probability in the module
 on the top to the lowest probability on the bottom.
}
\examples{
data(sceCeldaCG)
moduleHeatmap(sceCeldaCG)
data(celdaCGSim, celdaCGMod)
moduleHeatmap(celdaCGSim$counts, celdaCGMod)
}
