% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{modules}
\alias{modules}
\alias{modules,SingleCellExperiment-method}
\alias{modules<-}
\alias{modules<-,SingleCellExperiment-method}
\title{Get or set the feature module labels from a celda
 \linkS4class{SingleCellExperiment}  object.}
\usage{
modules(sce)

\S4method{modules}{SingleCellExperiment}(sce)

modules(sce) <- value

\S4method{modules}{SingleCellExperiment}(sce) <- value
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object returned by
\link{celda_G}, or \link{celda_CG}, with the matrix
located in the \code{useAssay} assay slot.
Rows represent features and columns represent cells.}
}
\value{
Character vector. Contains feature module labels for each
 feature in x.
}
\description{
Return or set the feature module cluster labels determined
 by \link{celda_G} or \link{celda_CG} models.
}
\examples{
data(celdaCGMod)
modules(celdaCGMod)
}
