% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{normalizeCounts function}
\usage{
normalizeCounts(counts, scale.factor = 1e+06)
}
\arguments{
\item{counts}{A count matrix}

\item{scale.factor}{the scalor for the normalization}
}
\description{
normalizeCounts function
}
