% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Normalize a counts matrix by a scalar factor}
\usage{
normalizeCounts(counts, scale.factor = 1e+06)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{scale.factor}{the scalar for the normalization}
}
\description{
Normalize a counts matrix by a scalar factor
}
