% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\docType{methods}
\name{params,celdaModel-method}
\alias{params,celdaModel-method}
\title{Get parameter values provided for celdaModel creation}
\usage{
\S4method{params}{celdaModel}(celdaMod)
}
\arguments{
\item{celdaMod}{celdaModel. Options available in `celda::availableModels`.}
}
\value{
List. Contains the model-specific parameters for the provided celda
 model object depending on its class.
}
\description{
Retrieves the K/L, model priors (e.g. alpha, beta),
 and count matrix checksum parameters provided during the creation of the
 provided celdaModel.
}
\examples{
params(celdaCGMod)
}
