% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{params}
\alias{params}
\title{Get parameter values provided for celda model creation}
\usage{
params(celda.mod)
}
\value{
List. Contains the model-specific parameters for the provided celda model object depending on its class.
}
\description{
Retrieves the K/L, model priors (e.g. alpha, beta), random seed, and count matrix checksum parameters provided during the creation of the provided celda model.
}
\examples{
params(celda.CG.mod)
}
