% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_res}
\alias{pbmc_res}
\title{pbmc_res}
\format{An object of class \code{celda_list} of length 4.}
\source{
\url{https://support.10xgenomics.com/single-cell/software/pipelines/latest/rkit}
}
\usage{
pbmc_res
}
\description{
A celda model made from pbmc_data, run with parameters K = 10:30 and L = 20,30.
}
\details{
Celda model created from 3000 Peripheral Blood Mononuclear Cells (PBMC) from 10X Genomics.
}
\keyword{datasets}
