% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{perplexity}
\alias{perplexity}
\alias{perplexity,SingleCellExperiment-method}
\title{Calculate the perplexity of a celda model}
\usage{
perplexity(sce, ...)

\S4method{perplexity}{SingleCellExperiment}(sce, useAssay = "counts")
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object returned by
\link{celda_C}, \link{celda_G}, or \link{celda_CG}, with the matrix
located in the \code{useAssay} assay slot.
Rows represent features and columns represent cells.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use. Default "counts".}
}
\value{
Numeric. The perplexity for the provided
 \linkS4class{SingleCellExperiment}.
}
\description{
Perplexity is a statistical measure of how well a probability
 model can predict new data. Lower perplexity indicates a better model.
}
\examples{
data(sceCelda_CG)
clusterProb <- clusterProbability(sceCelda_CG)
}
