% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{perplexity}
\alias{perplexity}
\title{Calculate the perplexity from a single celda model}
\usage{
perplexity(counts, celda.mod, new.counts = NULL)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.}

\item{celda.mod}{Celda object of class "celda_C", "celda_G" or "celda_CG".}

\item{new.counts}{A new counts matrix used to calculate perplexity. If NULL, perplexity will be calculated for the 'counts' matrix. Default NULL.}
}
\value{
Numeric. The perplexity for the provided count data and model.
}
\description{
Perplexity can be seen as a measure of how well a provided set of 
cluster assignments fit the data being clustered.
}
\examples{
perplexity = perplexity(celda.CG.sim$counts, celda.CG.mod)
}
