% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotCeldaViolin}
\alias{plotCeldaViolin}
\title{Feature Expression Violin Plot}
\usage{
plotCeldaViolin(
  counts,
  celdaMod,
  features,
  exactMatch = TRUE,
  plotDots = TRUE,
  dotSize = 0.1
)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells.}

\item{celdaMod}{Celda object of class "celda_G" or "celda_CG".}

\item{features}{Character vector. Uses these genes for plotting.}

\item{exactMatch}{Logical. Whether an exact match or a partial match using
\code{grep()} is used to look up the feature in the rownames of the counts
 matrix. Default \code{TRUE}.}

\item{plotDots}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the
expression of features will be plotted as points in addition to the violin
curve. Default TRUE.}
}
\value{
Violin plot for each feature, grouped by celda cluster
}
\description{
Outputs a violin plot for feature expression data.
}
\examples{
data(celdaCGSim, celdaCGMod)
plotCeldaViolin(counts = celdaCGSim$counts,
    celdaMod = celdaCGMod, features = "Gene_1")
}
