% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decontx.R
\name{plotDecontXMarkerPercentage}
\alias{plotDecontXMarkerPercentage}
\title{Plots percentage of cells cell types expressing markers}
\usage{
plotDecontXMarkerPercentage(x, markers, ...)
}
\arguments{
\item{x}{Either a \link[SingleCellExperiment:SingleCellExperiment-class]{SingleCellExperiment} or
a matrix-like object of counts.}

\item{markers}{List. A named list indicating the marker genes
for each cell type of interest.}

\item{...}{Additional arguments passed to
\code{\link[decontX]{plotDecontXMarkerPercentage}} (for example
\code{groupClusters}, \code{assayName}, \code{z}, \code{threshold},
\code{exactMatch}, \code{by}, \code{ncol}, \code{labelBars},
\code{labelSize}).}
}
\value{
Returns a \code{ggplot} object.
}
\description{
Generates a barplot that shows the percentage of
cells within clusters or cell types that have detectable levels
of given marker genes. Can be used to view the expression of
marker genes in different cell types before and after
decontamination with \code{\link{decontX}}. As of celda 1.23.0 this is a
thin wrapper around \code{\link[decontX]{plotDecontXMarkerPercentage}} in the
\pkg{decontX} package, which must be installed.
}
\seealso{
\code{\link[decontX]{plotDecontXMarkerPercentage}}. See
\code{\link{decontX}} for a full example of how to estimate and plot
contamination.
}
\author{
Shiyi Yang, Joshua Campbell
}
