% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceCluster}
\alias{plotDimReduceCluster}
\title{Create a scatterplot based on celda cluster labels.}
\usage{
plotDimReduceCluster(dim1, dim2, cluster, size = 1, xlab = "Dimension_1",
  ylab = "Dimension_2", specific_clusters = NULL)
}
\arguments{
\item{dim1}{Numeric vector. First dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector. Second dimension from data dimensionality reduction output.}

\item{cluster}{Integer vector. Contains cluster labels for each cell.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default "Dimension_1".}

\item{ylab}{Character vector. Label for the y-axis. Default "Dimension_2".}

\item{specific_clusters}{Numeric vector. Only color cells in the specified clusters. All other cells will be grey. If NULL, all clusters will be colored. Default NULL.}
}
\value{
The plot as a ggplot object
}
\description{
Create a scatterplot based on celda cluster labels.
}
\examples{
\donttest{
sim.res = simulateCells(model="celda_CG", K = 5, L = 5)
celda_cg <- celda_CG(counts = sim.res$counts, K = 5, L = 5)
celda.tsne <- celdaTsne(counts = sim.res$counts, celda.mod = celda_cg)
plotDimReduceCluster(dim1 = celda.tsne[,1], dim2 = celda.tsne[,2],
                     cluster = as.factor(celda_cg$z),
                     specific_clusters = c(1,2,3))
}
}
