% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceCluster}
\alias{plotDimReduceCluster}
\alias{plotDimReduceCluster,SingleCellExperiment-method}
\alias{plotDimReduceCluster,vector-method}
\title{Plotting the cell labels on a dimensionality reduction plot}
\usage{
plotDimReduceCluster(x, ...)

\S4method{plotDimReduceCluster}{SingleCellExperiment}(
  dim1,
  dim2,
  x,
  size = 1,
  xlab = "Dimension_1",
  ylab = "Dimension_2",
  specificClusters = NULL,
  labelClusters = FALSE,
  groupBy = NULL,
  labelSize = 3.5
)

\S4method{plotDimReduceCluster}{vector}(
  dim1,
  dim2,
  x,
  size = 1,
  xlab = "Dimension_1",
  ylab = "Dimension_2",
  specificClusters = NULL,
  labelClusters = FALSE,
  groupBy = NULL,
  labelSize = 3.5
)
}
\arguments{
\item{x}{Integer vector of cell cluster labels or a
\linkS4class{SingleCellExperiment} object
containing cluster labels for each cell in \code{"celda_cell_cluster"}
column in \code{colData(x)}.}

\item{dim1}{Numeric vector. First dimension from data
dimensionality reduction output.}

\item{dim2}{Numeric vector. Second dimension from data
dimensionality reduction output.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default "Dimension_1".}

\item{ylab}{Character vector. Label for the y-axis. Default "Dimension_2".}

\item{specificClusters}{Numeric vector.
Only color cells in the specified clusters.
All other cells will be grey.
If NULL, all clusters will be colored. Default NULL.}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.
Default FALSE.}

\item{groupBy}{Character vector. Contains sample labels for each cell.
If NULL, all samples will be plotted together. Default NULL.}

\item{labelSize}{Numeric. Sets size of label if labelClusters is TRUE.
Default 3.5.}

\item{cluster}{Integer vector. Contains cluster labels for each cell.}
}
\value{
The plot as a ggplot object
}
\description{
Create a scatterplot for each row of a normalized
 gene expression matrix where x and y axis are from a
 data dimensionality reduction tool.
 The cells are colored by its given `cluster` label.
}
\examples{
\donttest{
data(sceCeldaCG)
celdaTsne <- celdaTsne(sceCeldaCG)
plotDimReduceCluster(
  dim1 = celdaTsne[, 1],
  dim2 = celdaTsne[, 2],
  x = sceCeldaCG,
  specificClusters = c(1, 2, 3))
}
\donttest{
data(sceCeldaCG, celdaCGMod)
celdaTsne <- celdaTsne(sceCeldaCG)
plotDimReduceCluster(
  dim1 = celdaTsne[, 1],
  dim2 = celdaTsne[, 2],
  x = celdaClusters(celdaCGMod),
  specificClusters = c(1, 2, 3))
}
}
