% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceCluster}
\alias{plotDimReduceCluster}
\title{Plotting the cell labels on a dimensionality reduction plot}
\usage{
plotDimReduceCluster(dim1, dim2, cluster, size = 1,
  xlab = "Dimension_1", ylab = "Dimension_2",
  specificClusters = NULL, labelClusters = FALSE, labelSize = 3.5)
}
\arguments{
\item{dim1}{Numeric vector. First dimension from data
dimensionality reduction output.}

\item{dim2}{Numeric vector. Second dimension from data
dimensionality reduction output.}

\item{cluster}{Integer vector. Contains cluster labels for each cell.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default "Dimension_1".}

\item{ylab}{Character vector. Label for the y-axis. Default "Dimension_2".}

\item{specificClusters}{Numeric vector.
Only color cells in the specified clusters.
All other cells will be grey.
If NULL, all clusters will be colored. Default NULL.}

\item{labelClusters}{Logical. Whether the cluster labels are plotted.
Default FALSE.}

\item{labelSize}{Numeric. Sets size of label if labelClusters is TRUE.
Default 3.5.}
}
\value{
The plot as a ggplot object
}
\description{
Create a scatterplot for each row of a normalized
 gene expression matrix where x and y axis are from a
 data dimensionality reduction tool.
 The cells are colored by its given `cluster` label.
}
\examples{
\donttest{
data(celdaCGSim, celdaCGMod)
celdaTsne <- celdaTsne(counts = celdaCGSim$counts,
    celdaMod = celdaCGMod)
plotDimReduceCluster(dim1 = celdaTsne[, 1],
    dim2 = celdaTsne[, 2],
    cluster = as.factor(z(celdaCGMod)),
    specificClusters = c(1, 2, 3))
}
}
