% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceCluster}
\alias{plotDimReduceCluster}
\title{Plotting the cell labels on a dimensionality reduction plot}
\usage{
plotDimReduceCluster(dim1, dim2, cluster, size = 1, xlab = "Dimension_1",
  ylab = "Dimension_2", specific_clusters = NULL)
}
\arguments{
\item{dim1}{Numeric vector. First dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector. Second dimension from data dimensionality reduction output.}

\item{cluster}{Integer vector. Contains cluster labels for each cell.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default "Dimension_1".}

\item{ylab}{Character vector. Label for the y-axis. Default "Dimension_2".}

\item{specific_clusters}{Numeric vector. Only color cells in the specified clusters. All other cells will be grey. If NULL, all clusters will be colored. Default NULL.}
}
\value{
The plot as a ggplot object
}
\description{
Create a scatterplot for each row of a normalized gene expression matrix where x and y axis are from a data dimensionality reduction tool. The cells are colored by its given `cluster` label.
}
\examples{
\donttest{
celda.tsne <- celdaTsne(counts = celda.CG.sim$counts, celda.mod = celda.CG.mod)
plotDimReduceCluster(dim1 = celda.tsne[,1], dim2 = celda.tsne[,2],
                     cluster = as.factor(z(celda.CG.mod)),
                     specific_clusters = c(1,2,3))
}
}
