% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceFeature}
\alias{plotDimReduceFeature}
\title{Plotting feature expression on a dimensionality reduction plot}
\usage{
plotDimReduceFeature(
  dim1,
  dim2,
  counts,
  features,
  headers = NULL,
  normalize = FALSE,
  zscore = TRUE,
  exactMatch = TRUE,
  trim = c(-1, 1),
  size = 1,
  xlab = "Dimension_1",
  ylab = "Dimension_2",
  colorLow = "blue4",
  colorMid = "white",
  colorHigh = "firebrick1",
  midpoint = NULL,
  ncol = NULL
)
}
\arguments{
\item{dim1}{Numeric vector. First dimension from data
dimensionality reduction output.}

\item{dim2}{Numeric vector. Second dimension from data dimensionality
reduction output.}

\item{counts}{Integer matrix. Rows represent features and columns
represent cells.}

\item{features}{Character vector. Features in the rownames of counts to plot.}

\item{headers}{Character vector. If `NULL`, the corresponding rownames are
used as labels. Otherwise, these headers are used to label the features.}

\item{normalize}{Logical. Whether to normalize the columns of `counts`.
Default \code{FALSE}.}

\item{zscore}{Logical. Whether to scale each feature to have a mean 0
and standard deviation of 1. Default \code{TRUE}.}

\item{exactMatch}{Logical. Whether an exact match or a partial match using
\code{grep()} is used to look up the feature in the rownames of the counts
 matrix. Default TRUE.}

\item{trim}{Numeric vector. Vector of length two that specifies the lower
and upper bounds for the data. This threshold is applied after row scaling.
Set to NULL to disable. Default \code{c(-1,1)}.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default "Dimension_1".}

\item{ylab}{Character vector. Label for the y-axis. Default "Dimension_2".}

\item{colorLow}{Character. A color available from `colors()`. The color
will be used to signify the lowest values on the scale. Default 'blue'.}

\item{colorMid}{Character. A color available from `colors()`. The color
will be used to signify the midpoint on the scale. Default 'white'.}

\item{colorHigh}{Character. A color available from `colors()`. The color
will be used to signify the highest values on the scale. Default 'red'.}

\item{midpoint}{Numeric. The value indicating the midpoint of the
diverging color scheme. If \code{NULL}, defaults to the mean
with 10 percent of values trimmed. Default \code{NULL}.}

\item{ncol}{Integer. Passed to \link[ggplot2]{facet_wrap}. Specify the
number of columns for facet wrap.}
}
\value{
The plot as a ggplot object
}
\description{
Create a scatterplot for each row of a normalized gene
 expression matrix where x and y axis are from a data dimensionality
 reduction tool. The cells are colored by expression of
 the specified feature.
}
\examples{
\donttest{
data(celdaCGSim, celdaCGMod)
celdaTsne <- celdaTsne(counts = celdaCGSim$counts,
    celdaMod = celdaCGMod)
plotDimReduceFeature(dim1 = celdaTsne[, 1],
    dim2 = celdaTsne[, 2],
    counts = celdaCGSim$counts,
    normalize = TRUE,
    features = c("Gene_99"),
    exactMatch = TRUE)
}
}
