% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceGrid}
\alias{plotDimReduceGrid}
\title{Mapping the dimensionality reduction plot}
\usage{
plotDimReduceGrid(dim1, dim2, matrix, size, xlab, ylab, colorLow, colorMid,
  colorHigh, varLabel)
}
\arguments{
\item{dim1}{Numeric vector. First dimension from data dimensionality
reduction output.}

\item{dim2}{Numeric vector. Second dimension from data dimensionality
reduction output.}

\item{matrix}{Numeric matrix. Each row of the matrix will be plotted as
a separate facet.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default 'Dimension_1'.}

\item{ylab}{Character vector. Label for the y-axis. Default 'Dimension_2'.}

\item{colorLow}{Character. A color available from `colors()`.
The color will be used to signify the lowest values on the scale.
Default 'grey'.}

\item{colorMid}{Character. A color available from `colors()`.
The color will be used to signify the midpoint on the scale.}

\item{colorHigh}{Character. A color available from `colors()`.
The color will be used to signify the highest values on the scale.
 Default 'blue'.}

\item{varLabel}{Character vector. Title for the color legend.}
}
\value{
The plot as a ggplot object
}
\description{
Creates a scatterplot given two dimensions from a data
 dimensionality reduction tool (e.g tSNE) output.
}
\examples{
\donttest{
data(celdaCGSim, celdaCGMod)
celdaTsne <- celdaTsne(counts = celdaCGSim$counts,
    celdaMod = celdaCGMod)
plotDimReduceGrid(celdaTsne[, 1],
    celdaTsne[, 2],
    matrix = celdaCGSim$counts,
    xlab = "Dimension1",
    ylab = "Dimension 2",
    varLabel = "tsne",
    size = 1,
    colorLow = "grey",
    colorMid = NULL,
    colorHigh = "blue")
}
}
