% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceGrid}
\alias{plotDimReduceGrid}
\alias{plotDimReduceGrid,SingleCellExperiment-method}
\alias{plotDimReduceGrid,matrix-method}
\title{Mapping the dimensionality reduction plot}
\usage{
plotDimReduceGrid(x, ...)

\S4method{plotDimReduceGrid}{SingleCellExperiment}(
  x,
  reducedDimName,
  dim1 = NULL,
  dim2 = NULL,
  useAssay = "counts",
  size = 1,
  xlab = "Dimension_1",
  ylab = "Dimension_2",
  colorLow = "blue4",
  colorMid = "white",
  colorHigh = "firebrick1",
  midpoint = NULL,
  varLabel = NULL,
  ncol = NULL,
  headers = NULL
)

\S4method{plotDimReduceGrid}{matrix}(
  x,
  dim1,
  dim2,
  size = 1,
  xlab = "Dimension_1",
  ylab = "Dimension_2",
  colorLow = "blue4",
  colorMid = "white",
  colorHigh = "firebrick1",
  midpoint = NULL,
  varLabel = NULL,
  ncol = NULL,
  headers = NULL
)
}
\arguments{
\item{x}{Numeric matrix or a \linkS4class{SingleCellExperiment} object
with the matrix located in the assay slot under \code{useAssay}. Each
row of the matrix will be plotted as a separate facet.}

\item{reducedDimName}{The name of the dimension reduction slot in
\code{reducedDimNames(x)} if \code{x} is a
\linkS4class{SingleCellExperiment} object. Ignored if both \code{dim1} and
\code{dim2} are set.}

\item{dim1}{Numeric vector. First dimension from data dimensionality
reduction output.}

\item{dim2}{Numeric vector. Second dimension from data dimensionality
reduction output.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use if \code{x} is a
\linkS4class{SingleCellExperiment} object. Default "counts".}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default 'Dimension_1'.}

\item{ylab}{Character vector. Label for the y-axis. Default 'Dimension_2'.}

\item{colorLow}{Character. A color available from `colors()`.
The color will be used to signify the lowest values on the scale.
Default "blue4".}

\item{colorMid}{Character. A color available from `colors()`.
The color will be used to signify the midpoint on the scale. Default
"white".}

\item{colorHigh}{Character. A color available from `colors()`.
The color will be used to signify the highest values on the scale.
Default "firebrick1".}

\item{midpoint}{Numeric. The value indicating the midpoint of the
diverging color scheme. If \code{NULL}, defaults to the mean
with 10 percent of values trimmed. Default \code{NULL}.}

\item{varLabel}{Character vector. Title for the color legend.}

\item{ncol}{Integer. Passed to \link[ggplot2]{facet_wrap}. Specify the
number of columns for facet wrap.}

\item{headers}{Character vector. If `NULL`, the corresponding rownames are
used as labels. Otherwise, these headers are used to label the genes.}
}
\value{
The plot as a ggplot object
}
\description{
Creates a scatterplot given two dimensions from a data
 dimensionality reduction tool (e.g tSNE) output.
}
\examples{
data(sceCeldaCG)
sce <- celdaTsne(sceCeldaCG)
plotDimReduceGrid(x = sce,
  reducedDimName = "celda_tSNE",
  xlab = "Dimension1",
  ylab = "Dimension2",
  varLabel = "tSNE")
data(sceCeldaCG)
sce <- celdaTsne(sceCeldaCG)
plotDimReduceGrid(x = counts(sce),
  dim1 = reducedDim(sce, "celda_tSNE")[, 1],
  dim2 = reducedDim(sce, "celda_tSNE")[, 2],
  xlab = "Dimension1",
  ylab = "Dimension2",
  varLabel = "tSNE")
}
