% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr.R
\name{plotDimReduceModule}
\alias{plotDimReduceModule}
\alias{plotDimReduceModule,SingleCellExperiment-method}
\alias{plotDimReduceModule,matrix-method}
\title{Plotting the Celda module probability on a
 dimensionality reduction plot}
\usage{
plotDimReduceModule(x, ...)

\S4method{plotDimReduceModule}{SingleCellExperiment}(
  x,
  reducedDimName,
  dim1 = NULL,
  dim2 = NULL,
  useAssay = "counts",
  modules = NULL,
  rescale = TRUE,
  size = 1,
  xlab = "Dimension_1",
  ylab = "Dimension_2",
  colorLow = "blue4",
  colorMid = "white",
  colorHigh = "firebrick1",
  ncol = NULL
)

\S4method{plotDimReduceModule}{matrix}(
  x,
  dim1,
  dim2,
  celdaMod,
  modules = NULL,
  rescale = TRUE,
  size = 1,
  xlab = "Dimension_1",
  ylab = "Dimension_2",
  colorLow = "blue4",
  colorMid = "white",
  colorHigh = "firebrick1",
  ncol = NULL
)
}
\arguments{
\item{x}{Numeric matrix or a \linkS4class{SingleCellExperiment} object
with the matrix located in the assay slot under \code{useAssay}. Rows
represent features and columns represent cells.}

\item{reducedDimName}{The name of the dimension reduction slot in
\code{reducedDimNames(x)} if \code{x} is a
\linkS4class{SingleCellExperiment} object. Ignored if both \code{dim1} and
\code{dim2} are set.}

\item{dim1}{Numeric vector.
First dimension from data dimensionality reduction output.}

\item{dim2}{Numeric vector.
Second dimension from data dimensionality reduction output.}

\item{useAssay}{A string specifying which \link[SummarizedExperiment]{assay}
slot to use if \code{x} is a
\linkS4class{SingleCellExperiment} object. Default "counts".}

\item{modules}{Character vector. Module(s) from celda model to be plotted.
e.g. c("1", "2").}

\item{rescale}{Logical.
Whether rows of the matrix should be rescaled to [0, 1]. Default TRUE.}

\item{size}{Numeric. Sets size of point on plot. Default 1.}

\item{xlab}{Character vector. Label for the x-axis. Default "Dimension_1".}

\item{ylab}{Character vector. Label for the y-axis. Default "Dimension_2".}

\item{colorLow}{Character. A color available from `colors()`.
The color will be used to signify the lowest values on the scale.
Default "blue4".}

\item{colorMid}{Character. A color available from `colors()`.
The color will be used to signify the midpoint on the scale. Default
"white".}

\item{colorHigh}{Character. A color available from `colors()`.
The color will be used to signify the highest values on the scale.
Default "firebrick1".}

\item{ncol}{Integer. Passed to \link[ggplot2]{facet_wrap}. Specify the
number of columns for facet wrap.}

\item{celdaMod}{Celda object of class "celda_G" or "celda_CG". Used only if
\code{x} is a matrix object.}
}
\value{
The plot as a ggplot object
}
\description{
Create a scatterplot for each row of a normalized
 gene expression matrix where x and y axis are from a data
 dimensionality reduction tool.
 The cells are colored by the module probability(s).
}
\examples{
data(sceCeldaCG)
sce <- celdaTsne(sceCeldaCG)
plotDimReduceModule(x = sce,
  reducedDimName = "celda_tSNE",
  modules = c("1", "2"))
library(SingleCellExperiment)
data(sceCeldaCG, celdaCGMod)
sce <- celdaTsne(sceCeldaCG)
plotDimReduceModule(x = counts(sce),
  dim1 = reducedDim(sce, "celda_tSNE")[, 1],
  dim2 = reducedDim(sce, "celda_tSNE")[, 2],
  celdaMod = celdaCGMod,
  modules = c("1", "2"))
}
