% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity.celda_C}
\alias{plotGridSearchPerplexity.celda_C}
\title{Plot perplexity as a function of K from celda_C models}
\usage{
\method{plotGridSearchPerplexity}{celda_C}(celda.list)
}
\arguments{
\item{celda.list}{Object of class 'celda_list'.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters.
}
\description{
Plots perplexity as a function of the cell (K) clusters as generated by celdaGridSearch().
}
\examples{
## Simulate a small dataset with 5 cell clusters and 10 feature modules
celda.sim = simulateCells(model="celda_C", K=5, G=100)
## Run various combinations of parameters with 'celdaGridSearch'
cgs = celdaGridSearch(celda.sim$counts, model="celda_C", params.test=list(K=3:5))
cgs = resamplePerplexity(celda.sim$counts, cgs)
plotGridSearchPerplexity(cgs)
}
