% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity.celda_C}
\alias{plotGridSearchPerplexity.celda_C}
\title{Plot perplexity as a function of K from celda_C models}
\usage{
plotGridSearchPerplexity.celda_C(celdaList, sep)
}
\arguments{
\item{celdaList}{Object of class 'celdaList'.}

\item{sep}{Numeric. Breaks in the x axis of the resulting plot.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering
  parameters.
}
\description{
Plots perplexity as a function of the cell (K) clusters as
  generated by celdaGridSearch().
}
\examples{
data(celdaCGSim, celdaCGGridSearchRes)
celdaCGGridSearchRes <- resamplePerplexity(
  celdaCGSim$counts,
  celdaCGGridSearchRes
)
plotGridSearchPerplexity(celdaCGGridSearchRes)
}
