% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity.celda_C}
\alias{plotGridSearchPerplexity.celda_C}
\title{Plot perplexity as a function of K from celda_C models}
\usage{
\method{plotGridSearchPerplexity}{celda_C}(celda.list)
}
\arguments{
\item{celda.list}{An object returned from 'celdaGridSearch'}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters
}
\description{
Plots perplexity as a function of the cell (K) clusters as generated by celdaGridSearch().
}
