% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity.celda_CG}
\alias{plotGridSearchPerplexity.celda_CG}
\title{Plot perplexity as a function of K and L from celda_CG models}
\usage{
plotGridSearchPerplexity.celda_CG(celdaList, sep)
}
\arguments{
\item{celdaList}{Object of class 'celdaList'.}

\item{sep}{Numeric. Breaks in the x axis of the resulting plot.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering
 parameters.
}
\description{
This function plots perplexity as a function of the cell/gene
 (K/L) clusters as generated by celdaGridSearch().
}
\examples{
celdaCGGridSearchRes <- resamplePerplexity(
  celdaCGSim$counts,
  celdaCGGridSearchRes
)
plotGridSearchPerplexity(celdaCGGridSearchRes)
}
