% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity.celda_G}
\alias{plotGridSearchPerplexity.celda_G}
\title{Plot perplexity as a function of L from a celda_G model}
\usage{
\method{plotGridSearchPerplexity}{celda_G}(celda.list)
}
\arguments{
\item{celda.list}{An object returned from 'celdaGridSearch'}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters
}
\description{
Plots perplexity as a function of the gene (L) clusters as generated by celdaGridSearch().
}
