% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity}
\alias{plotGridSearchPerplexity}
\title{Visualize perplexity of every model in a celda_list, by unique K/L combinations}
\usage{
plotGridSearchPerplexity(celda.list)
}
\arguments{
\item{celda.list}{Object of class 'celda_list'.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters.
}
\description{
Visualize perplexity of every model in a celda_list, by unique K/L combinations
}
\examples{
## Simulate a small dataset with 5 cell clusters and 10 feature modules
celda.sim = simulateCells(model="celda_CG", K=5, L=10, G=100)

## Run various combinations of parameters with 'celdaGridSearch'
cgs = celdaGridSearch(celda.sim$counts, model="celda_CG", params.test=list(K=4:6, L=9:11), 
                      params.fixed=list(sample.label=celda.sim$sample.label),
                      best.only=TRUE, nchains=1)
cgs = resamplePerplexity(celda.sim$counts, cgs)
plotGridSearchPerplexity(cgs)
}
