% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity}
\alias{plotGridSearchPerplexity}
\title{Visualize perplexity of every model in a celda_list, by unique K/L combinations}
\usage{
plotGridSearchPerplexity(celda.list)
}
\arguments{
\item{celda.list}{Object of class 'celda_list'.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters.
}
\description{
Visualize perplexity of every model in a celda_list, by unique K/L combinations
}
\examples{
celda.sim = simulateCells(model="celda_CG", K=5, L=10)
cgs = celdaGridSearch(celda.sim$counts, model="celda_CG", params=list(K=3:5, L=9:11))
cgs = calculatePerplexityWithResampling(celda.sim$counts, cgs)
plotGridSearchPerplexity(cgs)
}
