% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity}
\alias{plotGridSearchPerplexity}
\title{Visualize perplexity of a list of celda models}
\usage{
plotGridSearchPerplexity(celdaList, sep = 1)
}
\arguments{
\item{celdaList}{Object of class 'celdaList'.}

\item{sep}{Numeric. Breaks in the x axis of the resulting plo.t.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering
 parameters.
}
\description{
Visualize perplexity of every model in a celdaList, by unique
 K/L combinations
}
\examples{
## Run various combinations of parameters with 'celdaGridSearch'
celdaCGGridSearchRes <- resamplePerplexity(
  celdaCGSim$counts,
  celdaCGGridSearchRes)
plotGridSearchPerplexity(celdaCGGridSearchRes)
}
