% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity}
\alias{plotGridSearchPerplexity}
\title{Generate factorized matrices showing each feature's influence on cell / gene clustering}
\usage{
plotGridSearchPerplexity(celda.list)
}
\arguments{
\item{celda.list}{An object returned from 'celdaGridSearch'}
}
\description{
Generate factorized matrices showing each feature's influence on cell / gene clustering
}
