% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{plotGridSearchPerplexity}
\alias{plotGridSearchPerplexity}
\title{Visualize perplexity of every model in a celda_list, by unique K/L combinations}
\usage{
plotGridSearchPerplexity(celda.list)
}
\arguments{
\item{celda.list}{Object of class 'celda_list'.}
}
\value{
A ggplot plot object showing perplexity as a function of clustering parameters.
}
\description{
Visualize perplexity of every model in a celda_list, by unique K/L combinations
}
\examples{
## Run various combinations of parameters with 'celdaGridSearch'
celda.CG.grid.search.res = resamplePerplexity(celda.CG.sim$counts, 
                                              celda.CG.grid.search.res)
plotGridSearchPerplexity(celda.CG.grid.search.res)
}
