% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexity.R
\name{plotGridSearchPerplexityDiff}
\alias{plotGridSearchPerplexityDiff}
\alias{plotGridSearchPerplexityDiff,SingleCellExperiment-method}
\alias{plotGridSearchPerplexityDiff,celdaList-method}
\title{Visualize perplexity differences of a list of celda models}
\usage{
plotGridSearchPerplexityDiff(x, ...)

\S4method{plotGridSearchPerplexityDiff}{SingleCellExperiment}(x, sep = 1)

\S4method{plotGridSearchPerplexityDiff}{celdaList}(x, sep = 1)
}
\arguments{
\item{x}{Can be one of
\itemize{
 \item A \linkS4class{SingleCellExperiment} object returned from
 \code{celdaGridSearch}, \code{recursiveSplitModule},
 or \code{recursiveSplitCell}. Must contain a list named
 \code{"celda_grid_search"} in \code{metadata(x)}.
 \item celdaList object.}}

\item{sep}{Numeric. Breaks in the x axis of the resulting plot.}
}
\value{
A ggplot plot object showing perplexity diferences as a function of
 clustering parameters.
}
\description{
Visualize perplexity differences of every model in a celdaList,
 by unique K/L combinations.
}
\examples{
data(sceCeldaCGGridSearch)
sce <- resamplePerplexity(sceCeldaCGGridSearch)
plotGridSearchPerplexityDiff(sce)
data(celdaCGSim, celdaCGGridSearchRes)
## Run various combinations of parameters with 'celdaGridSearch'
celdaCGGridSearchRes <- resamplePerplexity(
  celdaCGSim$counts,
  celdaCGGridSearchRes)
plotGridSearchPerplexityDiff(celdaCGGridSearchRes)
}
