% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{recodeClusterY}
\alias{recodeClusterY}
\title{Re-code gene cluster labels by provided mapping scheme}
\usage{
recodeClusterY(celda.mod, from, to)
}
\arguments{
\item{celda.mod}{An object of class celda_G or celda_CG}

\item{from}{A numeric vector of the "keys", corresponding to the "values" in the to parameter}

\item{to}{A numeric vector of the "values"; what each corresponding number in "from" should be mapped to}
}
\description{
This function will re-code _gene_ cluster labels based off of a mapping provided by the user,
for all fields on a celda object involving cluster labels.
e.g. if Y (gene cluster) values range from 1-4 and the user would like all 3's switched to 1's and
vice versa, this function can be useful. NOTE: it is recommended that this function's results
aren't used to overwrite the original celda model object provided, in the event of a mis-mapping.
}
