% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{recodeClusterY}
\alias{recodeClusterY}
\title{Recode feature module clusters}
\usage{
recodeClusterY(celdaMod, from, to)
}
\arguments{
\item{celdaMod}{Celda object of class `celda_G` or `celda_CG`.}

\item{from}{Numeric vector. Unique values in the range of 1:L that
correspond to the original cluster labels in `celdaMod`.}

\item{to}{Numeric vector. Unique values in the range of 1:L that correspond
to the new cluster labels.}
}
\value{
Celda object with recoded feature module clusters, with class
 corresponding to that of `celdaMod`.
}
\description{
Recode feature module clusters using a mapping in the
 `from` and `to` arguments.
}
\examples{
celdaMod.reordered.y <- recodeClusterY(celdaCGMod, c(1, 3), c(3, 1))
}
