% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{recodeClusterZ}
\alias{recodeClusterZ}
\title{Recode cell cluster labels}
\usage{
recodeClusterZ(celda.mod, from, to)
}
\arguments{
\item{celda.mod}{Celda object of class `celda_C` or `celda_CG`.}

\item{from}{Numeric vector. Unique values in the range of 1:K that correspond to the original cluster labels in `celda.mod`.}

\item{to}{Numeric vector. Unique values in the range of 1:K that correspond to the new cluster labels.}
}
\value{
Celda object with cell subpopulation clusters, with class corresponding to that of `celda.mod`.
}
\description{
Recode cell subpopulaton clusters using a mapping in the `from` and `to` arguments.
}
\examples{
celda.mod.reordered.z = recodeClusterZ(celda.CG.mod, c(1, 3), c(3, 1))
}
