% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_heatmap.R
\name{render_celda_heatmap}
\alias{render_celda_heatmap}
\title{plot the heatmap of the counts data}
\usage{
render_celda_heatmap(counts, z = NULL, y = NULL, scale.log = FALSE,
  scale.row = TRUE, scale_function = scale, normalize = "none",
  z.trim = c(-2, 2), cluster.row = TRUE, cluster.column = TRUE,
  annotation_cell = NULL, annotation_gene = NULL, col = NULL,
  breaks = NULL, legend = TRUE, annotation_legend = TRUE,
  annotation_names_gene = TRUE, annotation_names_cell = TRUE,
  show_genenames = FALSE, show_cellnames = FALSE)
}
\arguments{
\item{counts}{the counts matrix}

\item{z}{A numeric vector of cluster assignments for cell}

\item{y}{A numeric vector of cluster assignments for gene}

\item{scale.log}{specify the transformation type of the matrix for (semi-)heatmap, can be "log","row"(z-acore by row),"col"(z-score by column), etc. #To be completed}

\item{scale.row}{specify the transformation type of the matrix for (semi-)heatmap, can be "log","row"(z-acore by row),"col"(z-score by column), etc. #To be completed}

\item{z.trim}{two element vector to specify the lower and upper cutoff of the z-score normalization result by default it is set to NULL so no trimming will be done.}

\item{cluster.row}{boolean values determining if rows should be clustered}

\item{cluster.column}{boolean values determining if columns should be clustered}

\item{annotation_cell}{a dataframe for the cell annotations (columns)}

\item{annotation_gene}{a dataframe for the gene annotations (rows)}

\item{col}{color for the heatmap}

\item{breaks}{a sequence of numbers that covers the range of values in mat and is one 
element longer than color vector. Used for mapping values to colors. Useful, if needed 
to map certain values to certain colors, to certain values. If value is NA then the 
breaks are calculated automatically.}

\item{legend}{logical to determine if legend should be drawn or not}

\item{annotation_legend}{boolean value showing if the legend for annotation tracks should be drawn}

\item{annotation_names_gene}{boolean value showing if the names for gene annotation tracks should be drawn}

\item{annotation_names_cell}{boolean value showing if the names for cell annotation tracks should be drawn}

\item{show_genenames}{boolean specifying if gene names are be shown}

\item{show_cellnames}{boolean specifying if cell names are be shown}

\item{scale_fun}{specify the function for scaling}
}
\description{
plot the heatmap of the counts data
}
\examples{

}
