% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{runParams}
\alias{runParams}
\title{Get run parameters for a celda run.}
\usage{
runParams(celda.list)
}
\arguments{
\item{celda.list}{Object of class "celda_list". An object containing celda models returned from `celdaGridSearch()`.}
}
\value{
A data.frame containing the run parameters used to generate the provided celda_list.
}
\description{
Get run parameters for a celda run.
}
\examples{
runParams(celda.CG.mod)
}
