% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{runParams}
\alias{runParams}
\title{Get run parameters for a celda run.}
\usage{
runParams(celda.res)
}
\arguments{
\item{celda.res}{A result object returned from celda()}
}
\description{
Get run parameters for a celda run.
}
