% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{runParams}
\alias{runParams}
\title{Get run parameters for a celda run.}
\usage{
runParams(celda.list)
}
\arguments{
\item{celda.res}{A celda_list object, as returned from celda()}
}
\description{
Get run parameters for a celda run.
}
