% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{runParams}
\alias{runParams}
\title{Get run parameters provided to `celdaGridSearch()`}
\usage{
runParams(celdaMod)
}
\arguments{
\item{celdaMod}{An object of class celdaList.}
}
\value{
Data Frame. Contains details on the various K/L parameters, chain
 parameters, and final log-likelihoods derived for each model in the provided
 celdaList.
}
\description{
Returns details on the clustering parameters, and model priors
 provided to `celdaGridSearch()` when the provided celdaList was
 created.
}
\examples{
data(celdaCGGridSearchRes)
runParams(celdaCGGridSearchRes)
}
