% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{runParams}
\alias{runParams}
\title{Get run parameters provided to `celdaGridSearch()`}
\usage{
runParams(celda.mod)
}
\value{
Data Frame. Contains details on the various K/L parameters, chain parameters, and final log-likelihoods derived for each model in the provided celdaList.
}
\description{
Returns details on the clustering parameters, model priors, and seeds provided to `celdaGridSearch()` when the provided celdaList was created.
}
\examples{
runParams(celda.CG.grid.search.res)
}
