% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{sampleLabel}
\alias{sampleLabel}
\alias{sampleLabel,SingleCellExperiment-method}
\alias{sampleLabel<-}
\alias{sampleLabel<-,SingleCellExperiment-method}
\title{Get or set sample labels from a celda
 \link[SingleCellExperiment]{SingleCellExperiment} object}
\usage{
sampleLabel(sce)

\S4method{sampleLabel}{SingleCellExperiment}(sce)

sampleLabel(sce) <- value

\S4method{sampleLabel}{SingleCellExperiment}(sce) <- value
}
\arguments{
\item{sce}{A \link[SingleCellExperiment]{SingleCellExperiment} object
returned by \link{celda_C}, \link{celda_G}, or \link{celda_CG}.}
}
\value{
Character. Contains the sample labels provided at model creation,
 or those automatically generated by celda.
}
\description{
Return or set the sample labels for the cells in \code{sce}.
}
\examples{
data(sceCelda_CG)
sampleLabel(sceCelda_CG)
}
