% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{sampleLabel}
\alias{sampleLabel}
\title{Get sampleLabels from a celdaModel}
\usage{
sampleLabel(celdaMod)
}
\arguments{
\item{celdaMod}{celdaModel. Options available in `celda::availableModels`.}
}
\value{
Character. Contains the sampleLabels provided at model
creation time, or those automatically generated by celda.
}
\description{
Returns the sampleLabels for the count matrix provided for
 generation of a given celdaModel.
}
\examples{
data(celdaCGMod)
sampleLabel(celdaCGMod)
}
