% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{sampleLabel}
\alias{sampleLabel}
\title{Get sample labels from a celda model}
\usage{
sampleLabel(celda.mod)
}
\value{
Character. Contains the sample labels provided at model creation time, or those automatically generated by celda.
}
\description{
Returns the sample labels for the count matrix provided for generation of a given celda model.
}
\examples{
sampleLabel(celda.CG.mod)
}
