% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCeldaC}
\alias{sceCeldaC}
\title{sceCeldaC}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
sceCeldaC
}
\description{
A \linkS4class{SingleCellExperiment} object containing the
 results of running \link{selectFeatures} and \link{celda_C} on
 \link{celdaCSim}.
}
\examples{
data(celdaCSim)
sceCeldaC <- selectFeatures(celdaCSim$counts)
sceCeldaC <- celda_C(sceCeldaC,
    K = celdaCSim$K,
    sampleLabel = celdaCSim$sampleLabel,
    nchains = 1)
}
\keyword{datasets}
