% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCeldaC}
\alias{sceCeldaC}
\title{sceCeldaC}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
sceCeldaC
}
\description{
A \linkS4class{SingleCellExperiment} object containing the
 results of running \link{celda_C} on \link{celdaCSim}.
}
\examples{
data(celdaCSim)
sceCeldaC = celda_C(celdaCSim$counts,
    K = celdaCSim$K,
    sampleLabel = celdaCSim$sampleLabel)
}
\keyword{datasets}
