% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCeldaCG}
\alias{sceCeldaCG}
\title{sceCeldaCG}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
data(sceCeldaCG)
}
\description{
A \linkS4class{SingleCellExperiment} object containing the
 results of running \link{selectFeatures} and \link{celda_CG} on
 \link{celdaCGSim}.
}
\examples{
data(celdaCGSim)
sceCeldaCG <- selectFeatures(celdaCGSim$counts)
sceCeldaCG <- celda_CG(sceCeldaCG,
    K = celdaCGSim$K,
    L = celdaCGSim$L,
    sampleLabel = celdaCGSim$sampleLabel,
    nchains = 1)
}
\keyword{datasets}
