% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCeldaCG}
\alias{sceCeldaCG}
\title{sceCeldaCG}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
sceCeldaCG
}
\description{
A \linkS4class{SingleCellExperiment} object containing the
 results of running \link{selectFeatures} and \link{celda_CG} on
 \link{celdaCGSim}.
}
\examples{
data(celdaCGSim)
sceCeldaCG = celda_CG(celdaCGSim$counts,
    K = celdaCGSim$K,
    L = celdaCGSim$L,
    sampleLabel = celdaCGSim$sampleLabel)
}
\keyword{datasets}
