% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCeldaCGGridSearch}
\alias{sceCeldaCGGridSearch}
\title{sceCeldaCGGridSearch}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
sceCeldaCGGridSearch
}
\description{
A \linkS4class{SingleCellExperiment} object containing the
 results of running \link{selectFeatures} and \link{celdaGridSearch} on
 \link{celdaCGSim}.
}
\examples{
data(celdaCGSim)
sce <- selectFeatures(celdaCGSim$counts)
sceCeldaCGGridSearch <- celdaGridSearch(sce,
    model = "celda_CG",
    paramsTest = list(K = seq(4, 6), L = seq(9, 11)),
    paramsFixed = list(sampleLabel = celdaCGSim$sampleLabel),
    bestOnly = TRUE,
    nchains = 1,
    cores = 1,
    verbose = FALSE)
}
\keyword{datasets}
