% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceCeldaG}
\alias{sceCeldaG}
\title{sceCeldaG}
\format{
A \linkS4class{SingleCellExperiment} object
}
\usage{
sceCeldaG
}
\description{
A \linkS4class{SingleCellExperiment} object containing the
 results of running \link{celda_G} on \link{celdaGSim}.
}
\examples{
data(celdaGSim)
sceCeldaG = celda_G(celdaGSim$counts, L = celdaGSim$L)
}
\keyword{datasets}
