% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_list.R
\name{selectBestModel}
\alias{selectBestModel}
\title{Select the best model from a celda_list by final log-likelihood.}
\usage{
selectBestModel(celda.list, K = c(), L = c())
}
\arguments{
\item{celda.list}{Object of class "celda_list". An object containing celda models returned from `celdaGridSearch()`.}

\item{K}{Limit search for best model to models with this number of cell clusters.}

\item{L}{Limit search for best model to models with this number of feature clusters.}
}
\value{
The celda model object with the highest finalLogLik attribute, meeting any K/L criteria provided
}
\description{
This function returns the celda model from a celda_list object containing
the maxiumim final log-likelihood. K, L, or combination of K and L must be 
provided for celda_list objects containing celda_C, celda_G, and celda_CG 
models, respectively.
}
