% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{selectBestModel}
\alias{selectBestModel}
\title{Select models with the best log likelihood from a 'celda_list' object gererated by celdaGridSearch.}
\usage{
selectBestModel(celda.list)
}
\arguments{
\item{celda.list}{Object of class "celda_list". An object containing celda models returned from `celdaGridSearch()`.}
}
\value{
The celda model object with the highest finalLogLik attribute, meeting any K/L criteria provided
}
\description{
This function returns the celda model from a celda_list object containing
the maxiumim final log-likelihood. K, L, or combination of K and L must be 
provided for celda_list objects containing celda_C, celda_G, and celda_CG 
models, respectively.
}
\examples{
celda.sim = simulateCells(model="celda_CG", K=5, L=10, G=100)
## Run various combinations of parameters with 'celdaGridSearch'
cgs = celdaGridSearch(celda.sim$counts, model="celda_CG", params.test=list(K=4:6, L=9:11), 
                      params.fixed=list(sample.label=celda.sim$sample.label),
                      best.only=FALSE, nchains=1)

## Returns same result as running celdaGridSearch with "best.only = TRUE"
cgs.best = selectBestModel(cgs)
}
