% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{selectBestModel}
\alias{selectBestModel}
\title{Select best chain within each combination of parameters}
\usage{
selectBestModel(celdaList)
}
\arguments{
\item{celdaList}{Object of class `celdaList`. An object containing celda
models returned from `celdaGridSearch()`.}
}
\value{
A new `celdaList` object containing one model with the best log
 likelihood for each set of parameters. If only one set of parameters is in
 the `celdaList`, the best model will be returned directly instead of a
 `celdaList` object.
}
\description{
Select the chain with the best log likelihood for each
 combination of tested parameters from a `celdaList` object gererated by
 `celdaGridSearch()`.
}
\examples{
data(celdaCGGridSearchRes)
## Returns same result as running celdaGridSearch with "bestOnly = TRUE"
cgsBest <- selectBestModel(celdaCGGridSearchRes)
}
\seealso{
`celdaGridSearch()` can run Celda with multiple parameters and
 chains in parallel. `subsetCeldaList()` can subset the `celdaList` object.
}
