% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{selectBestModel}
\alias{selectBestModel}
\title{Select best chain within each combination of parameters}
\usage{
selectBestModel(celda.list)
}
\arguments{
\item{celda.list}{Object of class `celda_list`. An object containing celda models returned from `celdaGridSearch()`.}
}
\value{
A new `celda_list` object containing one model with the best log likelihood for each set of parameters. If only one set of parameters is in the `celda_list`, the best model will be returned directly instead of a `celda_list` object.
}
\description{
Select the chain with the best log likelihood for each combination of tested parameters from a `celda_list` object gererated by `celdaGridSearch()`.
}
\examples{
## Returns same result as running celdaGridSearch with "best.only = TRUE"
cgs.best = selectBestModel(celda.CG.grid.search.res)
}
\seealso{
`celdaGridSearch()` can run Celda with multiple parameters and chains in parallel. `subsetCeldaList()` can subset the `celda_list` object.
}
