% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{simulateCells.celda_C}
\alias{simulateCells.celda_C}
\title{Simulate cells from the cell clustering generative model}
\usage{
\method{simulateCells}{celda_C}(model, S = 10, C.Range = c(10, 100),
  N.Range = c(500, 5000), G = 500, K = 5, alpha = 1, beta = 1,
  seed = 12345, ...)
}
\arguments{
\item{model}{Character. Options available in `celda::available.models`.}

\item{S}{Integer. Number of samples to simulate.}

\item{C.Range}{Vector of length 2 given the range (min,max) of number of cells for each sample to be randomly generated from the uniform distribution.}

\item{N.Range}{Integer vector. A vector of length 2 that specifies the lower and upper bounds of the number of counts generated for each cell. Default c(500, 5000).}

\item{G}{Numeric. The total number of features to be simulated.}

\item{K}{Integer. Number of cell populations.}

\item{alpha}{Numeric. Concentration parameter for Theta. Adds a pseudocount to each cell population in each sample. Default 1.}

\item{beta}{Numeric. Concentration parameter for Phi. Adds a pseudocount to each feature in each cell population. Default 1.}

\item{seed}{Integer. Passed to set.seed(). Default 12345.}

\item{...}{Additional parameters.}
}
\description{
Simulate cells from the cell clustering generative model
}
