% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{simulateCells.celda_CG}
\alias{simulateCells.celda_CG}
\title{Simulate cells from the cell/gene clustering generative model}
\usage{
\method{simulateCells}{celda_CG}(model, S = 10, C.Range = c(50, 100),
  N.Range = c(500, 5000), G = 1000, K = 3, L = 10, alpha = 1,
  beta = 1, gamma = 5, delta = 1, seed = 12345, ...)
}
\arguments{
\item{model}{Celda model to use for simulation. One of 'available_models'.}

\item{S}{The number of samples}

\item{C.Range}{two element vector to specify the lower and upper bound of the counts of cells for each sample}

\item{N.Range}{two element vector to specify the lower and upper bound of the counts of the transcripts}

\item{G}{The number of genes}

\item{K}{The number of cell populations}

\item{L}{The number of gene clusters being considered}

\item{alpha}{Non-zero concentration parameter for sample Dirichlet distribution}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state}

\item{seed}{starting point used for generating simulated data}

\item{...}{Other arguments}
}
\description{
Simulate cells from the cell/gene clustering generative model
}
