% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{simulateCells.celda_CG}
\alias{simulateCells.celda_CG}
\title{Simulate cells from the celda_CG model}
\usage{
simulateCells.celda_CG(model, S = 5, C.Range = c(50, 100),
  N.Range = c(500, 1000), G = 100, K = 5, L = 10, alpha = 1,
  beta = 1, gamma = 5, delta = 1, seed = 12345, ...)
}
\arguments{
\item{model}{Character. Options available in `celda::available.models`.}

\item{S}{Integer. Number of samples to simulate. Default 5.}

\item{C.Range}{Integer vector. A vector of length 2 that specifies the lower and upper bounds of the number of cells to be generated in each sample. Default c(50, 100).}

\item{N.Range}{Integer vector. A vector of length 2 that specifies the lower and upper bounds of the number of counts generated for each cell. Default c(500, 1000).}

\item{G}{Integer. The total number of features to be simulated. Default 100.}

\item{K}{Integer. Number of cell populations. Default 5.}

\item{L}{Integer. Number of feature modules. Default 10.}

\item{alpha}{Numeric. Concentration parameter for Theta. Adds a pseudocount to each cell population in each sample. Default 1.}

\item{beta}{Numeric. Concentration parameter for Phi. Adds a pseudocount to each feature module in each cell population. Default 1.}

\item{gamma}{Numeric. Concentration parameter for Eta. Adds a pseudocount to the number of features in each module. Default 5.}

\item{delta}{Numeric. Concentration parameter for Psi. Adds a pseudocount to each feature in each module. Default 1.}

\item{seed}{Integer. Passed to `set.seed()`. Default 12345.}

\item{...}{Additional parameters.}
}
\value{
List. Contains the simulated matrix `counts`, cell population clusters `z`, feature module clusters `y`, sample assignments `sample.label`, and input parameters.
}
\description{
Generates a simulated counts matrix, cell subpopulation clusters, sample labels, and feature module clusters
according to the generative process of the celda_CG model.
}
\examples{
celda.sim = simulateCells(model="celda_CG")
}
\seealso{
`celda_C()` for simulating cell subpopulations and `celda_G()` for simulating feature modules.
}
