% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{simulateCells.celda_G}
\alias{simulateCells.celda_G}
\title{Simulate cells from the gene clustering generative model}
\usage{
\method{simulateCells}{celda_G}(model, C = 100, N.Range = c(500, 5000),
  G = 1000, L = 5, beta = 1, gamma = 5, delta = 1, seed = 12345,
  ...)
}
\arguments{
\item{model}{Celda model to use for simulation. One of 'available_models'.}

\item{C}{The number of cells}

\item{N.Range}{Vector of length 2 given the range (min,max) of number of counts for each cell to be randomly generated from the uniform distribution}

\item{G}{The number of genes for which to simulate counts}

\item{L}{The number of transcriptional states}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state.}

\item{delta}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state}

\item{seed}{Parameter to set.seed() for random number generation}

\item{...}{Other arguments}
}
\description{
Generate a simulated count matrix, based off a generative distribution whose 
parameters can be provided by the user.
}
