% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_G.R
\name{simulateCells.celda_G}
\alias{simulateCells.celda_G}
\title{Generate Count Data}
\usage{
simulateCells.celda_G(C = 100, N.Range = c(500, 5000), G = 1000, L = 5,
  beta = 1, gamma = 1, delta = 1, seed = 12345)
}
\arguments{
\item{C}{The number of cells}

\item{N.Range}{The range of counts each gene should have}

\item{G}{The number of genes for which to simulate counts}

\item{L}{The number of transcriptional states}

\item{beta}{The Dirichlet distribution parameter for Phi; adds a pseudocount to each transcriptional state within each cell}

\item{gamma}{The Dirichlet distribution parameter for Psi; adds a pseudocount to each gene within each transcriptional state}

\item{delta}{The Dirichlet distribution parameter for Eta; adds a gene pseudocount to the numbers of genes each state}

\item{seed}{Parameter to set.seed() for random number generation}
}
\description{
Generate a simulated count matrix, based off a generative distribution whose 
parameters can be provided by the user.
}
\examples{
TODO
}
