% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{simulateCells}
\alias{simulateCells}
\title{Simulate count data from the celda generative models.}
\usage{
simulateCells(model, ...)
}
\arguments{
\item{model}{Character. Options available in `celda::available.models`.}

\item{...}{Additional parameters.}
}
\value{
List. Contains the simulated counts matrix, derived cell cluster assignments, the provided parameters, and estimated Dirichlet distribution parameters for the model.
}
\description{
This function generates a list containing a simulated counts matrix, as well as various parameters
used in the simulation which can be useful for running celda. The user must provide the desired model
(one of celda_C, celda_G, celda_CG) as well as any desired tuning parameters for those model's simulation
functions as detailed below.
}
\examples{
dim(celda.CG.sim$counts)
}
