% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{simulateContamination}
\alias{simulateContamination}
\title{Simulate contaminated count matrix}
\usage{
simulateContamination(...)
}
\arguments{
\item{...}{Arguments passed to
\code{\link[decontX]{simulateContamination}} (for example \code{C},
\code{G}, \code{K}, \code{NRange}, \code{beta}, \code{delta},
\code{numMarkers}, \code{seed}).}
}
\value{
A list containing the \code{nativeMatrix} (real expression),
\code{observedMatrix} (real expression + contamination), as well as other
parameters used in the simulation.
}
\description{
This function generates a list containing two count matrices --
 one for real expression, the other one for contamination, as well as other
 parameters used in the simulation which can be useful for running
 decontamination.

 As of celda 1.23.0 this is a thin wrapper around
 \code{\link[decontX]{simulateContamination}} in the \pkg{decontX} package,
 which must be installed.
}
\examples{
\dontshow{if (requireNamespace("decontX", quietly = TRUE)) withAutoprint(\{ # examplesIf}
contaminationSim <- simulateContamination(K = 3, delta = c(1, 10))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[decontX]{simulateContamination}}
}
\author{
Shiyi Yang, Yuan Yin, Joshua Campbell
}
