% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{subsetCeldaList}
\alias{subsetCeldaList}
\title{Select a subset of models from a 'celda_list' object generated by celdaGridSearch.}
\usage{
subsetCeldaList(celda.list, params)
}
\arguments{
\item{celda.list}{celda.list Object of class "celda_list". An object containing celda models returned from `celdaGridSearch()`.}

\item{params}{List. List of parameters used to subset celda.list.}
}
\value{
A new 'celda_list' object containing all models matching the provided criteria in 'params'. If entry in the list matches, the results for the matching model will be returned.
}
\description{
Convenience function for picking out specific models from a celda_list object.
Models can be selected by various parameters, most importantly the K/L parameters (number of cell
 clusters / number of feature clusters).
}
\examples{
celda.sim = simulateCells(model="celda_CG", K=5, L=10)
cgs = celdaGridSearch(celda.sim$counts, model="celda_CG", params.test=list(K=4:6, L=9:11), params.fixed=list(sample.label=celda.sim$sample.label))
res.K5.L10 = subsetCeldaList(cgs, params=list(K=5, L=10))
}
