% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{subsetCeldaList}
\alias{subsetCeldaList}
\title{Subset celdaList object from celdaGridSearch}
\usage{
subsetCeldaList(celdaList, params)
}
\arguments{
\item{celdaList}{celdaList Object of class `celdaList`. An object
containing celda models returned from `celdaGridSearch`.}

\item{params}{List. List of parameters used to subset celdaList.}
}
\value{
A new `celdaList` object containing all models matching the
 provided criteria in `params`. If only one item in the `celdaList` matches
 the given criteria, the matching model will be returned directly instead of
 a `celdaList` object.
}
\description{
Select a subset of models from a `celdaList` object generated
 by `celdaGridSearch()` that match the criteria in the argument `params`.
}
\examples{
data(celdaCGGridSearchRes)
resK5L10 <- subsetCeldaList(celdaCGGridSearchRes, params = list(
  K = 5,
  L = 10
))
}
\seealso{
`celdaGridSearch()` can run Celda with multiple parameters and
 chains in parallel. `selectBestModel()` can get the best model for each
 combination of parameters.
}
