% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celdaGridSearch.R
\name{subsetCeldaList}
\alias{subsetCeldaList}
\title{Subset celda_list object from celdaGridSearch}
\usage{
subsetCeldaList(celda.list, params)
}
\arguments{
\item{celda.list}{celda.list Object of class `celda_list`. An object containing celda models returned from `celdaGridSearch`.}

\item{params}{List. List of parameters used to subset celda.list.}
}
\value{
A new `celda_list` object containing all models matching the provided criteria in `params`. If only one item in the `celda_list` matches the given criteria, the matching model will be returned directly instead of a `celda_list` object.
}
\description{
Select a subset of models from a `celda_list` object generated by `celdaGridSearch()` that match the criteria in the argument `params`.
}
\examples{
res.K5.L10 = subsetCeldaList(celda.CG.grid.search.res, params=list(K=5, L=10))
}
\seealso{
`celdaGridSearch()` can run Celda with multiple parameters and chains in parallel. `selectBestModel()` can get the best model for each combination of parameters.
}
