% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda.R
\name{validate_counts}
\alias{validate_counts}
\title{Perform some simple checks on the counts matrix, to ensure celda won't choke.}
\usage{
validate_counts(counts, K, L)
}
\arguments{
\item{counts}{A count matrix}

\item{K}{the number of cell clusters requested}

\item{L}{the number of gene clusters requested}
}
\description{
Perform some simple checks on the counts matrix, to ensure celda won't choke.
}
