% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{violinPlot}
\alias{violinPlot}
\title{Feature Expression Violin Plot}
\usage{
violinPlot(counts, celda.mod, features)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent cells.}

\item{celda.mod}{Celda object of class "celda_G" or "celda_CG".}

\item{features}{Character vector. Uses these genes for plotting.}
}
\value{
Violin plot for each feature, grouped by celda cluster
}
\description{
Outputs a violin plot for feature expression data.
}
\examples{
violinPlot(counts = celda.CG.sim$counts,
           celda.mod = celda.CG.mod, features = "Gene_1")
}
