% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_functions.R
\name{violinPlot}
\alias{violinPlot}
\title{Feature Expression Violin Plot}
\usage{
violinPlot(counts, celdaMod, features, plotDots = FALSE)
}
\arguments{
\item{counts}{Integer matrix. Rows represent features and columns represent
cells.}

\item{celdaMod}{Celda object of class "celda_G" or "celda_CG".}

\item{features}{Character vector. Uses these genes for plotting.}

\item{plotDots}{\strong{TRUE} or \strong{FALSE}. If \strong{TRUE}, the
expression of features will be plotted as points in addition to the violin
curve.}
}
\value{
Violin plot for each feature, grouped by celda cluster
}
\description{
Outputs a violin plot for feature expression data.
}
\examples{
violinPlot(counts = celdaCGSim$counts,
    celdaMod = celdaCGMod, features = "Gene_1")
}
