% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{visualize_model_performance.celda_CG}
\alias{visualize_model_performance.celda_CG}
\title{visualize_model_performance for Celda Cell and Gene clustering model}
\usage{
\method{visualize_model_performance}{celda_CG}(celda.list,
  method = "perplexity", title = "Model Performance (All Chains)")
}
\arguments{
\item{celda.list}{A celda model object of "Celda_CG"}

\item{method}{One of "perplexity", "harmonic", or "loglik"}

\item{title}{Title for the visualize_model_performance}
}
\description{
visualize_model_performance for Celda Cell and Gene clustering model
}
