% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_CG.R
\name{visualize_model_performance.celda_CG}
\alias{visualize_model_performance.celda_CG}
\title{Visualize the performance of celda_CG models grouped by L and K}
\usage{
\method{visualize_model_performance}{celda_CG}(celda.list,
  method = "perplexity", title = "Model Performance (All Chains)")
}
\arguments{
\item{celda.list}{A list of celda_CG objects returned from celda function}

\item{method}{One of "perplexity", "harmonic", or "loglik"}

\item{title}{Title for the visualize_model_performance}
}
\description{
Plot the performance of a list of celda_CG models returned 
from running celda function. For each number of gene clusters 
L (cell clusters K), plot the performance of each number of cell
clusters K (gene clusters L).
}
