% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celda_C.R
\name{visualizeModelPerformance.celda_C}
\alias{visualizeModelPerformance.celda_C}
\title{visualizeModelPerformance for celda Cell clustering function}
\usage{
\method{visualizeModelPerformance}{celda_C}(celda.list, counts,
  method = "perplexity", title = "Model Performance (All Chains)",
  log = F)
}
\arguments{
\item{celda.list}{A celda_list object returned from celda()}

\item{counts}{The counts used to generate the celda.list results}

\item{method}{One of "perplexity", "loglik"}

\item{title}{Title for the plot}

\item{log}{Currently not working for celda_C objects}
}
\description{
visualizeModelPerformance for celda Cell clustering function
}
